### 2021-06-09 #3349 @whedon
[REVIEW]: Makie.jl: Flexible high-performance data visualization for Julia
**Submitting author:** @jkrumbiegel (<a href="http://orcid.org/0000-0002-8409-5590">Julius Krumbiegel</a>)
**Repository:** <a href="https://github.com/JuliaPlots/Makie.jl" target ="_blank">https://github.com/JuliaPlots/Makie.jl</a>
**Version:** v.0.15.0
