### 2021-06-10 50e4f7a
make series work with palette of colors
