### 2021-06-10 cb829d9
add tests for categorical
