+++ src/basic_recipes/series.jl
@@ -42,6 +42,10 @@ function categorical_colors(cols::AbstractVector{<: Colorant}, categories::Integ
     return to_colormap(cols)
 end
 
+function categorical_colors(cols::AbstractVector, categories::Integer)
+    return categorical_colors(to_color.(cols), categories)
+end
+
 function categorical_colors(cs::Union{String, Symbol}, categories::Integer)
     cs_string = string(cs)
     if cs_string in all_gradient_names
