### 2021-06-10 d0a5f1f
[failing] change color API to use data length
