### 2021-06-10 cb6dc35
make sure we have one texture atlas per resolution
