### 2021-06-10 23aaf59
Add boxplot coloring examples
