+++ src/camera/camera3d.jl
@@ -1,329 +1,532 @@
-
-@enum ProjectionEnum Perspective Orthographic
-
 struct Camera3D <: AbstractCamera
-    rotationspeed::Node{Float32}
-    translationspeed::Node{Float32}
     eyeposition::Node{Vec3f0}
     lookat::Node{Vec3f0}
     upvector::Node{Vec3f0}
-    fov::Node{Float32}
+
+    zoom_mult::Node{Float32}
     near::Node{Float32}
     far::Node{Float32}
-    projectiontype::Node{ProjectionEnum}
-    pan_button::Node{ButtonTypes}
-    rotate_button::Node{ButtonTypes}
-    move_key::Node{ButtonTypes}
+    pulser::Node{Float64}
+
+    attributes::Attributes
 end
 
 """
-    cam3d_cad!(scene; kw_args...)
+    keyboard_cam!(scene[; attributes...])
 
-Creates a 3D camera for `scene` which rotates around
-the _viewer_'s "up" axis - similarly to how it's done
-in CAD software cameras.
+Creates a 3d camera with a lot of controls.
+
+# Attributes:
+
+## Camera settings:
+- `fov = 45f0`: "Neutral" field of view.
+- `near = automatic`: Value of the near clip. By default this will be chosen 
+    based on the scenes bounding box. The final value is outside attributes.
+- `far = automatic`: Value of the far clip. By default this will be chosen 
+based on the scenes bounding box. The final value is outside attributes.
+- `rotation_center = :lookat`: Use `lookat` or `eyeposition` as rotation center.
+- `projectiontype = Perspective`: Switch between `Orthographic` and `Perspective` projection.
+- `fixed_axis = true`: If true panning uses the (world) z-axis instead of the vertical direction.
+- `zoom_shift_lookat = true`: If true attempt to keep data under the cursor in view when zooming
+- `cad = false`: If true rotate view when zooming off-center.
+
+## Keyboard controlled camera
+- `up_key   = Keyboard.left_shift`: Translation towards the top of the screen.
+- `down_key = Keyboard.left_control`: Translation towards the bottom of the screen.
+- `left_key  = Keyboard.a`: Translation towards the left of the screen.
+- `right_key = Keyboard.d`: Translation towards the right of the screen.
+- `forward_key  = Keyboard.w`: Translation into the screen.
+- `backward_key = Keyboard.s`: Translation out of the screen.
+
+- `zoom_in_key   = Keyboard.i`: Zoom into the scene (enlarge, via fov).
+- `zoom_out_key  = Keyboard.k`: Zoom out of the scene (shrink, via fov).
+- `stretch_view_key  = Keyboard.page_up`: Moves `eyepostion` away from `lookat`.
+- `contract_view_key = Keyboard.page_down`: Moves `eyeposition` towards `lookat`.
+
+- `pan_left_key  = Keyboard.j`: Rotation around the screens vertical direction.
+- `pan_right_key = Keyboard.l`: Rotation around the screens vertical direction.
+- `tilt_up_key   = Keyboard.r`: Rotation around the screens horizontal direction.
+- `tilt_down_key = Keyboard.f`: Rotation around the screens horizontal direction.
+- `roll_clockwise_key        = Keyboard.e`: Rotation of the screen.
+- `roll_counterclockwise_key = Keyboard.q`: Rotation of the screen.
+
+- `keyboard_rotationspeed = 1f0`: Set the speed of keyboard based rotations.
+- `keyboard_translationspeed = 0.5f0`: Set the speed of keyboard based translations.
+- `keyboard_zoomspeed = 1f0`: Set the speed of keyboard based zooms.
+- `update_rate = 1/30`: Rate at which keyboard based camera updates trigger.
+
+## Buttons/keys for mouse controls
[…]
