### 2021-06-10 e500e2d
Update docs/src/plotting_functions/boxplot.md
Co-authored-by: Pietro Vertechi <pietro.vertechi@protonmail.com>
