### 2021-06-10 928f086
cleaner outliercolor code
