### 2021-06-10 cf8fdb1
Update src/stats/boxplot.jl
Co-authored-by: Pietro Vertechi <pietro.vertechi@protonmail.com>
