+++ docs/src/cameras.md
@@ -4,47 +4,99 @@ A `Camera` is simply a viewport through which the Scene is visualized.  `Makie` 
 
 To specify the camera you want to use for your Scene, you can set the `camera` attribute.  Currently, we offer four types of camera:
 
+[`campixel!`](@ref)
 [`cam2d!`](@ref)
 [`cam3d!`](@ref)
-[`campixel!`](@ref)
 [`cam3d_cad!`](@ref)
 
 which will mutate the camera of the Scene into the specified type.
 
-## Controlling the camera
+## Pixel Camera
 
-We offer several functions to control the camera programatically.  
-You can rotate, translate, zoom and change the speed of the camera, as well as setting it to "look at" a certain point.
+The pixel camera ([`campixel!(scene)`](@ref)) projects the scene in pixel space, i.e. each integer step in the displayed data will correspond to one pixel. There are no controls for this camera. The clipping limits are set to `(-10_000, 10_000)`.
 
-[`translate_cam!`](@ref)
-[`rotate_cam!`](@ref)
-[`zoom!`](@ref)
+## 2D Camera
 
+The 2D camera ([`cam2d!(scene)`](@ref)) uses an orthographic projection with a fixed rotation and aspect ratio. You can set the following attributes via keyword arguments in `cam2d!` or by accessing the camera struct `cam = cameracontrols(scene)`:
 
-In addition, the camera can be configured by setting the fields of `cam = cameracontrols(scene)`:
+- `zoomspeed = 0.10f0` sets the speed of mouse wheel zooms.
+- `zoombutton = nothing` sets an additional key that needs to be pressed in order to zoom. Defaults to no key.
+- `panbutton = Mouse.right` sets the mouse button that needs to be pressed to translate the view.
+- `selectionbutton = (Keyboard.space, Mouse.left)` sets a set of buttons that need to be pressed to perform rectangle zooms.
 
-- `pan_button`: The mouse button used for panning.
-- `rotate_button`: The mouse button used for rotating the camera.
-- `translationspeed`: This parameter allows to customize the panning speed.
-- `rotationspeed`: This parameter allows to customize the camera rotation speed
-
-The other fields control camera positioning, and the camera needs to be updated (with [`update_cam!`](@ref)) after changing these:
-
-- `lookat`: The camera is directed at this point.
-- `eyeposition`: The camera is at this point.
-- `upvector`: The camera's "up" direction follows this vector (normalized internally).
-- `fov`: The field of view of the camera (not used for zooming internally).
-
-As an example, to set the rotation speed of the camera, simply execute `cam.rotationspeed[] = 0.05`.
+Note that this camera is not used by MakieLayout `Axis`. It is used, by default, for 2D `LScene`s and `Scene`s.
 
 ## 3D Camera
+
+The 3D camera is (or can be) unrestricted in terms of rotations and translations. Both [`cam3d!(scene)`](@ref) and [`cam3d_cad!(scene)`](@ref) create this camera type. Unlike the 2D camera, settings and controls are stored in the `cam.attributes` field rather than in the struct directly, but can still be passed as keyword arguments. The general camera settings include
+
+- `fov = 45f0` sets the "neutral" field of view, i.e. the fov corresponding to no zoom. This is irrelevant if the camera uses an orthographic projection. 
+- `near = automatic` sets the value of the near clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.near`.
+- `far = automatic` sets the value of the far clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.far`.
+- `rotation_center = :lookat` sets the default center for camera rotations. Currently allows `:lookat` or `:eyeposition`.
[…]
