+++ src/interfaces.jl
@@ -827,7 +827,7 @@ function plot!(scene::Combined, P::PlotFunc, attributes::Attributes, input::NTup
 end
 
 function apply_camera!(scene::Scene, cam_func)
-    if cam_func in (cam2d!, cam3d!, campixel!, cam3d_cad!, keyboard_cam!)
+    if cam_func in (cam2d!, cam3d!, campixel!, cam3d_cad!)
         cam_func(scene)
     else
         error("Unrecognized `camera` attribute type: $(typeof(cam_func)). Use automatic, cam2d! or cam3d!, campixel!, cam3d_cad!")
