+++ src/camera/camera3d.jl
@@ -213,8 +213,12 @@ function Camera3D(scene; kwargs...)
 end
 
 # These imitate the old camera
-cam3d!(scene) = Camera3D(scene, zoom_shift_lookat = true, fixed_axis = true)
-cam3d_cad!(scene) = Camera3D(scene, cad = true, zoom_shift_lookat = false, fixed_axis = false)
+function cam3d!(scene; zoom_shift_lookat = true, fixed_axis = true, kwargs...)
+    Camera3D(scene, zoom_shift_lookat = zoom_shift_lookat, fixed_axis = fixed_axis; kwargs...)
+end
+function cam3d_cad!(scene; cad = true, zoom_shift_lookat = false, fixed_axis = false, kwargs...)
+    Camera3D(scene, cad = cad, zoom_shift_lookat = zoom_shift_lookat, fixed_axis = fixed_axis; kwargs...)
+end
 
 
 function add_translation!(scene, cam::Camera3D)
