### 2021-06-11 513780c
use best filtering when possible and use filter when downsampling
