### 2021-06-11 e8fd585
Apply suggestions from code review
Co-authored-by: Eric Hanson <5846501+ericphanson@users.noreply.github.com>
