+++ test/cairomakie.jl
@@ -1,0 +1,55 @@
+using CairoMakie
+
+CairoMakie.activate!()
+recorded = joinpath(@__DIR__, "cairo_images")
+rm(recorded; force=true, recursive=true); mkdir(recorded)
+
+excludes = Set([
+    "Colored Mesh",
+    "Line GIF",
+    "Streamplot animation",
+    "Line changing colour",
+    "Axis + Surface",
+    "Streamplot 3D",
+    "Meshscatter Function",
+    "Hollow pie chart",
+    "Record Video",
+    "Image on Geometry (Earth)",
+    "Comparing contours, image, surfaces and heatmaps",
+    "Textured Mesh",
+    "Simple pie chart",
+    "Animated surface and wireframe",
+    "Open pie chart",
+    "image scatter",
+    "surface + contour3d",
+    "Orthographic Camera",
+    "Legend",
+    "rotation",
+    "3D Contour with 2D contour slices",
+    "Surface with image",
+    "Test heatmap + image overlap",
+    "Text Annotation",
+    "step-2",
+    "FEM polygon 2D.png",
+    "Text rotation",
+    "Image on Surface Sphere",
+    "FEM mesh 2D",
+    "Hbox",
+    "Stars",
+    "Subscenes",
+    "Arrows 3D",
+    "Layouting",
+    # sigh this is actually super close,
+    # but doesn't interpolate the values inside the
+    # triangles, so looks pretty different
+    "FEM polygon 2D",
+    "Connected Sphere",
+    # markers too big, close otherwise, needs to be assimilated with glmakie
+    "Unicode Marker",
+])
+excludes2 = Set(["short_tests_90", "short_tests_111", "short_tests_35", "short_tests_13", "short_tests_3"])
+
+functions = [:volume, :volume!, :uv_mesh]
+database = database_filtered!(excludes, excludes2, functions=functions)
+ReferenceTests.record_tests(database, recording_dir=recorded)
+ReferenceTests.reference_tests(recorded)
