+++ ReferenceTests/src/image_download.jl
@@ -7,7 +7,7 @@ end
 
 # Well, to be more precise, last non patch
 function last_major_version()
-    path = basedir("..", "..", "Project.toml")
+    path = basedir("..", "Project.toml")
     version = VersionNumber(TOML.parse(String(read(path)))["version"])
     return "v" * string(VersionNumber(version.major, version.minor))
 end
