+++ test/wglmakie.jl
@@ -1,0 +1,39 @@
+using ElectronDisplay
+ElectronDisplay.CONFIG.showable = showable
+ElectronDisplay.CONFIG.single_window = true
+ElectronDisplay.CONFIG.focus = false
+using WGLMakie
+WGLMakie.activate!()
+recorded = joinpath(@__DIR__, "wgl_images")
+rm(recorded; force=true, recursive=true); mkdir(recorded)
+
+excludes = Set([
+    "Streamplot animation",
+    "Transforming lines",
+    "image scatter",
+    "Line GIF",
+    "surface + contour3d",
+    # Hm weird, looks like some internal JSServe error missing an Observable:
+    "Errorbars x y low high",
+    "Rangebars x y low high",
+    # These are a bit sad, since it's just missing interpolations
+    "FEM mesh 2D",
+    "FEM polygon 2D",
+    # missing transparency & image
+    "Wireframe of a Surface",
+    "Image on Surface Sphere",
+    "Surface with image",
+    # Marker size seems wrong in some occasions:
+    "Hbox",
+    "UnicodeMarker",
+    # Not sure, looks pretty similar to me! Maybe blend mode?
+    "Test heatmap + image overlap",
+    "Stars",
+    "heatmaps & surface",
+    "OldAxis + Surface"
+])
+excludes2 = Set(["short_tests_83", "short_tests_78", "short_tests_40", "short_tests_13", "short_tests_5"])
+database = database_filtered!(excludes, excludes2)
+ReferenceTests.record_tests(database, recording_dir=recorded)
+ReferenceTests.reference_tests(recorded)
+# ReferenceTests.generate_test_summary("preview.html", recorded)

[…]
