+++ WGLMakie/test/runtests.jl
@@ -4,18 +4,9 @@ ElectronDisplay.CONFIG.single_window = true
 ElectronDisplay.CONFIG.focus = false
 using ImageMagick, FileIO
 using WGLMakie, Makie, Test
-using Pkg
-
-# ImageIO seems broken on 1.6 ... and there doesn't
-# seem to be a clean way anymore to force not to use a loader library?
-filter!(x-> x !== :ImageIO, FileIO.sym2saver[:PNG])
-filter!(x-> x !== :ImageIO, FileIO.sym2loader[:PNG])
-Makie.set_theme!(resolution=(400, 400))
-
-path = normpath(joinpath(dirname(pathof(Makie)), "..", "test", "ReferenceTests"))
-Pkg.develop(PackageSpec(path = path))
 using ReferenceTests
-using ReferenceTests: nice_title
+using ReferenceTests: database_filtered
+
 excludes = Set([
     "Streamplot animation",
     "Transforming lines",
@@ -37,19 +28,13 @@ excludes = Set([
     "UnicodeMarker",
     # Not sure, looks pretty similar to me! Maybe blend mode?
     "Test heatmap + image overlap",
-    "Stars"
+    "Stars",
+    "heatmaps & surface",
+    "OldAxis + Surface"
 ])
+excludes2 = Set(["short_tests_83", "short_tests_78", "short_tests_40", "short_tests_13", "short_tests_5", "short_tests_41"])
+database = database_filtered(excludes, excludes2)
 
-database = ReferenceTests.load_database()
-filter!(database) do (name, entry)
-    !(entry.title in excludes) &&
-    nice_title(entry) !== "short_tests_83" &&
-    nice_title(entry) !== "short_tests_78" &&
-    # difference in line rendering
-    nice_title(entry) !== "short_tests_40" &&
-    # Difference in text marker scaling (needs proper fixing in GLMakie + WGLMakie)
-    nice_title(entry) !== "short_tests_13"
-end
 recorded = joinpath(@__DIR__, "recorded")
 rm(recorded; force=true, recursive=true); mkdir(recorded)
 ReferenceTests.record_tests(database; recording_dir=recorded)
