+++ WGLMakie/test/Project.toml
@@ -6,6 +6,7 @@ ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
 JSServe = "824d6782-a2ef-11e9-3a09-e5662e0c26f9"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+ReferenceTests = "d37af2e0-5618-4e00-9939-d430db56ee94"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]

[…]
