+++ CairoMakie/Project.toml
@@ -30,6 +30,7 @@ julia = "1.3"
 [extras]
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+ReferenceTests = "d37af2e0-5618-4e00-9939-d430db56ee94"
 
 [targets]
-test = ["Test", "Pkg"]
+test = ["Test", "Pkg", "ReferenceTests"]
