### 2021-06-12 cb50493
remove hacks
