+++ .github/workflows/glmakie.yaml
@@ -57,7 +57,9 @@ jobs:
         uses: actions/upload-artifact@v2
         with:
           name: ReferenceImages
-          path: ./test/recorded
+          path: |
+            ./test/recorded
+            ./test/recorded_glmakie
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
