### 2021-06-12 66edd42
clean up urls + readmes
