+++ src/interaction/inspector.jl
@@ -42,7 +42,7 @@ function color2text(c::RGBAf0)
         @sprintf("RGB(%0.2f, %0.2f, %0.2f)", c.r, c.g, c.b)
     else
         @sprintf("RGBA(%0.2f, %0.2f, %0.2f, %0.2f)", c.r, c.g, c.b, c.alpha)
-    end 
+    end
 end
 
 color2text(name, i::Integer, j::Integer, c) = "$name[$i, $j] = $(color2text(c))"
@@ -224,8 +224,8 @@ end
         # Text
         text_padding = Vec4f0(5, 5, 3, 3), # LRBT
         text_align = (:left, :bottom),
-        textcolor = :black, 
-        textsize = 18, 
+        textcolor = :black,
+        textsize = 18,
         font = theme(scene, :font),
         _display_text = " ",
         _text_position = Point2f0(0),
@@ -242,7 +242,7 @@ end
         indicator_linestyle = nothing,
         _bbox2D = FRect2D(Vec2f0(0), Vec2f0(0)),
         _px_bbox_visible = true,
-        
+
         # general
         tooltip_align = (:center, :top), # default tooltip position relative to cursor
         tooltip_offset = Vec2f0(20), # default offset in alignment direction
@@ -266,19 +266,19 @@ function plot!(plot::_Inspector)
         text_padding, text_align, textcolor, textsize, font,
         background_color, outline_color, outline_linestyle, outline_linewidth,
         indicator_linestyle, indicator_linewidth, indicator_color,
-        tooltip_offset, depth, 
-        _display_text, _text_position, _bbox2D, _px_bbox_visible, 
+        tooltip_offset, depth,
+        _display_text, _text_position, _bbox2D, _px_bbox_visible,
         _tooltip_align, _root_px_projection, _visible
     )
 
     # tooltip text
     _aligned_text_position = Node(Point2f0(0))
     id = Mat4f0(1,0,0,0,0,1,0,0,0,0,1,0,0,0,0,1)
-    text_plot = text!(plot, _display_text, 
+    text_plot = text!(plot, _display_text,
         position = _aligned_text_position, visible = _visible, align = text_align,
         color = textcolor, font = font, textsize = textsize, show_axis = false,
-        inspectable = false, 
-        # with https://github.com/JuliaPlots/GLMakie.jl/pull/183 this should
+        inspectable = false,
+        # with https://github.com/JuliaPlots/Makie.jl/tree/master/GLMakie/pull/183 this should
         # allow the tooltip to work in any scene.
         space = :data,
         projection = _root_px_projection, view = id, projectionview = _root_px_projection
@@ -289,7 +289,7 @@ function plot!(plot::_Inspector)
         rect = Bbox_from_glyphlayout(_display_text[], gl)
         l, r, b, t = pad
         FRect2D(
-            origin(rect) .+ Vec2f0(pos[1] - l, pos[2] - b), 
+            origin(rect) .+ Vec2f0(pos[1] - l, pos[2] - b),
             widths(rect) .+ Vec2f0(l + r, b + t)
         )
     end
@@ -318,9 +318,9 @@ function plot!(plot::_Inspector)
     end
 
 
-    # tooltip background and frame 
+    # tooltip background and frame
     background = mesh!(
-        plot, bbox, color = background_color, shading = false, #fxaa = false, 
+        plot, bbox, color = background_color, shading = false, #fxaa = false,
         # TODO with fxaa here the text above becomes seethrough on a heatmap
         visible = _visible, show_axis = false, inspectable = false,
         projection = _root_px_projection, view = id, projectionview = _root_px_projection
@@ -328,15 +328,15 @@ function plot!(plot::_Inspector)
     outline = wireframe!(
         plot, bbox,
[…]
