+++ test/Project.toml
@@ -1,18 +1,12 @@
 [deps]
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 CategoricalArrays = "324d7699-5711-5eae-9e2f-1d82baa6b597"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-ElectronDisplay = "d872a56f-244b-5cc9-b574-2017b5b909a8"
-GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
-ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
 KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 MeshIO = "7269a6da-0436-5bbc-96c2-40638cbb6118"
 Observables = "510215fc-4207-5dde-b226-833fc4488ee2"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-ReferenceTests = "d37af2e0-5618-4e00-9939-d430db56ee94"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-WGLMakie = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
