+++ test/runtests.jl
@@ -1,9 +1,6 @@
 using Test
-using MeshIO
 using StaticArrays
 using Makie
-using ImageMagick
-
 using Makie.Observables
 using Makie.GeometryBasics
 using Makie.PlotUtils
@@ -11,4 +8,24 @@ using Makie.FileIO
 using Makie.IntervalSets
 using GeometryBasics: Pyramid
 
-include("unit_tests/runtests.jl")
+using Makie: volume
+
+@testset "Unit tests" begin
+    @testset "#659 Volume errors if data is not a cube" begin
+        fig, ax, vplot = volume(1:8, 1:8, 1:10, rand(8, 8, 10))
+        lims = Makie.data_limits(vplot)
+        lo, hi = extrema(lims)
+        @test all(lo .<= 1)
+        @test all(hi .>= (8,8,10))
+    end
+
+    include("conversions.jl")
+    include("quaternions.jl")
+    include("projection_math.jl")
+    include("liftmacro.jl")
+    include("makielayout.jl")
+    include("figures.jl")
+    include("transformations.jl")
+    include("stack.jl")
+    include("events.jl")
+end
