### 2021-06-12 08efbe2
CompatHelper: bump compat for "MakieCore" to "0.1"
