+++ Project.toml
@@ -68,7 +68,7 @@ ImageIO = "0.2, 0.3, 0.4, 0.5"
 IntervalSets = "0.3, 0.4, 0.5"
 Isoband = "0.1"
 KernelDensity = "0.5, 0.6"
-MakieCore = "0.1.3"
+MakieCore = "0.1.3, 0.1"
 Match = "1.1"
 Observables = "0.4"
 Packing = "0.4"
