+++ WGLMakie/Project.toml
@@ -1,6 +1,6 @@
-authors = ["SimonDanisch <sdanisch@gmail.com>"]
 name = "WGLMakie"
 uuid = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
+authors = ["SimonDanisch <sdanisch@gmail.com>"]
 version = "0.4"
 
 [deps]
@@ -25,7 +25,7 @@ GeometryBasics = "0.3"
 Hyperscript = "0.0.3, 0.0.4"
 ImageMagick = "1.1"
 JSServe = "1.2"
-Makie = "=0.14.0"
+Makie = "=0.14.0, 0.13"
 Observables = "0.4"
 ShaderAbstractions = "0.2.1"
 StaticArrays = "0.12, 1.0"
