### 2021-06-13 0ed7f87
added reference to beautiful makie gallery
