### 2021-06-15 d3cb791
Remove deprecated MakieLayout from tutorial.
