### 2021-06-16 #3349 @jkrumbiegel
@Kevin-Mattheus-Moerman I just wanted to let you know that we have recently consolidated our codebase into the https://github.com/JuliaPlots/Makie.jl/ repository, whereas before, most code lived in `AbstractPlotting.jl`. The paper lies in the AbstractPlotting repository, still, and I'm not sure if it's worth the trouble of moving and relinking it. But if you think that would be better long term, let me know. It wasn't foreseeable for me at the time I submitted the paper that we would start this major internal refactor so soon.
