### 2021-06-18 #3349 @arfon
@jkrumbiegel – please move your paper into the new repository (https://github.com/JuliaPlots/Makie.jl)
