### 2021-06-18 4b1ff5d
add menu auto-direction
automatic
