+++ WGLMakie/Project.toml
@@ -25,7 +25,7 @@ GeometryBasics = "0.3"
 Hyperscript = "0.0.3, 0.0.4"
 ImageMagick = "1.1"
 JSServe = "1.2"
-Makie = "=0.14.0"
+Makie = "=0.14.1"
 Observables = "0.4"
 ShaderAbstractions = "0.2.1"
 StaticArrays = "0.12, 1.0"
