+++ assets/paper.bib
@@ -1,0 +1,28 @@
+@article{Hunter:2007,
+  title={Matplotlib: A 2D graphics environment},
+  author={Hunter, John D},
+  journal={IEEE Annals of the History of Computing},
+  volume={9},
+  number={03},
+  pages={90--95},
+  year={2007},
+  publisher={IEEE Computer Society}
+}
+
+@article{Wickham:2011,
+  title={ggplot2},
+  author={Wickham, Hadley},
+  journal={Wiley Interdisciplinary Reviews: Computational Statistics},
+  volume={3},
+  number={2},
+  pages={180--185},
+  year={2011},
+  publisher={Wiley Online Library}
+}
+
+@article{Bezanson:2012,
+  title={Julia: A fast dynamic language for technical computing},
+  author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
+  journal={arXiv preprint arXiv:1209.5145},
+  year={2012}
+}
