+++ assets/paper.md
@@ -1,0 +1,146 @@
+---
+title: 'Makie.jl: Flexible high-performance data visualization for Julia'
+tags:
+  - Julia
+  - plotting
+  - data visualization
+  - GPU
+  - OpenGL
+authors:
+  - name: Simon Danisch
+    affiliation: 1
+  - name: Julius Krumbiegel
+    orcid: 0000-0002-8409-5590
+    affiliation: 2
+affiliations:
+ - name: Beacon Biosignals
+   index: 1
+ - name: Department of Systems Neuroscience, University Medical Center Hamburg-Eppendorf
+   index: 2
+date: 10 April 2021
+bibliography: paper.bib
+---
+
+# Summary
+
+`Makie.jl` is a cross-platform plotting ecosystem for the Julia programming language [@Bezanson:2012] , which enables researchers to create high-performance, GPU-powered, interactive visualizations, as well as publication-quality vector graphics with one unified interface.
+The infrastructure based on `Observables.jl` allows users to express how a visualization depends on multiple parameters and data sources, which can then be updated live, either programmatically, or through sliders, buttons and other GUI elements.
+A sophisticated layout system makes it easy to assemble complex figures.
+It is designed to avoid common difficulties when aligning nested subplots of different sizes, or placing colorbars or legends freely without spacing issues.
+`Makie.jl` leverages the Julia type system to automatically convert many kinds of input arguments which results in a very flexible API that reduces the need to manually prepare data.
+Finally, users can extend every step of this pipeline for their custom types through Julia's powerful multiple dispatch mechanism, making Makie a highly productive and generic visualization system.
+
+# Statement of need
+
+Good data visualization is crucial for the work of every scientist around the world.
+To effectively understand and communicate results, different disciplines ranging from publication-quality static vector graphics, through animated movies, to interactive data exploration tools, require flexible and powerful plotting software.
+Most tools available today lack one or more of the following attributes: High performance for the responsive rendering of large datasets, interactive visualizations with flexible dynamic languages, the ability to extend the plotting pipeline to handle user-defined data structures and types, implementations of both 2D and 3D rendering, and the power to construct complex figures without having to tweak positions after the fact in image editing software.
+Therefore, researchers have to switch between tools which means they have to spend more time to learn unfamiliar syntax and redo work if one software turns out to lack critical abilities for the task at hand.
+
+`Makie.jl` is a new plotting package which is built from the ground up to leverage the power of `Julia`, a relatively young programming language which excels at technical computing and has seen steady growth of its user base since reaching the 1.0 milestone in 2018.
+Julia users have historically often used plotting software from other ecosystems, such as `matplotlib` [@Hunter:2007] or `ggplot2` [@Wickham:2011] through interface packages to Python and R like `PyCall.jl` and `RCall.jl`.
+But these wrapper packages cannot take full advantage of Julia's type system and multiple dispatch paradigm, so they leave both performance and flexibility on the table.
+`Makie.jl` aims to fill this gap in the Julia ecosystem.
+
+# Example
+
+The following figure illustrates a few key features of `Makie.jl`:
+The map example augments a dataset of the world's airports to 100 million points, a size that many other plotting frameworks would struggle with.
+
+The mandelbrot fractal heatmap on the right is plotted by simply passing an x- and y-range, as well as the mandelbrot function, which is then automatically evaluated for all xy combinations.
+Julia's multiple dispatch mechanism is, in very simple terms, a way to overload functions with different versions or methods, that are chosen depending on the set of input arguments.
[…]
