### 2021-06-18 #3349 @jkrumbiegel
@arfon All right, I have moved the paper to https://github.com/JuliaPlots/Makie.jl/tree/joss-paper
