### 2021-06-24 31f1830
port @ndinsmore osx retina fix
