+++ .github/workflows/cairomakie.yaml
@@ -1,15 +1,15 @@
-name: CairoMakie CI
+name: CairoMakie
 on:
   pull_request:
-    paths-ignore:
-      - 'docs/**'
-      - '*.md'
+    paths:
+      - CairoMakie/**
+      - src/**
     branches:
       - master
   push:
-    paths-ignore:
-      - 'docs/**'
-      - '*.md'
+    paths:
+      - CairoMakie/**
+      - src/**
     branches:
       - master
     tags: '*'
