### 2021-06-24 d25104c
allow colorrange as an attribute consistent with plot types
