### 2021-06-24 108e20c
add exports for old 3d camera
