+++ src/basic_recipes/barplot.jl
@@ -1,3 +1,7 @@
+function bar_label_formatter(value::Number)
+    return string(round(value; digits=3))
+end
+
 """
     barplot(x, y; kwargs...)
 
@@ -25,17 +29,28 @@ $(ATTRIBUTES)
         visible = theme(scene, :visible),
         inspectable = theme(scene, :inspectable),
         cycle = [:color => :patchcolor],
+
+        bar_labels = nothing,
+        flip_labels_at = Inf,
+        label_color = theme(scene, :textcolor),
+        color_over_background = automatic,
+        color_over_bar = automatic,
+        label_offset = 5,
+        label_font = theme(scene, :font),
+        label_size = 20,
+        label_formatter = bar_label_formatter
     )
 end
 
 conversion_trait(::Type{<: BarPlot}) = PointBased()
 
-function bar_rectangle(x, y, width, fillto)
+function bar_rectangle(x, y, width, fillto, in_y_direction)
     # y could be smaller than fillto...
     ymin = min(fillto, y)
     ymax = max(fillto, y)
     w = abs(width)
-    return FRect(x - (w / 2f0), ymin, w, ymax - ymin)
+    rect = FRect(x - (w / 2f0), ymin, w, ymax - ymin)
+    return in_y_direction ? rect : flip(rect)
 end
 
 flip(r::Rect2D) = Rect2D(reverse(origin(r)), reverse(widths(r)))
@@ -55,19 +70,113 @@ function xw_from_dodge(x, width, minimum_distance, x_gap, dodge, n_dodge, dodge_
     return x .+ width .* shifts, width * dodge_width
 end
 
-function Makie.plot!(p::BarPlot)
+scale_width(dodge_gap, n_dodge) = (1 - (n_dodge - 1) * dodge_gap) / n_dodge
 
-    in_y_direction = lift(p.direction) do dir
-        if dir == :y
-            true
-        elseif dir == :x
-            false
-        else
-            error("Invalid direction $dir. Options are :x and :y.")
-        end
+function shift_dodge(i, dodge_width, dodge_gap)
+    (dodge_width - 1) / 2 + (i - 1) * (dodge_width + dodge_gap)
+end
+
+function stack_from_to_sorted(y)
+    to = cumsum(y)
+    from = [0.0; to[firstindex(to):end-1]]
+
+    (from = from, to = to)
+end
+
+function stack_from_to(i_stack, y)
+    # save current order
+    order = 1:length(y)
+    # sort by i_stack
+    perm = sortperm(i_stack)
+    # restore original order
+    inv_perm = sortperm(order[perm])
+
+    from, to = stack_from_to_sorted(view(y, perm))
+
+    (from = view(from, inv_perm), to = view(to, inv_perm))
+end
+
+function stack_grouped_from_to(i_stack, y, grp)
+
+    from = Array{Float64}(undef, length(y))
+    to   = Array{Float64}(undef, length(y))
+
+    groupby = StructArray((; grp..., is_pos = y .> 0))
+
+    grps = StructArrays.finduniquesorted(groupby)
+
+    for (grp, inds) in grps
+
+        fromto = stack_from_to(i_stack[inds], y[inds])
+
+        from[inds] .= fromto.from
+        to[inds] .= fromto.to
+
     end
 
-    bars = lift(p[1], p.fillto, p.width, p.dodge, p.n_dodge, p.x_gap, p.dodge_gap, p.stack, in_y_direction) do xy, fillto, width, dodge, n_dodge, x_gap, dodge_gap, stack, in_y_direction
+    (from = from, to = to)
+end
+
[…]
