+++ docs/src/plotting_functions/hist.md
@@ -6,7 +6,7 @@ hist
 
 ### Examples
 
-```@example
+```@example hist
 using GLMakie
 GLMakie.activate!() # hide
 Makie.inline!(true) # hide
@@ -19,4 +19,15 @@ hist(f[1, 2], data, bins = 20, color = :red, strokewidth = 1, strokecolor = :bla
 hist(f[2, 1], data, bins = [-5, -2, -1, 0, 1, 2, 5], color = :gray)
 hist(f[2, 2], data, normalization = :pdf)
 f
+```
+
+#### Histogram with labels
+
+You can use all the same arguments as [`barplot`](@ref):
+```@example hist
+using CairoMakie
+CairoMakie.activate!()
+hist(data, normalization = :pdf, bar_labels = :y,
+     label_formatter=x-> round(x, digits=2), label_size = 15,
+     strokewidth = 0.5, strokecolor = (:black, 0.5), color = :y)
 ```
