+++ src/camera/camera3d.jl
@@ -13,68 +13,73 @@ struct Camera3D <: AbstractCamera
 end
 
 """
-    Camera3D!(scene[; attributes...])
+    Camera3D(scene[; attributes...])
 
 Creates a 3d camera with a lot of controls.
 
-# Attributes:
+The 3D camera is (or can be) unrestricted in terms of rotations and translations. Both `cam3d!(scene)` and `cam3d_cad!(scene)` create this camera type. Unlike the 2D camera, settings and controls are stored in the `cam.attributes` field rather than in the struct directly, but can still be passed as keyword arguments. The general camera settings include
 
-## Camera settings:
-- `fov = 45f0`: "Neutral" field of view.
-- `near = automatic`: Value of the near clip. By default this will be chosen
-    based on the scenes bounding box. The final value is outside attributes.
-- `far = automatic`: Value of the far clip. By default this will be chosen
-based on the scenes bounding box. The final value is outside attributes.
-- `rotation_center = :lookat`: Use `lookat` or `eyeposition` as rotation center.
-- `projectiontype = Perspective`: Switch between `Orthographic` and `Perspective` projection.
-- `fixed_axis = true`: If true panning uses the (world) z-axis instead of the vertical direction.
-- `zoom_shift_lookat = true`: If true attempt to keep data under the cursor in view when zooming
-- `cad = false`: If true rotate view when zooming off-center.
+- `fov = 45f0` sets the "neutral" field of view, i.e. the fov corresponding to no zoom. This is irrelevant if the camera uses an orthographic projection.
+- `near = automatic` sets the value of the near clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.near`.
+- `far = automatic` sets the value of the far clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.far`.
+- `rotation_center = :lookat` sets the default center for camera rotations. Currently allows `:lookat` or `:eyeposition`.
+- `projectiontype = Perspective` sets the type of the projection. Can be `Orthographic` or `Perspective`.
+- `fixed_axis = false`: If true panning uses the (world/plot) z-axis instead of the camera up direction.
+- `zoom_shift_lookat = true`: If true attempts to keep data under the cursor in view when zooming.
+- `cad = false`: If true rotates the view around `lookat` when zooming off-center.
 
-## Keyboard controlled camera
-- `up_key   = Keyboard.left_shift`: Translation towards the top of the screen.
-- `down_key = Keyboard.left_control`: Translation towards the bottom of the screen.
-- `left_key  = Keyboard.a`: Translation towards the left of the screen.
-- `right_key = Keyboard.d`: Translation towards the right of the screen.
-- `forward_key  = Keyboard.w`: Translation into the screen.
-- `backward_key = Keyboard.s`: Translation out of the screen.
+The camera view follows from the position of the camera `eyeposition`, the point which the camera focuses `lookat` and the up direction of the camera `upvector`. These can be accessed as `cam.eyeposition` etc and adjusted via `update_cam!(scene, cameracontrols(scene), eyeposition, lookat[, upvector = Vec3f0(0, 0, 1)])`. They can also be passed as keyword arguments when the camera is constructed.
 
-- `zoom_in_key   = Keyboard.i`: Zoom into the scene (enlarge, via fov).
-- `zoom_out_key  = Keyboard.k`: Zoom out of the scene (shrink, via fov).
-- `stretch_view_key  = Keyboard.page_up`: Moves `eyepostion` away from `lookat`.
-- `contract_view_key = Keyboard.page_down`: Moves `eyeposition` towards `lookat`.
[…]
