+++ docs/src/cameras.md
@@ -6,8 +6,8 @@ To specify the camera you want to use for your Scene, you can set the `camera` a
 
 [`campixel!`](@ref)
 [`cam2d!`](@ref)
-[`cam3d!`]
-[`cam3d_cad!`]
+`cam3d!`
+`cam3d_cad!`
 
 which will mutate the camera of the Scene into the specified type.
 
@@ -28,69 +28,9 @@ Note that this camera is not used by MakieLayout `Axis`. It is used, by default,
 
 ## 3D Camera
 
-The 3D camera is (or can be) unrestricted in terms of rotations and translations. Both [`cam3d!(scene)`] and [`cam3d_cad!(scene)`] create this camera type. Unlike the 2D camera, settings and controls are stored in the `cam.attributes` field rather than in the struct directly, but can still be passed as keyword arguments. The general camera settings include
-
-- `fov = 45f0` sets the "neutral" field of view, i.e. the fov corresponding to no zoom. This is irrelevant if the camera uses an orthographic projection. 
-- `near = automatic` sets the value of the near clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.near`.
-- `far = automatic` sets the value of the far clip. By default this will be chosen based on the scenes bounding box. The final value is in `cam.far`.
-- `rotation_center = :lookat` sets the default center for camera rotations. Currently allows `:lookat` or `:eyeposition`.
-- `projectiontype = Perspective` sets the type of the projection. Can be `Orthographic` or `Perspective`.
-- `fixed_axis = false`: If true panning uses the (world/plot) z-axis instead of the camera up direction.
-- `zoom_shift_lookat = true`: If true attempts to keep data under the cursor in view when zooming.
-- `cad = false`: If true rotates the view around `lookat` when zooming off-center.
-
-The camera view follows from the position of the camera `eyeposition`, the point which the camera focuses `lookat` and the up direction of the camera `upvector`. These can be accessed as `cam.eyeposition` etc and adjusted via `update_cam!(scene, cameracontrols(scene), eyeposition, lookat[, upvector = Vec3f0(0, 0, 1)])`. They can also be passed as keyword arguments when the camera is constructed.
-
-The camera can be controlled by keyboard and mouse. The keyboard has the following available attributes
-
-- `up_key   = Keyboard.left_shift` sets the key for translations towards the top of the screen.
-- `down_key = Keyboard.left_control` sets the key for translations towards the bottom of the screen.
-- `left_key  = Keyboard.a` sets the key for translations towards the left of the screen.
-- `right_key = Keyboard.d` sets the key for translations towards the right of the screen.
-- `forward_key  = Keyboard.w` sets the key for translations into the screen.
-- `backward_key = Keyboard.s` sets the key for translations out of the screen.
-
-- `zoom_in_key   = Keyboard.i` sets the key for zooming into the scene (enlarge, via fov).
-- `zoom_out_key  = Keyboard.k` sets the key for zooming out of the scene (shrink, via fov).
-- `stretch_view_key  = Keyboard.page_up` sets the key for moving `eyepostion` away from `lookat`.
-- `contract_view_key = Keyboard.page_down` sets the key for moving `eyeposition` towards `lookat`.
-
-- `pan_left_key  = Keyboard.j` sets the key for rotations around the screens vertical axis.
-- `pan_right_key = Keyboard.l` sets the key for rotations around the screens vertical axis.
-- `tilt_up_key   = Keyboard.r` sets the key for rotations around the screens horizontal axis.
-- `tilt_down_key = Keyboard.f` sets the key for rotations around the screens horizontal axis.
[…]
