### 2021-06-26 8e117ab
make sure surface gets transformed correctly
