+++ src/stats/hist.jl
@@ -39,7 +39,6 @@ $(ATTRIBUTES)
     )
 end
 
-
 function Makie.plot!(plot::Hist)
 
     values = plot[:values]
