+++ CairoMakie/src/primitives.jl
@@ -676,7 +676,7 @@ function draw_mesh3D(
     func = Makie.transform_func_obs(scene)[]
     # pass func as argument to function, so that we get a function barrier
     # and have `func` be fully typed inside closure
-    vs = broadcast(decompose(Point, mesh), func) do v, f
+    vs = broadcast(decompose(Point, mesh), (func,)) do v, f
         # Should v get a nan2zero?
         v = Makie.apply_transform(f, v)
         p4d = to_ndim(Vec4f0, scale .* to_ndim(Vec3f0, v, 0f0), 1f0)
