+++ GLMakie/Project.toml
@@ -29,7 +29,7 @@ FixedPointNumbers = "0.7, 0.8"
 FreeTypeAbstraction = "0.8, 0.9"
 GLFW = "3"
 GeometryBasics = "0.3"
-Makie = "=0.14.1"
+Makie = "=0.14.2"
 MeshIO = "0.4"
 ModernGL = "1"
 Observables = "0.4"
