+++ WGLMakie/Project.toml
@@ -1,7 +1,7 @@
 authors = ["SimonDanisch <sdanisch@gmail.com>"]
 name = "WGLMakie"
 uuid = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
-version = "0.4.1"
+version = "0.4.2"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
