### 2021-06-27 6f5e421
remove figureposition and use normal gridpositions with parents
