+++ src/makielayout/layoutables.jl
@@ -51,11 +51,11 @@ end
 
 can_be_current_axis(x) = false
 
-get_top_parent(gp::GridLayoutBase.GridPosition) = GridLayoutBase.top_parent(gp.layout)
-get_top_parent(gp::GridLayoutBase.GridSubposition) = GridLayoutBase.top_parent(gp.parent)
+get_top_parent(gp::GridPosition) = GridLayoutBase.top_parent(gp.layout)
+get_top_parent(gp::GridSubposition) = GridLayoutBase.top_parent(gp.parent)
 
 function _layoutable(T::Type{<:Layoutable},
-        gp::Union{GridPosition, GridLayoutBase.GridSubposition}, args...; kwargs...)
+        gp::Union{GridPosition, GridSubposition}, args...; kwargs...)
 
     top_parent = get_top_parent(gp)
     if top_parent === nothing
