### 2021-06-27 e40e606
change all figurepositions to gridpositions etc
