+++ src/makielayout/MakieLayout.jl
@@ -17,6 +17,7 @@ using Match
 import Animations
 import PlotUtils
 using GridLayoutBase
+using GridLayoutBase: GridSubposition
 import Showoff
 using Colors
 
