+++ docs/src/plot_method_signatures.md
@@ -10,11 +10,11 @@ We use `Scatter` as our example, but the principles apply to every plot type.
 
 ## Non-Mutating
 
-The non-mutating methods create and return something in addition to the plot object, either a figure with an axis in default position, or an axis at a given figure position.
+The non-mutating methods create and return something in addition to the plot object, either a figure with an axis in default position, or an axis at a given GridPosition or GridSubposition.
 
 ```julia
 scatter(args...; kwargs...) -> ::FigureAxisPlot
-scatter(figureposition, args...; kwargs...) -> ::AxisPlot
+scatter(gridposition, args...; kwargs...) -> ::AxisPlot
 ```
 
 `FigureAxisPlot` is just a collection of the new figure, axis and plot.
@@ -22,7 +22,7 @@ For convenience it has the same display overload as `Figure`, so that `scatter(a
 It can be destructured at assignment like `fig, ax, plotobj = scatter(args...)`.
 
 `AxisPlot` is a collection of a new axis and plot.
-It has no special display overload but can also be destructured like `ax, plotobj = scatter(figureposition, args...)`.
+It has no special display overload but can also be destructured like `ax, plotobj = scatter(gridposition, args...)`.
 
 ### Special Keyword Arguments
 
@@ -58,17 +58,18 @@ If no figure is passed, the `current_figure()` is used, if no axis or scene is g
 ```julia
 scatter!(args...; kwargs...) -> ::Scatter
 scatter!(figure, args...; kwargs...) -> ::Scatter
-scatter!(figureposition, args...; kwargs...) -> ::Scatter
+scatter!(gridposition, args...; kwargs...) -> ::Scatter
 scatter!(axis, args...; kwargs...) -> ::Scatter
 scatter!(scene, args...; kwargs...) -> ::Scatter
 ```
 
-## FigurePositions
+## GridPositions
 
 In the background, each `Figure` has a `GridLayout` from [GridLayoutBase.jl](https://github.com/jkrumbiegel/GridLayoutBase.jl), which takes care of layouting plot elements nicely.
 For convenience, you can index into a figure multiple times to refer to nested grid positions, which makes it easy to quickly assemble complex layouts.
 
-For example, `fig[1, 2]` creates a `FigurePosition` referring to row 1 and column 2, while `fig[1, 2][3, 1:2]` creates a `FigureSubposition` that refers to row 3 and columns 1 to 2 in a nested GridLayout which is located at row 1 and column 2.
+For example, `fig[1, 2]` creates a `GridPosition` referring to row 1 and column 2, while `fig[1, 2][3, 1:2]` creates a `GridSubposition` that refers to row 3 and columns 1 to 2 in a nested GridLayout which is located at row 1 and column 2.
+The link to the Figure is in the parent field of the top layout.
 
 ### With Non-Mutating Plotting Functions
 

[…]
