+++ CairoMakie/test/runtests.jl
@@ -70,3 +70,5 @@ recorded = joinpath(@__DIR__, "recorded")
 rm(recorded; force=true, recursive=true); mkdir(recorded)
 ReferenceTests.record_tests(database; recording_dir=recorded)
 ReferenceTests.reference_tests(recorded)
+@show recorded
+@show readdir(recorded)
