+++ .github/workflows/cairomakie.yaml
@@ -56,7 +56,7 @@ jobs:
         uses: actions/upload-artifact@v2
         with:
           name: ReferenceImages
-          path: ./test/recorded
+          path: ./CairoMakie/test/recorded
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
