+++ .github/workflows/glmakie.yaml
@@ -56,10 +56,10 @@ jobs:
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
-          name: ReferenceImages
+          name: ReferenceImages_${{ matrix.os }}_${{ matrix.arch }}_${{ matrix.version }}
           path: |
-            ./test/recorded
-            ./test/recorded_glmakie
+            ./GLMakie/test/recorded
+            ./GLMakie/test/recorded_glmakie
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
