+++ .github/workflows/cairomakie.yaml
@@ -55,7 +55,7 @@ jobs:
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
-          name: ReferenceImages
+          name: ReferenceImages_${{ matrix.os }}_${{ matrix.arch }}_${{ matrix.version }}
           path: ./CairoMakie/test/recorded
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
