### 2021-06-27 d8adf8d
continue on error so artifacts still upload
