+++ .github/workflows/wglmakie.yaml
@@ -42,6 +42,7 @@ jobs:
           # dev mono repo versions
           pkg"dev . MakieCore WGLMakie ReferenceTests"
       - name: Run the tests
+        continue-on-error: true
         run: DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("WGLMakie", coverage=true)'
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
