+++ .github/workflows/glmakie.yaml
@@ -52,6 +52,7 @@ jobs:
           # dev mono repo versions
           pkg"dev . MakieCore GLMakie ReferenceTests"
       - name: Run the tests
+        continue-on-error: true
         run: DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("GLMakie", coverage=true)'
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
