### 2021-06-27 39093b5
fix test artifact creation (#1077)
* test why artifacts don't show up

* try different artifacts folder

* fix artifact names & os and version specific names
