+++ .github/workflows/cairomakie.yaml
@@ -55,8 +55,8 @@ jobs:
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
-          name: ReferenceImages
-          path: ./test/recorded
+          name: ReferenceImages_${{ matrix.os }}_${{ matrix.arch }}_${{ matrix.version }}
+          path: ./CairoMakie/test/recorded
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
