+++ CairoMakie/test/runtests.jl
@@ -3,6 +3,10 @@ using CairoMakie
 
 # Before changing Pkg environment, try the test in #864
 @testset "Runs without error" begin
+    open(joinpath(@__DIR__, "recorded", "testfile.txt"), "w") do file
+        println(file, "test")
+    end
+    error()
     fig = Figure()
     scatter(fig[1, 1], rand(10))
     fn = tempname()*".png"
@@ -70,6 +74,3 @@ recorded = joinpath(@__DIR__, "recorded")
 rm(recorded; force=true, recursive=true); mkdir(recorded)
 ReferenceTests.record_tests(database; recording_dir=recorded)
 ReferenceTests.reference_tests(recorded)
-@show pwd()
-@show recorded
-@show readdir(recorded)
