+++ .github/workflows/cairomakie.yaml
@@ -52,12 +52,19 @@ jobs:
           pkg"dev . MakieCore CairoMakie ReferenceTests"
       - name: Run the tests
         continue-on-error: true
-        run: julia --project=monorepo -e 'using Pkg; Pkg.test("CairoMakie", coverage=true)'
+        run: >
+          julia --project=monorepo -e 'using Pkg; Pkg.test("CairoMakie", coverage=true)'
+          && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV 
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
           name: ReferenceImages_${{ matrix.os }}_${{ matrix.arch }}_${{ matrix.version }}
           path: ./CairoMakie/test/recorded
+      - name: Fail after artifacts if tests failed
+        run: |
+         if ["$TESTS_SUCCESSFUL" != "true"]; then
+            exit 1
+         fi
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
