+++ .github/workflows/wglmakie.yaml
@@ -43,12 +43,17 @@ jobs:
           pkg"dev . MakieCore WGLMakie ReferenceTests"
       - name: Run the tests
         continue-on-error: true
-        run: DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("WGLMakie", coverage=true)'
+        run: >
+          DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("WGLMakie", coverage=true)'
+          && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV 
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
           name: ReferenceImages_${{ matrix.os }}_${{ matrix.arch }}_${{ matrix.version }}
           path: ./WGLMakie/test/recorded
+      - name: Fail after artifacts if tests failed
+        if: ${{ env.TESTS_SUCCESSFUL != 'true' }}
+        run: exit 1
       - uses: actions/cache@v1
         env:
           cache-name: cache-artifacts
