### 2021-06-28 e60cc72
upload artifacts even if tests fail (#1080)
