+++ CairoMakie/test/runtests.jl
@@ -70,6 +70,3 @@ recorded = joinpath(@__DIR__, "recorded")
 rm(recorded; force=true, recursive=true); mkdir(recorded)
 ReferenceTests.record_tests(database; recording_dir=recorded)
 ReferenceTests.reference_tests(recorded)
-@show pwd()
-@show recorded
-@show readdir(recorded)
