+++ .github/workflows/glmakie.yaml
@@ -53,7 +53,9 @@ jobs:
           pkg"dev . MakieCore GLMakie ReferenceTests"
       - name: Run the tests
         continue-on-error: true
-        run: DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("GLMakie", coverage=true)'
+        run: >
+          DISPLAY=:0 xvfb-run -s '-screen 0 1024x768x24' julia --project=monorepo -e 'using Pkg; Pkg.test("GLMakie", coverage=true)'
+          && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV 
       - name: Upload test Artifacts
         uses: actions/upload-artifact@v2
         with:
@@ -61,6 +63,9 @@ jobs:
           path: |
             ./GLMakie/test/recorded
             ./GLMakie/test/recorded_glmakie
+      - name: Fail after artifacts if tests failed
+        if: ${{ env.TESTS_SUCCESSFUL != 'true' }}
+        run: exit 1
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v1
         with:
