### 2021-06-28 b2d87a9
only update axis limits on real changes
