+++ src/basic_recipes/barplot.jl
@@ -211,7 +211,7 @@ function Makie.plot!(p::BarPlot)
         end
 
         # --------------------------------
-        # ----------- Labesl -------------
+        # ----------- Labels -------------
         # --------------------------------
 
         if !isnothing(bar_labels)
