### 2021-06-28 #3349 @gaelforget
> 👋 @gaelforget, please update us on how your review is going (this is an automated reminder).

Hi @Kevin-Mattheus-Moerman

I will work on the review & check-list this week. Have already been reading through the package docs which look really good now I feel
