### 2021-06-28 fc08bee
fix inlining videostream as text/html
