### 2021-06-29 e423cfa
treat negativity correctly + flip with (low, high)
