### 2021-06-29 c6aab1f
remove check that messed up plots
