### 2021-06-29 74c54c7
start plots type refactor
