### 2021-06-30 #3349 @fverdugo
Hi @jkrumbiegel @Kevin-Mattheus-Moerman

In this check box

>  - [ ] Repository: Is the source code for this software available at the repository url?

the link to the repo url points to [AbstractPlotting.jl](https://github.com/JuliaPlots/AbstractPlotting.jl) but I guess (based on previous comment by @jkrumbiegel ) that the repo I should review is [Makie.jl](https://github.com/JuliaPlots/Makie.jl). Another indicator is that  [AbstractPlotting.jl](https://github.com/JuliaPlots/AbstractPlotting.jl) has 90 stars vs [Makie.jl](https://github.com/JuliaPlots/Makie.jl) with 895 stars. So, I believe that the repo for this review should be [Makie.jl](https://github.com/JuliaPlots/Makie.jl)

Can you please confirm or clarify this?
