### 2021-06-30 #3349 @jkrumbiegel
Yes all code is in the Makie.jl monorepo now, while most of it was in AbstractPlotting.jl when I submitted the paper.
