### 2021-06-30 #3349 @fverdugo
Hi @jkrumbiegel @Kevin-Mattheus-Moerman  ! find my review below 👇 There are some rather small issues marked with **TODO**. @jkrumbiegel please fix (or clarify) them and I'll update my review accordingly.

## Review

> - [x] Repository: Is the source code for this software available at the repository url?

Yes, source code available at https://github.com/JuliaPlots/Makie.jl

> - [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

MIT license file available at https://github.com/JuliaPlots/Makie.jl/blob/master/LICENSE.md

> - [x] Contribution and authorship: Has the submitting author (@jkrumbiegel) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@SimonDanish (with +2K commits) and @jkrumbiegel (with +1K commits) are the 2 top contributors to the package.

> - [x] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

The package has more that 5.8K commits, 97 contributors, and 36 releases. This is clearly more than 3 person months as the minimum requested.

> - [ ] Installation: Does installation proceed as outlined in the documentation?

The package can be easily installed via the Julia package manager with the command `pkg> add GLMakie`. After this I am able to use Makie to generate plots. **TODO** However, the tests launched with `pkg> test GLMakie` fail with error message `ERROR: expected package ReferenceTests [d37af2e0] to be registered`.

> - [ ] Functionality: Have the functional claims of the software been confirmed?

**TODO**: I cannot run the code snipped given in the manuscript to generate Figure 1. The problem is that file `"airportlocations.csv"` is not given in the local folder provided by `assetpath()` nor in here https://github.com/JuliaPlots/Makie.jl/tree/master/assets. I can generate the rest of the subplots without any problem though.

> - [x] Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The paper shows how the package can deal with plots with 100M data points.

> - [x] A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The first page of the documentation webpage states "Makie is a high-performance, extendable, and multi-platform plotting ecosystem for the Julia programming language." which clearly shows the goal (high-performance, extendable and multi-platform plotting system) and the target audience (Julia programmers).

 > - [x] Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Installation is straight-forward via the Julia package manager `Pkg`. List of dependencies are defined the Julia-way via a `Project.toml` file.

> - [x] Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The documentation provides tutorials (e.g., https://makie.juliaplots.org/dev/basic-tutorial.html). Function signatures are also documented using numerous examples. In addition, the README.md file (https://github.com/JuliaPlots/Makie.jl/blob/master/README.md) contains some practical use cases.

> - [x] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Yes, see https://makie.juliaplots.org/dev/index.html .

> - [x] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

A complete test suite is in place to test the different parts of the library. Test are written via the testing tools provided by Julia and run automatically via Github Actions.

> - [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

**TODO** Some extra entries in the README.md file (https://github.com/JuliaPlots/Makie.jl/blob/master/README.md) detailing 1), 2) and 3) would be welcome.

> - [x] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Yes, sections "Summary" and "Overview" of the manuscript address this point.

> - [x] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

Yes. See section "Statement of need" of the manuscript.

> - [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?

The authors show how this package benefits from the Julia dispatch system (i.e., it has superior extendability), from Julia performance (i.e., it can handle larger data sets), and also show how it allows users to generate interactive plots in a more convenient way (i.e., reactive design). **TODO** Some comment on how Makie.jl compares with Plots.jl (https://github.com/JuliaPlots/Plots.jl), the other major plotting package in Julia, would be welcome. A citation to Plots.jl would be also welcome.

> - [x] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is well written.

> - [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Yes.
