### 2021-06-30 #3349 @SimonDanisch
>The package can be easily installed via the Julia package manager with the command pkg> add GLMakie. After this I am able to use Makie to generate plots. TODO However, the tests launched with pkg> test GLMakie fail with error message ERROR: expected package ReferenceTests [d37af2e0] to be registered.

Good point... Test are supposed to mainly run in the CI, so the local setup has been bitrotting. Should be easy to fix, although a bit annoying with the current monorepo setup...

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

That's a good idea in general.. I hope we can find some time in the next days to add those.
