+++ MakieCore/src/basic_plots/others.jl
@@ -1,0 +1,247 @@
+
+"""
+`calculated_attributes!(trait::Type{<: AbstractPlot}, plot)`
+trait version of calculated_attributes
+"""
+calculated_attributes!(trait, plot) = nothing
+
+"""
+`calculated_attributes!(plot::AbstractPlot)`
+Fill in values that can only be calculated when we have all other attributes filled
+"""
+calculated_attributes!(plot::T) where T = calculated_attributes!(T, plot)
+
+"""
+image(x, y, image)
+image(image)
+
+Plots an image on range `x, y` (defaults to dimensions).
+
+"""
+@recipe(Image, x, y, image) do scene
+Attributes(;
+    default_theme(scene)...,
+    colormap = [:black, :white],
+    colorrange = automatic,
+    interpolate = true,
+    fxaa = false,
+    lowclip = nothing,
+    highclip = nothing,
+    inspectable = theme(scene, :inspectable)
+)
+end
+
+"""
+heatmap(x, y, values)
+heatmap(values)
+
+Plots a heatmap as an image on `x, y` (defaults to interpretation as dimensions).
+
+"""
+@recipe(Heatmap, x, y, values) do scene
+Attributes(;
+    default_theme(scene)...,
+    colormap = theme(scene, :colormap),
+    colorrange = automatic,
+    linewidth = 0.0,
+    interpolate = false,
+    levels = 1,
+    fxaa = true,
+    lowclip = nothing,
+    highclip = nothing,
+    inspectable = theme(scene, :inspectable)
+)
+end
+
+"""
+volume(volume_data)
+
+Plots a volume. Available algorithms are:
+* `:iso` => IsoValue
+* `:absorption` => Absorption
+* `:mip` => MaximumIntensityProjection
+* `:absorptionrgba` => AbsorptionRGBA
+* `:additive` => AdditiveRGBA
+* `:indexedabsorption` => IndexedAbsorptionRGBA
+"""
+@recipe(Volume, x, y, z, volume) do scene
+Attributes(;
+    default_theme(scene)...,
+    algorithm = :mip,
+    isovalue = 0.5,
+    isorange = 0.05,
+    color = nothing,
+    colormap = theme(scene, :colormap),
+    colorrange = (0, 1),
+    fxaa = true,
+    inspectable = theme(scene, :inspectable)
+)
+end
+
+"""
+surface(x, y, z)
+
+Plots a surface, where `(x, y)`  define a grid whose heights are the entries in `z`.
+`x` and `y` may be `Vectors` which define a regular grid, **or** `Matrices` which define an irregular grid.
+
+"""
+@recipe(Surface, x, y, z) do scene
+Attributes(;
+    default_theme(scene)...,
+    color = nothing,
+    colormap = theme(scene, :colormap),
+    colorrange = automatic,
+    shading = true,
+    fxaa = true,
+    lowclip = nothing,
+    highclip = nothing,
+    invert_normals = false,
+    inspectable = theme(scene, :inspectable)
+)
+end
+
+"""
+lines(positions)
+lines(x, y)
+lines(x, y, z)
+
+Creates a connected line plot for each element in `(x, y, z)`, `(x, y)` or `positions`.
+
+!!! tip
+You can separate segments by inserting `NaN`s.
+"""
+@recipe(Lines, positions) do scene
+Attributes(;
+    default_theme(scene)...,
+    linewidth = theme(scene, :linewidth),
+    color = theme(scene, :linecolor),
+    colormap = theme(scene, :colormap),
+    colorrange = automatic,
+    linestyle = nothing,
+    fxaa = false,
+    cycle = [:color],
[…]
