+++ MakieCore/Project.toml
@@ -1,10 +1,11 @@
-authors = ["Simon Danisch"]
 name = "MakieCore"
 uuid = "20f20a25-4f0e-4fdf-b5d1-57303727442b"
+authors = ["Simon Danisch"]
 version = "0.1.3"
 
 [deps]
 Observables = "510215fc-4207-5dde-b226-833fc4488ee2"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 
 [compat]
 Observables = "0.4"
