### 2021-06-30 dfbf612
Only call each index once
