+++ src/basic_recipes/stairs.jl
@@ -1,3 +1,4 @@
+# adapted from https://gist.github.com/jkrumbiegel/529936a3a6f158715343601d9487880b
 @recipe(Stairs) do scene
     a = Attributes(
         step = :pre, # :center :post
@@ -13,29 +14,35 @@ function plot!(p::Stairs{<:Tuple{<:AbstractVector{<:Point2}}})
     steppoints = lift(points, p.step) do points, step
         if step == :pre
             s_points = Vector{Point2f0}(undef, length(points) * 2 - 1)
-            s_points[1] = points[1]
+            s_points[1] = point = points[1]
             for i in 1:length(points)-1
-                s_points[2i] = Point2f0(points[i][1], points[i+1][2])
-                s_points[2i + 1] = points[i+1]
+                nextpoint = points[i + 1]
+                s_points[2i] = Point2f0(point[1], nextpoint[2])
+                s_points[2i + 1] = nextpoint
+                point = nextpoint
             end
             s_points
         elseif step == :post
             s_points = Vector{Point2f0}(undef, length(points) * 2 - 1)
-            s_points[1] = points[1]
+            s_points[1] = point = points[1]
             for i in 1:length(points)-1
-                s_points[2i] = Point2f0(points[i+1][1], points[i][2])
-                s_points[2i + 1] = points[i+1]
+                nextpoint = points[i+1]
+                s_points[2i] = Point2f0(nextpoint[1], point[2])
+                s_points[2i + 1] = nextpoint
+                point = nextpoint
             end
             s_points
         elseif step == :center
             s_points = Vector{Point2f0}(undef, length(points) * 2)
-            s_points[1] = points[1]
-            s_points[end] = points[end]
+            s_points[1] = point = points[1]
             for i in 1:length(points)-1
-                halfx = 0.5 * (points[i][1] + points[i+1][1])
-                s_points[2i] = Point2f0(halfx, points[i][2])
-                s_points[2i + 1] = Point2f0(halfx, points[i+1][2])
+                nextpoint = points[i+1]
+                halfx = (point[1] + nextpoint[1]) / 2
+                s_points[2i] = Point2f0(halfx, point[2])
+                s_points[2i + 1] = Point2f0(halfx, nextpoint[2])
+                point = nextpoint
             end
+            s_points[end] = point
             s_points
         else
             error("Invalid step $step. Valid options are :pre, :post and :center")
