### 2021-07-05 970c458
Avoid same name for lowe and upper
This change avoids an error where lower and upper defined below as Node (lines 89-90) are change to Point2f0, resulting in: `type Array has no field val`
