+++ ReferenceTests/src/runtests.jl
@@ -4,7 +4,34 @@ function extract_frames(video, frame_folder)
     FFMPEG.ffmpeg_exe(`-loglevel quiet -i $video -y $path`)
 end
 
-function compare_media(a, b; sigma=[1,1], eps=0.02)
+
+function get_frames(a, b)
+    return (get_frames(a), get_frames(b))
+end
+
+function get_frames(video)
+    mktempdir() do folder
+        afolder = joinpath(folder, "a")
+        mkpath(afolder)
+        extract_frames(video, afolder)
+        aframes = joinpath.(afolder, readdir(afolder))
+        if length(aframes) > 10
+            # we don't want to compare too many frames since it's time costly
+            # so we just compare 10 random frames if more than 10
+            samples = range(1, stop=length(aframes), length=10)
+            istep = round(Int, length(aframes) / 10)
+            samples = 1:istep:length(aframes)
+            aframes = aframes[samples]
+        end
+        return load.(aframes)
+    end
+end
+
+function compare_media(a::Matrix, b::Matrix; sigma=[1,1])
+    Images.test_approx_eq_sigma_eps(a, b, sigma, Inf)
+end
+
+function compare_media(a, b; sigma=[1,1])
     file, ext = splitext(a)
     if ext in (".png", ".jpg", ".jpeg", ".JPEG", ".JPG")
         imga = load(a)
@@ -14,26 +41,10 @@ function compare_media(a, b; sigma=[1,1], eps=0.02)
             return Inf
         end
         conv(x) = convert(Matrix{RGBf0}, x)
-        return Images.test_approx_eq_sigma_eps(conv(imga), conv(imgb), sigma, Inf)
+        return compare_media(conv(imga), conv(imgb), sigma=sigma)
     elseif ext in (".mp4", ".gif")
-        mktempdir() do folder
-            afolder = joinpath(folder, "a")
-            bfolder = joinpath(folder, "b")
-            mkpath(afolder); mkpath(bfolder)
-            extract_frames(a, afolder)
-            extract_frames(b, bfolder)
-            aframes = joinpath.(afolder, readdir(afolder))
-            bframes = joinpath.(bfolder, readdir(bfolder))
-            if length(aframes) > 10
-                # we don't want to compare too many frames since it's time costly
-                # so we just compare 10 random frames if more than 10
-                samples = rand(1:length(aframes), 10)
-                aframes = aframes[samples]
-                bframes = bframes[samples]
-            end
-            # test by maximum diff
-            return mean(compare_media.(aframes, bframes; sigma=sigma, eps=eps))
-        end
+        aframes, bframes = get_frames(a, b)
+        return mean(compare_media.(aframes, bframes; sigma=sigma))
     else
         error("Unknown media extension: $ext")
     end
@@ -51,6 +62,8 @@ function compare(test_files::Vector{String}, reference_dir::String; missing_refi
         elseif isfile(test_path)
             if !isfile(ref_path)
                 push!(missing_refimages, test_path)
+            elseif endswith(test_path, ".html")
+                # ignore
             else
                 diff = compare_media(test_path, ref_path)
                 name = replace(ref_path, reference_dir => "")[2:end]
@@ -62,15 +75,32 @@ function compare(test_files::Vector{String}, reference_dir::String; missing_refi
     return missing_refimages, scores
 end
 
-# missing_imgs, scores = compare(recording_dir, joinpath(@__DIR__, "refimages"))
+function run_reference_tests(db, recording_folder, difference=0.03)
+    ref_images = download_refimages()
+    record_tests(db, recording_dir=recording_folder)
+    missing_files, scores = compare(joinpath.(recording_folder, readdir(recording_folder)), ref_images)
 
[…]
