+++ CairoMakie/test/runtests.jl
@@ -68,5 +68,4 @@ database = database_filtered(excludes, excludes2, functions=functions)
 
 recorded = joinpath(@__DIR__, "recorded")
 rm(recorded; force=true, recursive=true); mkdir(recorded)
-ReferenceTests.record_tests(database; recording_dir=recorded)
-ReferenceTests.reference_tests(recorded)
+ReferenceTests.run_reference_tests(database, recorded)
