+++ ReferenceTests/src/runtests.jl
@@ -75,8 +75,7 @@ function compare(test_files::Vector{String}, reference_dir::String; missing_refi
     return missing_refimages, scores
 end
 
-function run_reference_tests(db, recording_folder; difference=0.03)
-    ref_images = download_refimages()
+function run_reference_tests(db, recording_folder; difference=0.03, ref_images = download_refimages())
     record_tests(db, recording_dir=recording_folder)
     missing_files, scores = compare(joinpath.(recording_folder, readdir(recording_folder)), ref_images)
 
