+++ GLMakie/test/runtests.jl
@@ -25,7 +25,10 @@ include("glmakie_tests.jl")
 recorded_glmakie = joinpath(@__DIR__, "recorded_glmakie")
 rm(recorded_glmakie; force=true, recursive=true); mkdir(recorded_glmakie)
 ref_images = ReferenceTests.download_refimages(; name="glmakie_refimages")
-ReferenceTests.run_reference_tests(ReferenceTests.DATABASE, recorded_glmakie; ref_images=ref_images, difference=0.01)
+
+ReferenceTests.record_tests(ReferenceTests.DATABASE, recording_dir=recorded_glmakie)
+missing_files, scores = ReferenceTests.compare(joinpath.(recording_dir, readdir(recording_dir)), ref_images)
+ReferenceTests.reference_tests(scores; difference=0.01)
 # needs GITHUB_TOKEN to be defined
 # First look at the generated refimages, to make sure they look ok:
 # ReferenceTests.generate_test_summary("index_gl.html", recorded_glmakie)
