### 2021-07-08 eec6252
basic working changes for arbitrary glyphlayout
