### 2021-07-08 f9fdc00
fix basic text stuff for glyphlayout text parameter
