### 2021-07-08 d4e3f45
legend accepts abstractstring
