+++ src/makielayout/layoutables/legend.jl
@@ -1,6 +1,6 @@
 function layoutable(::Type{Legend},
         fig_or_scene,
-        entry_groups::Node{Vector{Tuple{Optional{String}, Vector{LegendEntry}}}};
+        entry_groups::Node{Vector{Tuple{Optional{<:AbstractString}, Vector{LegendEntry}}}};
         bbox = nothing, kwargs...)
 
     topscene = get_topscene(fig_or_scene)
@@ -319,7 +319,7 @@ legendelements(le::LegendElement, legend) = LegendElement[le]
 legendelements(les::AbstractArray{<:LegendElement}, legend) = LegendElement[les...]
 
 
-function LegendEntry(label::String, contentelements::AbstractArray, legend; kwargs...)
+function LegendEntry(label::AbstractString, contentelements::AbstractArray, legend; kwargs...)
     attrs = Attributes(label = label)
 
     kwargattrs = Attributes(kwargs)
@@ -329,7 +329,7 @@ function LegendEntry(label::String, contentelements::AbstractArray, legend; kwar
     LegendEntry(elems, attrs)
 end
 
-function LegendEntry(label::String, contentelement, legend; kwargs...)
+function LegendEntry(label::AbstractString, contentelement, legend; kwargs...)
     attrs = Attributes(label = label)
 
     kwargattrs = Attributes(kwargs)
@@ -461,8 +461,8 @@ end
     Legend(
         fig_or_scene,
         contents::AbstractArray,
-        labels::AbstractArray{String},
-        title::Optional{String} = nothing;
+        labels::AbstractArray{<:AbstractString},
+        title::Optional{<:AbstractString} = nothing;
         kwargs...)
 
 Create a legend from `contents` and `labels` where each label is associated to
@@ -472,8 +472,8 @@ one content element. A content element can be an `AbstractPlot`, an array of
 """
 function layoutable(::Type{Legend}, fig_or_scene,
         contents::AbstractArray,
-        labels::AbstractArray{String},
-        title::Optional{String} = nothing;
+        labels::AbstractArray{<:AbstractString},
+        title::Optional{<:AbstractString} = nothing;
         kwargs...)
 
     if length(contents) != length(labels)
@@ -494,7 +494,7 @@ end
         fig_or_scene,
         contentgroups::AbstractArray{<:AbstractArray},
         labelgroups::AbstractArray{<:AbstractArray},
-        titles::AbstractArray{<:Optional{String}};
+        titles::AbstractArray{<:Optional{<:AbstractString}};
         kwargs...)
 
 Create a multi-group legend from `contentgroups`, `labelgroups` and `titles`.
@@ -508,7 +508,7 @@ or any other object for which the `legendelements` method is defined.
 function layoutable(::Type{Legend}, fig_or_scene,
         contentgroups::AbstractArray{<:AbstractArray},
         labelgroups::AbstractArray{<:AbstractArray},
-        titles::AbstractArray{<:Optional{String}};
+        titles::AbstractArray{<:Optional{<:AbstractString}};
         kwargs...)
 
     if !(length(titles) == length(contentgroups) == length(labelgroups))
@@ -580,13 +580,13 @@ end
 # convenience constructor for axis legend
 axislegend(ax = current_axis(); kwargs...) = axislegend(ax, ax; kwargs...)
 
-axislegend(title::String; kwargs...) = axislegend(current_axis(), current_axis(), title; kwargs...)
+axislegend(title::AbstractString; kwargs...) = axislegend(current_axis(), current_axis(), title; kwargs...)
 
 """
     axislegend(ax, args...; position = :rt, kwargs...)
     axislegend(ax, args...; position = (1, 1), kwargs...)
[…]
