### 2021-07-08 dce29f6
make some changes to preprojected glyphs for glmakie
