+++ src/layouting/layouting.jl
@@ -223,8 +223,8 @@ end
 
 
 function preprojected_glyph_arrays(
-        string::String, position::VecTypes, glyphlayout::Makie.GlyphLayout3,
-        font, textsize, space::Symbol, projview, resolution, offset::VecTypes, transfunc
+        position::VecTypes, glyphlayout::Makie.GlyphLayout3,
+        space::Symbol, projview, resolution, offset::VecTypes, transfunc
     )
     offset = to_ndim(Point3f0, offset, 0)
     pos3f0 = to_ndim(Point3f0, position, 0)
@@ -237,12 +237,12 @@ function preprojected_glyph_arrays(
     else
         error("Unknown space $space, only :data or :screen allowed")
     end
-    text_quads(positions, string, font, textsize)
+    text_quads(positions, glyphlayout.glyphs, glyphlayout.fonts, glyphlayout.scales)
 end
 
 function preprojected_glyph_arrays(
-        string::String, position::VecTypes, glyphlayout::Makie.GlyphLayout3,
-        font, textsize, space::Symbol, projview, resolution, offsets::Vector, transfunc
+        position::VecTypes, glyphlayout::Makie.GlyphLayout3,
+        space::Symbol, projview, resolution, offsets::Vector, transfunc
     )
 
     offsets = to_ndim.(Point3f0, offsets, 0)
@@ -261,8 +261,7 @@ function preprojected_glyph_arrays(
 end
 
 function preprojected_glyph_arrays(
-        strings::AbstractVector{<:String}, positions::AbstractVector, glyphlayouts::Vector, font,
-        textsize, space::Symbol, projview, resolution, offset, transfunc
+        positions::AbstractVector, glyphlayouts::Vector, space::Symbol, projview, resolution, offset, transfunc
     )
 
     if offset isa VecTypes
@@ -295,50 +294,54 @@ function preprojected_glyph_arrays(
         error("Unknown space $space, only :data or :screen allowed")
     end
 
-    text_quads(allpos, strings, font, textsize)
+    text_quads(
+        reduce(vcat, allpos),
+        reduce(vcat, x.glyphs for x in glyphlayouts),
+        reduce(vcat, x.fonts for x in glyphlayouts),
+        reduce(vcat, x.scales for x in glyphlayouts))
 end
 
 
-function preprojected_glyph_arrays(
-        strings::AbstractVector{<:String}, positions::AbstractVector, glyphlayouts::Vector, font,
-        textsize, space::Symbol, projview, resolution, offsets::Vector{<: Vector}, transfunc
-    )
+# function preprojected_glyph_arrays(
+#         strings::AbstractVector{<:String}, positions::AbstractVector, glyphlayouts::Vector, font,
+#         textsize, space::Symbol, projview, resolution, offsets::Vector{<: Vector}, transfunc
+#     )
 
-    if space == :data
-        allpos = broadcast(positions, glyphlayouts, offsets) do pos, glyphlayout::Makie.GlyphLayout3, offsets
-            p = to_ndim(Point3f0, pos, 0)
-            apply_transform(
-                transfunc,
-                Point3f0[p .+ to_ndim(Point3f0, offset, 0) .+ o for (o, offset) in zip(glyphlayout.origins, offsets)]
-            )
-        end
-    elseif space == :screen
-        allpos = broadcast(positions, glyphlayouts, offsets) do pos, glyphlayout::Makie.GlyphLayout3, offsets
-            projected = to_ndim(
-                Point3f0,
-                Makie.project(
-                    projview,
-                    resolution,
-                    apply_transform(transfunc, to_ndim(Point3f0, pos, 0))
-                ),
[…]
