### 2021-07-08 ad46d93
update for mathtexengine as a package
