+++ CairoMakie/src/primitives.jl
@@ -308,49 +308,56 @@ function p3_to_p2(p::Point3{T}) where T
     end
 end
 
-function draw_atomic(scene::Scene, screen::CairoScreen, primitive::Text)
+function draw_atomic(scene::Scene, screen::CairoScreen, primitive::Text{<:Tuple{<:G}}) where G <: Union{AbstractArray{<:Makie.GlyphLayout3}, Makie.GlyphLayout3}
     ctx = screen.context
     @get_attribute(primitive, (textsize, color, font, rotation, model, space, offset))
-    txt = to_value(primitive[1])
     position = primitive.attributes[:position][]
     # use cached glyph info
-    glyphlayouts = primitive._glyphlayout[]
+    glyphlayout = to_value(primitive[1])
 
-    draw_string(scene, ctx, txt, position, glyphlayouts, textsize, color, font,
+    draw_glyphlayout(scene, ctx, position, glyphlayout, textsize, color, font,
         remove_billboard(rotation), model, space, offset)
 
     nothing
 end
 
-function draw_string(scene, ctx, strings::AbstractArray, positions::AbstractArray, glyphlayouts, textsize, color, font, rotation, model::SMatrix, space, offset)
+
+function draw_glyphlayout(scene, ctx, positions, glyphlayouts::AbstractArray, textsize, color, font, rotation, model::SMatrix, space, offset)
 
     # TODO: why is the Ref around model necessary? doesn't broadcast_foreach handle staticarrays matrices?
-    broadcast_foreach(strings, positions, glyphlayouts, textsize, color, font, rotation,
-        Ref(model), space, offset) do str, pos, glayout, ts, c, f, ro, mo, sp, off
+    broadcast_foreach(positions, glyphlayouts, textsize, color, font, rotation,
+        Ref(model), space, offset) do pos, glayout, ts, c, f, ro, mo, sp, off
 
-        draw_string(scene, ctx, str, pos, glayout, ts, c, f, ro, mo, sp, off)
+        draw_glyphlayout(scene, ctx, pos, glayout, ts, c, f, ro, mo, sp, off)
     end
 end
 
-function draw_string(scene, ctx, str::String, position::VecTypes, glyphlayout, textsize, color, font, rotation, model, space, offset)
+function draw_glyphlayout(scene, ctx, position, glyphlayout, textsize, color, font, rotation, model, space, offset)
 
     glyphoffsets = glyphlayout.origins
+    fonts = [Makie.convert_attribute(f, key"font"()) for f in glyphlayout.fonts]
+    rotations = [Makie.convert_attribute(r, key"rotation"()) for r in glyphlayout.rotations]
+    scales = glyphlayout.scales
 
     Cairo.save(ctx)
-    cairoface = set_ft_font(ctx, font)
     Cairo.set_source_rgba(ctx, rgbatuple(color)...)
-    old_matrix = get_font_matrix(ctx)
 
 
-    for (i, char) in enumerate(str)
+    for (i, char) in enumerate(glyphlayout.glyphs)
+
+        cairoface = set_ft_font(ctx, fonts[i])
+        old_matrix = get_font_matrix(ctx)
+
         goffset = glyphoffsets[i]
         if offset isa Vector
             p3_offset = to_ndim(Point3f0, offset[i], 0)
         else
             p3_offset = to_ndim(Point3f0, offset, 0)
         end
-        ts = textsize isa Vector ? textsize[i] : textsize
+        # ts = textsize isa Vector ? textsize[i] : textsize
+        ts = scales[i]
 
+        rotation = rotations[i]
 
         char in ('\r', '\n') && continue
 
[…]
