### 2021-07-08 80f7c43
update testscript for glmakie
