### 2021-07-08 0d7da33
add latexstrings and mathtexengine
