+++ CairoMakie/src/primitives.jl
@@ -308,67 +308,63 @@ function p3_to_p2(p::Point3{T}) where T
     end
 end
 
-function draw_atomic(scene::Scene, screen::CairoScreen, primitive::Text{<:Tuple{<:G}}) where G <: Union{AbstractArray{<:Makie.GlyphLayout5}, Makie.GlyphLayout5}
+function draw_atomic(scene::Scene, screen::CairoScreen, primitive::Text{<:Tuple{<:G}}) where G <: Union{AbstractArray{<:Makie.GlyphCollection2}, Makie.GlyphCollection2}
     ctx = screen.context
-    @get_attribute(primitive, (textsize, color, font, rotation, model, space, offset))
+    @get_attribute(primitive, (rotation, model, space, offset))
     position = primitive.attributes[:position][]
     # use cached glyph info
-    glyphlayout = to_value(primitive[1])
+    glyph_collection = to_value(primitive[1])
 
-    draw_glyphlayout(scene, ctx, position, glyphlayout, textsize, color, font,
-        remove_billboard(rotation), model, space, offset)
+    draw_glyph_collection(scene, ctx, position, glyph_collection, remove_billboard(rotation), model, space, offset)
 
     nothing
 end
 
 
-function draw_glyphlayout(scene, ctx, positions, glyphlayouts::AbstractArray, textsize, color, font, rotation, model::SMatrix, space, offset)
+function draw_glyph_collection(scene, ctx, positions, glyph_collections::AbstractArray, rotation, model::SMatrix, space, offset)
 
     # TODO: why is the Ref around model necessary? doesn't broadcast_foreach handle staticarrays matrices?
-    broadcast_foreach(positions, glyphlayouts, textsize, color, font, rotation,
-        Ref(model), space, offset) do pos, glayout, ts, c, f, ro, mo, sp, off
+    broadcast_foreach(positions, glyph_collections, rotation,
+        Ref(model), space, offset) do pos, glayout, ro, mo, sp, off
 
-        draw_glyphlayout(scene, ctx, pos, glayout, ts, c, f, ro, mo, sp, off)
+        draw_glyph_collection(scene, ctx, pos, glayout, ro, mo, sp, off)
     end
 end
 
-function draw_glyphlayout(scene, ctx, position, glyphlayout, textsize, color, font, rotation, model, space, offset)
+function draw_glyph_collection(scene, ctx, position, glyph_collection, rotation, model, space, offsets)
 
-    glyphoffsets = glyphlayout.origins
-    fonts = [Makie.convert_attribute(f, key"font"()) for f in glyphlayout.fonts]
-    rotations = [Makie.convert_attribute(r, key"rotation"()) for r in glyphlayout.rotations]
-    scales = glyphlayout.scales
+    glyphs = glyph_collection.glyphs
+    glyphoffsets = glyph_collection.origins
+    fonts = glyph_collection.fonts
+    rotations = glyph_collection.rotations
+    scales = glyph_collection.scales
+    colors = glyph_collection.colors
+    strokewidths = glyph_collection.strokewidths
+    strokecolors = glyph_collection.strokecolors
 
     Cairo.save(ctx)
-    Cairo.set_source_rgba(ctx, rgbatuple(color)...)
 
+    broadcast_foreach(glyphs, glyphoffsets, fonts, rotations, scales, colors, strokewidths, strokecolors, offsets) do glyph,
+        glyphoffset, font, rotation, scale, color, strokewidth, strokecolor, offset
 
-    for (i, char) in enumerate(glyphlayout.glyphs)
-
-        cairoface = set_ft_font(ctx, fonts[i])
+        cairoface = set_ft_font(ctx, font)
         old_matrix = get_font_matrix(ctx)
 
-        goffset = glyphoffsets[i]
[…]
