+++ src/Makie.jl
@@ -75,25 +75,47 @@ include("documentation/docstringextension.jl")
 include("utilities/quaternions.jl")
 include("interaction/PriorityObservable.jl")
 
+struct ScalarOrVector{T}
+    sv::Union{T, Vector{T}}
+end
+
+Base.convert(::Type{<:ScalarOrVector}, v::AbstractVector{T}) where T = ScalarOrVector{T}(collect(v))
+Base.convert(::Type{<:ScalarOrVector}, x::T) where T = ScalarOrVector{T}(x)
+Base.convert(::Type{<:ScalarOrVector{T}}, x::ScalarOrVector{T}) where T = x
 
 """
-    GlyphLayout5
+    GlyphCollection2
 
 Stores information about the glyphs in a string that had a layout calculated for them.
 """
-struct GlyphLayout5
+struct GlyphCollection2
     glyphs::Vector{Char}
     fonts::Vector{FTFont}
     origins::Vector{Point3f0}
     extents::Vector{FreeTypeAbstraction.FontExtent{Float32}}
-    scales::Vector{Vec2f0}
-    rotations::Vector{Quaternionf0}
+    scales::ScalarOrVector{Vec2f0}
+    rotations::ScalarOrVector{Quaternionf0}
+    colors::ScalarOrVector{RGBAf0}
+    strokecolors::ScalarOrVector{RGBAf0}
+    strokewidths::ScalarOrVector{Float32}
 
-    function GlyphLayout5(glyphs, fonts, origins, extents, scales, rotations)
-        @assert(length(glyphs) == length(fonts) == length(origins) == length(extents) == length(scales) == length(rotations))
+    function GlyphCollection2(glyphs, fonts, origins, extents, scales, rotations,
+            colors, strokecolors, strokewidths)
+
+        n = length(glyphs)
+        @assert length(fonts)  == n
+        @assert length(origins)  == n
+        @assert length(extents)  == n
+        @assert attr_broadcast_length(scales) in (n, 1)
+        @assert attr_broadcast_length(rotations)  in (n, 1)
+        @assert attr_broadcast_length(colors) in (n, 1)
+
         rotations = convert_attribute(rotations, key"rotation"())
         fonts = [convert_attribute(f, key"font"()) for f in fonts]
-        new(glyphs, fonts, origins, extents, scales, rotations)
+        colors = convert_attribute(colors, key"color"())
+        strokecolors = convert_attribute(strokecolors, key"color"())
+        strokewidths = Float32.(strokewidths)
+        new(glyphs, fonts, origins, extents, scales, rotations, colors, strokecolors, strokewidths)
     end
 end
 
@@ -341,13 +363,35 @@ export heatmap!, image!, lines!, linesegments!, mesh!, meshscatter!, scatter!, s
 
 
 
-function plot!(plot::Text{<:Tuple{LaTeXString}})
+function plot!(plot::Text{<:Tuple{<:Union{LaTeXString, AbstractVector{<:LaTeXString}}}})
     # attach a function to any text that calculates the glyph layout and stores it
-    lineels_glyphlayout_offset = lift(plot[1], plot.textsize, plot.align, plot.rotation, plot.model) do mytex, ts, al, rot, mo
+    lineels_glyphlayout_offset = lift(plot[1], plot.textsize, plot.align, plot.rotation,
+            plot.model, plot.color, plot.strokecolor, plot.strokewidth) do latexstring,
+                ts, al, rot, mo, color, scolor, swidth
+
+
         ts = to_textsize(ts)
         rot = to_rotation(rot)
+        col = to_color(color)
 
-        tex_elements, glyphlayout = texelems_and_glyph_positions(mytex, ts, al[1], al[2], rot)
[…]
