+++ src/layouting/data_limits.jl
@@ -113,7 +113,7 @@ FRect3D_from_point(p::VecTypes{2}) = FRect3D(Point3f0(p..., 0), Point3f0(0, 0, 0
 FRect3D_from_point(p::VecTypes{3}) = FRect3D(Point3f0(p...), Point3f0(0, 0, 0))
 
 
-function atomic_limits(x::Text{<:Tuple{<:GlyphLayout5}})
+function atomic_limits(x::Text{<:Tuple{<:GlyphCollection2}})
     if x.space[] == :data
         boundingbox(x)
     elseif x.space[] == :screen
@@ -123,15 +123,19 @@ function atomic_limits(x::Text{<:Tuple{<:GlyphLayout5}})
     end
 end
 
-function atomic_limits(x::Text{<:Tuple{<:AbstractArray{<:GlyphLayout5}}})
+function atomic_limits(x::Text{<:Tuple{<:AbstractArray{<:GlyphCollection2}}})
     if x.space[] == :data
         boundingbox(x)
     elseif x.space[] == :screen
-        bb = FRect3D_from_point(x.position[][1])
-        for p in x.position[][2:end]
-            bb = union(bb, FRect3D_from_point(p))
+        if isempty(x.position[])
+            FRect3D()
+        else
+            bb = FRect3D_from_point(x.position[][1])
+            for p in x.position[][2:end]
+                bb = union(bb, FRect3D_from_point(p))
+            end
+            bb
         end
-        bb
     else
         error()
     end

[…]
