### 2021-07-08 ec1030c
move import after quaternions
