+++ testscript_mathtexengine.jl
@@ -182,15 +182,15 @@ end
 with_theme() do
     f, ax , l = lines(cumsum(randn(1000)),
         axis = (
-            title = MyTex(raw"\sum_k{x y_k}"),
-            xlabel = MyTex(raw"\lim_{x →\infty} A^j v_{(a + b)_k}^i \sqrt{23.5} x!= \sqrt{\frac{1+6}{4+a+g}}\int_{0}^{2π} \sin(x) dx"),
-            ylabel = MyTex(raw"x + y - sin(x) × tan(y) + \sqrt{2}"),
+            title = L"\sum_k{x y_k}",
+            xlabel = L"\lim_{x →\infty} A^j v_{(a + b)_k}^i \sqrt{23.5} x!= \sqrt{\frac{1+6}{4+a+g}}\int_{0}^{2π} \sin(x) dx",
+            ylabel = L"x + y - sin(x) × tan(y) + \sqrt{2}",
         ),
         figure = (fontsize = 18, font = raw".\dev\MathTeXEngine\assets\fonts\NewCM10-Regular.otf")
     )
-    text!(MyTex(raw"\int_{0}^{2π} \sin(x) dx"), position = (500, 0))
+    text!(L"\int_{0}^{2π} \sin(x) dx", position = (500, 0))
 
-    Legend(f[1, 2], [l, l, l], [MyTex(raw"\sum{xy}"), MyTex(raw"a\int_0^5x^2+2ab"), "hello you!"])
+    Legend(f[1, 2], [l, l, l], [L"\sum{xy}", L"a\int_0^5x^2+2ab", "hello you!"])
     f
 end
 # save("test.pdf", f)
@@ -200,7 +200,7 @@ Makie.attribute_per_char(str::MyTex, attr) = attr
 ##
 s = Scene(camera = campixel!)
 t = text!(s,
-    MyTex(raw"\sqrt{2}"),
+    L"\sqrt{2}",
     # MyTex(raw"x + y - sin(x) × tan(y) + \sqrt{2}"),
     position = (50, 50),
     rotation = pi/2,
@@ -239,7 +239,7 @@ s
 
 s = Scene(camera = campixel!)
 t = text!(s,
-    MyTex(raw"\int_0^5x^2+2ab"),
+    L"\int_0^5x^2+2ab",
     position = Point2f0(50, 50),
     rotation = 0.0,
     show_axis = false,
@@ -248,7 +248,31 @@ wireframe!(s, boundingbox(t))
 s
 
 ##
-lines(cumsum(randn(99)), figure = (fontsize = 20, font = "Times"),
+
+lines(0..25, x -> sin(x) / (cos(3x) + 4), figure = (fontsize = 20, font = "Times"),
     axis = (
-        title = MyTex(raw"\int_0^1{x^2}∫"),
-        xlabel = MyTex(raw"\sum_k{x_k ⋅ y_k}"), ylabel = "you too"))
+        xticks = (0:50:100, [L"10^{-3.5}", L"10^{-4.5}", L"10^{-5.5}"]),
+        yticks = ([-1, 0, 1], [L"\sum_%$i{xy}" for i in 1:3]),
+        yticklabelrotation = pi/8,
+        title = L"\int_0^1{x^2}",
+        xlabel = L"\sum_k{x_k ⋅ y_k}",
+        ylabel = L"\int_a^b{\sqrt{abx}}"
+    )
+)
+text!(L"f(x) = \frac{sin(x)}{cos(3x) + 4}", position = (15, 2))
+current_figure()
+
+##
+
+text(Tuple{AbstractString, Point2f0}[(L"xy", Point2f0(0, 0)), (L"\int^2_3", Point2f0(20, 20))],
[…]
