### 2021-07-08 6aa1565
glyphlayout length check and font / rotation convert
