+++ src/basic_recipes/axis.jl
@@ -311,8 +311,8 @@ end
 function text_bb(str, font, size)
     rot = Quaternionf0(0,0,0,1)
     layout = layout_text(str, size, font, Vec2f0(0), rot, Mat4f0(I), 0.5, 1.0)
-    @assert typeof(layout.bboxes) <: Vector{FRect2D}
-    return data_text_boundingbox(str, layout, rot, Point3f0(0))
+    # @assert typeof(layout.bboxes) <: Vector{FRect2D}
+    return boundingbox(layout, Point3f0(0), rot)
 end
 
 
