### 2021-07-08 b43f11b
fix plot recursion bug
