### 2021-07-08 5eeb9c9
fixes for latex ticks
