### 2021-07-08 d03e801
make scalar examples run through, but with hacked attributes
