### 2021-07-08 b922c0d
fix variable length attributes
