### 2021-07-08 470ab2b
update test script
