### 2021-07-08 89b7690
add 3d axis test
