### 2021-07-08 e512ad5
fix textbox and rename almost all glyphlayouts
