### 2021-07-08 774504b
add latex tests
