+++ ReferenceTests/src/tests/text.jl
@@ -1,3 +1,5 @@
+using LaTeXStrings
+
 @cell "heatmap_with_labels" begin
     fig = Figure(resolution = (600, 600))
     ax = fig[1, 1] = Axis(fig)
@@ -245,4 +247,103 @@ end
         offset = (0, -10), color = :white, align = (:center, :top))
     tightlimits!(current_axis(), Bottom())
     current_figure()
+end
+
+
+@cell "latex strings" begin
+    f, ax , l = lines(cumsum(randn(1000)),
+        axis = (
+                title = L"\sum_k{x y_k}",
+                xlabel = L"\lim_{x →\infty} A^j v_{(a + b)_k}^i \sqrt{23.5} x!= \sqrt{\frac{1+6}{4+a+g}}\int_{0}^{2π} \sin(x) dx",
+                ylabel = L"x + y - sin(x) × tan(y) + \sqrt{2}",
+            ),
+        figure = (fontsize = 18, font = raw".\dev\MathTeXEngine\assets\fonts\NewCM10-Regular.otf")
+    )
+    text!(L"\int_{0}^{2π} \sin(x) dx", position = (500, 0))
+    Legend(f[1, 2], [l, l, l], [L"\sum{xy}", L"a\int_0^5x^2+2ab", L"||x-y||^2"])
+    f
+end
+
+@cell "latex simple" begin
+    s = Scene(camera = campixel!)
+    t = text!(s,
+        L"\sqrt{2}",
+        position = (50, 50),
+        rotation = pi/2,
+        show_axis = false,
+        space = :data)
+    s
+end
+
+@cell "latex bb" begin
+    s = Scene(camera = campixel!)
+    t = text!(s,
+        L"\int_0^5x^2+2ab",
+        position = Point2f0(50, 50),
+        rotation = 0.0,
+        show_axis = false,
+        space = :data)
+    wireframe!(s, boundingbox(t))
+    s
+end
+
+@cell "latex updates" begin
+    s = Scene(camera = campixel!)
+    st = Stepper(s)
+    textnode = Node([L"\int_0^5x^2+2ab", L"\int_0^5x^2+2ab"])
+    posnode = Node(Point2f0[(50, 50), (100, 100)])
+
+    t = text!(s,
+        textnode,
+        position = posnode,
+        rotation = 0.0,
+        show_axis = false,
+        space = :data)
+
+    Makie.step!(st)
+    ## change lengths
+    textnode.val = push!(textnode[], L"\int_0^5x^2+2ab")
+    posnode[] = push!(posnode[], Point2f0(150, 150))
+    Makie.step!(st)
+    st
+end
+
+@cell "update annotation style" begin
+    s = Scene(camera = campixel!)
+    st = Stepper(s)
+    textposnode = Node([
+        (L"\int_0^5x^2+2ab", Point2f0(50, 50)),
+        (L"\int_0^5x^2+2ab", Point2f0(100, 100)),
+    ])
+
+    t = text!(s,
+        textposnode,
+        show_axis = false,
+        space = :data)
+
+    Makie.step!(st)
+    ## change lengths
+    textposnode[] = push!(textposnode[], (L"\int_0^5x^2+2ab", Point2f0(150, 150)))
+    Makie.step!(st)
+    st
+end
+
+@cell "latex ticks" begin
+    lines(0..25, x -> 4 * sin(x) / (cos(3x) + 4), figure = (fontsize = 25, font = "Times"),
+        axis = (
+            xticks = (0:10:20, [L"10^{-3.5}", L"10^{-4.5}", L"10^{-5.5}"]),
[…]
