### 2021-07-08 25458b7
remove text char offsets test, not supported anymore
