### 2021-07-08 5a14b56
add legend example and remove string conversion for labels
