### 2021-07-08 76978a2
fix iswhitespace so it works on empty strings
