### 2021-07-08 2e8385e
add news to docs
