### 2021-07-08 4d10f0f
trigger calculation of ticklabel width once
