### 2021-07-08 851275b
fix inspector glyphcollection
