### 2021-07-08 028581a
jk/news (#1098)
* remove changelog of old makie, add news file
* add news to docs
