### 2021-07-08 bb4d2ab
Update NEWS.md
