### 2021-07-08 4ddebf4
remove mathtexengine testscript
