+++ testscript_mathtexengine.jl
@@ -1,312 +1,0 @@
-using CairoMakie: Cairo
-# using CairoMakie
-using GLMakie
-
-CairoMakie.activate!(type = "svg")
-
-
-##
-
-struct MyTex <: AbstractString
-    s::String
-end
-
-MakieLayout.iswhitespace(t::MyTex) = t.s == ""
-Base.isempty(::MyTex) = false
-
-##
-t = Node(MyTex("\\sum_k{3.002}"))
-let
-    s = Scene(camera = campixel!)
-    text!(s,
-        t,
-        position = (100, 100),
-        textsize = 20,
-        show_axis = false)
-    sl = Slider(s, bbox = BBox(10, 200, 10, 30))
-    on(sl.value) do v
-        t[] = MyTex("\\sum_k{$v} + \\int_{$v}^{$v}xyz")
-    end
-    notify(sl.value)
-    s
-end
-
-##
-let
-    s = Scene(camera = campixel!)
-    text!(s,
-        ["hello", "what's up"],
-        position = [(100, 100), (200, 200)],
-        textsize = 20,
-        space = :data,
-        show_axis = false)
-    s
-end
-##
-
-function Makie.plot!(plot::Makie.Text{<:Tuple{MyTex}})
-
-    # attach a function to any text that calculates the glyph layout and stores it
-    lineels_glyphlayout_offset = lift(plot[1], plot.textsize, plot.align, plot.rotation, plot.model) do mytex, ts, al, rot, mo
-        ts = to_textsize(ts)
-        rot = to_rotation(rot)
-
-        tex_elements, glyphlayout = texelems_and_glyph_collection(mytex, ts, al[1], al[2], rot)
-    end
-
-    glyphlayout = @lift($lineels_glyphlayout_offset[2])
-
-
-    linepairs = Node(Tuple{Point2f0, Point2f0}[])
-    linewidths = Node(Float32[])
-
-    onany(lineels_glyphlayout_offset, plot.position, plot.textsize, plot.rotation) do (allels, _, offs), pos, ts, rot
-
-        ts = to_textsize(ts)
-        rot = convert_attribute(rot, key"rotation"())
-
-        offset = Point2f0(pos)
-
-        els = map(allels) do el
-            if el[1] isa VLine
-                h = el[1].height
-                t = el[1].thickness * ts
-                pos = el[2]
-                size = el[3]
-                ps = (Point2f0(pos[1], pos[2]) .* ts, Point2f0(pos[1], pos[2] + h) .* ts) .- Ref(offs)
-                ps = Ref(rot) .* to_ndim.(Point3f0, ps, 0)
-                ps = Point2f0.(ps) .+ Ref(offset)
-                (ps, t)
-            elseif el[1] isa HLine
-                w = el[1].width
-                t = el[1].thickness * ts
-                pos = el[2]
-                size = el[3]
-                ps = (Point2f0(pos[1], pos[2]) .* ts, Point2f0(pos[1] + w, pos[2]) .* ts) .- Ref(offs)
-                ps = Ref(rot) .* to_ndim.(Point3f0, ps, 0)
-                ps = Point2f0.(ps) .+ Ref(offset)
-                (ps, t)
-            else
-                nothing
-            end
-        end
-        pairs = filter(!isnothing, els)
-        linewidths.val = repeat(last.(pairs), inner = 2)
-        linepairs[] = first.(pairs)
-        # @show linepairs
-    end
-
[…]
