### 2021-07-09 de831f4
fix cairomakie reversed heatmap +clean up tests
