### 2021-07-09 1fde3a1
add three-vector variant to surfacelike conversions
