### 2021-07-09 f04bd77
add examples to heatmap docs
