+++ docs/src/plotting_functions/heatmap.md
@@ -4,7 +4,7 @@
 heatmap
 ```
 
-### Examples
+## Two vectors and a matrix
 
 ```@example
 using CairoMakie
@@ -12,14 +12,40 @@ CairoMakie.activate!() # hide
 Makie.inline!(true) # hide
 
 
-f = Figure()
-Axis(f[1, 1])
-
-xs = LinRange(0, 10, 25)
-ys = LinRange(0, 15, 25)
+xs = range(0, 10, length = 25)
+ys = range(0, 15, length = 25)
 zs = [cos(x) * sin(y) for x in xs, y in ys]
 
-heatmap!(xs, ys, zs)
+heatmap(xs, ys, zs)
+```
 
-f
+## Two ranges and a function
+
+```@example
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+function mandelbrot(x, y)
+    z = c = x + y*im
+    for i in 1:30.0; abs(z) > 2 && return i; z = z^2 + c; end; 0
+end
+
+heatmap(-2:0.1:1, -1.1:0.1:1.1, mandelbrot,
+    colormap = Reverse(:deep))
+```
+
+## Three vectors
+
+```@example
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+
+xs = [1, 2, 3, 1, 2, 3, 1, 2, 3]
+ys = [1, 1, 1, 2, 2, 2, 3, 3, 3]
+zs = [1, 2, 3, 4, 5, 6, 7, 8, NaN]
+
+heatmap(xs, ys, zs)
 ```
