### 2021-07-09 c3f93b2
improve diff render + fix log10 heatmap test
