### 2021-07-09 22dab9f
fix log heatmaps
