### 2021-07-09 c8c973c
jk/three vector heatmap (#1101)
* add three-vector variant to surfacelike conversions

* add examples to heatmap docs

* add caveat

* add test

* tests for errors

* add news entry
