### 2021-07-09 4798e92
refactor RectAngleZoom, to be able to customize it
