### 2021-07-09 ff1b131
fix visibility + transparency in wglmakie
