+++ test/makielayout.jl
@@ -24,6 +24,7 @@ end
     sc = scatter!(ax, randn(100, 2))
     li = lines!(ax, randn(100, 2))
     hm = heatmap!(ax, randn(20, 20))
+    # axis contains 3 + 1 plots, one for the zoomrectangle
     @test length(ax.scene.plots) == 4
     delete!(ax, sc)
     @test length(ax.scene.plots) == 3
