### 2021-07-10 8a2543c
clean up code structure a bit
