+++ src/basic_recipes/text.jl
@@ -1,0 +1,255 @@
+function plot!(plot::Text)
+    # attach a function to any text that calculates the glyph layout and stores it
+    glyphcollection = lift(plot[1], plot.textsize, plot.font, plot.align,
+            plot.rotation, plot.model, plot.justification, plot.lineheight,
+            plot.color, plot.strokecolor, plot.strokewidth) do str,
+                ts, f, al, rot, mo, jus, lh, col, scol, swi
+        ts = to_textsize(ts)
+        f = to_font(f)
+        rot = to_rotation(rot)
+        col = to_color(col)
+        scol = to_color(scol)
+
+        layout_text(str, ts, f, al, rot, mo, jus, lh, col, scol, swi)
+    end
+
+    text!(plot, glyphcollection; plot.attributes...)
+
+    plot
+end
+
+# TODO: is this necessary? there seems to be a recursive loop with the above
+# function without these two interceptions, but I didn't need it before merging
+# everything into the monorepo...
+plot!(plot::Text{<:Tuple{<:GlyphCollection}}) = plot
+plot!(plot::Text{<:Tuple{<:AbstractArray{<:GlyphCollection}}}) = plot
+
+function plot!(plot::Text{<:Tuple{<:AbstractArray{<:AbstractString}}})
+
+    glyphcollections = Node(GlyphCollection[])
+    position = Node{Any}(nothing)
+    rotation = Node{Any}(nothing)
+    model = Node{Any}(nothing)
+    # offset = Node{Any}(nothing)
+
+    onany(plot[1], plot.textsize, plot.position,
+            plot.font, plot.align, plot.rotation, plot.model, plot.justification,
+            plot.lineheight, plot.color, plot.strokecolor, plot.strokewidth) do str,
+                    ts, pos, f, al, rot, mo, jus, lh, col, scol, swi
+
+        ts = to_textsize(ts)
+        f = to_font(f)
+        rot = to_rotation(rot)
+        col = to_color(col)
+        scol = to_color(scol)
+
+        gcs = GlyphCollection[]
+        broadcast_foreach(str, ts, f, al, rot, Ref(mo), jus, lh, col, scol, swi) do str,
+                ts, f, al, rot, mo, jus, lh, col, scol, swi
+            subgl = layout_text(str, ts, f, al, rot, mo, jus, lh, col, scol, swi)
+            push!(gcs, subgl)
+        end
+        position.val = pos
+        rotation.val = rot
+        model.val = mo
+        glyphcollections[] = gcs
+    end
+
+    # run onany once to initialize
+    notify(plot[1])
+
+    text!(plot, glyphcollections; position = plot.position, rotation = rotation,
+        model = model, offset = plot.offset, space = plot.space)
+
+    plot
+end
+
+# overload text plotting for a vector of tuples of a string and a point each
+function plot!(plot::Text{<:Tuple{<:AbstractArray{<:Tuple{<:AbstractString, <:Point}}}})
+    strings_and_positions = plot[1]
+
+    strings = Node(first.(strings_and_positions[]))
+    positions = Node(to_ndim.(Ref(Point3f0), last.(strings_and_positions[]), 0))
+
+    attrs = plot.attributes
+    pop!(attrs, :position)
+
+    t = text!(plot, strings; position = positions, attrs...)
+
+    # update both text and positions together
+    on(strings_and_positions) do str_pos
+        strs = first.(str_pos)
+        poss = to_ndim.(Ref(Point3f0), last.(str_pos), 0)
+        # first mutate strings without triggering redraw
+        t[1].val = strs
+        # then update positions with trigger
+        positions[] = poss
+    end
+    plot
+end
+
+
[…]
