+++ MakieCore/src/types.jl
@@ -1,17 +1,9 @@
 
-"""
-    abstract type Transformable
-This is a bit of a weird name, but all scenes and plots are transformable,
-so that's what they all have in common. This might be better expressed as traits.
-"""
-abstract type Transformable end
-
-abstract type AbstractPlot{Typ} <: Transformable end
-abstract type AbstractScene <: Transformable end
-abstract type ScenePlot{Typ} <: AbstractPlot{Typ} end
-abstract type AbstractScreen <: AbstractDisplay end
-
-const SceneLike = Union{AbstractScene, ScenePlot}
+const TorVector{T} = Union{T, AbstractVector{T}}
+const RealVector{T} = AbstractVector{T} where T <: Number
+const RGBAf0 = RGBA{Float32}
+const RGBf0 = RGB{Float32}
+const NativeFont = FreeTypeAbstraction.FTFont
 
 """
 Main structure for holding attributes, for theming plots etc!
@@ -30,6 +22,9 @@ struct Combined{Typ, T} <: ScenePlot{Typ}
     plots::Vector{AbstractPlot}
 end
 
+# all the plotting functions that get a plot type
+const PlotFunc = Union{Type{Any}, Type{<: AbstractPlot}}
+
 function Base.show(io::IO, plot::Combined)
     print(io, typeof(plot))
 end
@@ -41,6 +36,7 @@ macro key_str(arg)
     :(Key{$(QuoteNode(Symbol(arg)))})
 end
 Base.broadcastable(x::Key) = (x,)
+Key(sym) = Key{sym}()
 
 """
 Type to indicate that an attribute will get calculated automatically
@@ -62,11 +58,7 @@ So in the end, this is just relative to some normed camera - the value on screen
 will not actually sit on those pixels. Only camera that guarantees the correct mapping is the
 `:pixel` camera type.
 """
-struct Pixel{T} <: Unit{T}
-    value::T
-end
-
-const px = Pixel(1)
+@enum Space Pixel Data
 
 """
     Billboard([angle::Real])
@@ -81,3 +73,85 @@ end
 Billboard() = Billboard(0f0)
 Billboard(angle::Real) = Billboard(Float32(angle))
 Billboard(angles::Vector) = Billboard(Float32.(angles))
+
+@enum AntiAliasing FXAA NOAA MULTISAMPLE
+
+mutable struct Camera
+    pixel_area::Observable{Rect2D}
+    pixel_space::Observable{Mat4f0}
+    view::Observable{Mat4f0}
+    projection::Observable{Mat4f0}
+    projectionview::Observable{Mat4f0}
+    eyeposition::Observable{Vec3f0}
+    steering_nodes::Vector{Observables.ObserverFunction}
+end
+
+function Camera(pixel_area = Observable(Rect(0.0, 0.0, 0.0, 0.0)))
+    pixel_space = map(pixel_area) do window_size
+        nearclip = -10_000f0
+        farclip = 10_000f0
+        w, h = Float32.(widths(window_size))
+        return orthographicprojection(0f0, w, 0f0, h, nearclip, farclip)
+    end
+    Camera(
+        pixel_area,
+        pixel_space,
+        Observable(Mat4f0(I)),
+        Observable(Mat4f0(I)),
+        Observable(Mat4f0(I)),
+        Observable(Vec3f0(1)),
+        Observables.ObserverFunction[]
+    )
+end
+
+"""
+Holds the transformations for Scenes.
+"""
+struct Transformation <: Transformable
+    parent::RefValue{Transformable}
+    translation::Observable{Vec3f0}
+    scale::Observable{Vec3f0}
+    rotation::Observable{Quaternionf0}
[…]
