### 2021-07-10 f976071
CompatHelper: add new compat entry for "MathTeXEngine" at version "0.1"
