+++ Project.toml
@@ -73,6 +73,7 @@ Isoband = "0.1"
 KernelDensity = "0.5, 0.6"
 MakieCore = "0.1.3"
 Match = "1.1"
+MathTeXEngine = "0.1"
 Observables = "0.4"
 Packing = "0.4"
 PlotUtils = "1"
