### 2021-07-10 a1831eb
make scatter not error
