+++ test/projection_math.jl
@@ -1,3 +1,4 @@
+
 @testset "Projection math" begin
     @test eltype(Makie.rotationmatrix_x(1)) == Float64
     @test eltype(Makie.rotationmatrix_x(1f0)) == Float32
