### 2021-07-11 63bbb4d
add symlog and pseudolog
