+++ src/makielayout/layoutables/axis.jl
@@ -334,7 +334,6 @@ function layoutable(::Type{<:Axis}, fig_or_scene::Union{Figure, Scene}; bbox = n
         font = titlefont,
         color = titlecolor,
         space = :data,
-        show_axis=false,
         inspectable = false)
     decorations[:title] = titlet
 
