### 2021-07-12 7bf7648
prerender with better highlight.js and ansicolor
