### 2021-07-12 1e1a244
fix ```example typos
