### 2021-07-12 fd1fe44
prerender with better highlight.js and ansicolor (#1111)
* prerender with better highlight.js and ansicolor
* change highlight js to self-compiled fekre fork v11
* fix ```example typos
