### 2021-07-12 f264ad4
fix path for stepper (#1113)
* fix path for stepper
* fix missing fonts
