### 2021-07-12 b5c1697
Update NEWS.md
