+++ GLMakie/src/drawing_primitives.jl
@@ -183,9 +183,19 @@ function Base.insert!(screen::GLScreen, scene::Scene, @nospecialize(x::Combined)
     end
 end
 
-
-function Base.insert!(screen::GLScreen, scene::Scene, plot::Scatter)
+function Base.insert!(screen::GLScreen, scene::Scene, plot::AbstractPlot)
     draw_atomic(screen, scene, plot)
+end
+
+function add_to_screen!(screen, scene, robj)
+    for key in (:pixel_space, :view, :projection, :eyeposition, :projectionview)
+        if !haskey(robj.uniforms, key)
+            robj[key] = getfield(scene.camera, key)
+        end
+    end
+    robj[:resolution] = map(x-> Vec2f0(widths(x)), scene.camera.pixel_area)
+    push!(screen, scene, robj)
+    return robj
 end
 
 function draw_atomic(screen::GLScreen, scene::Scene, plot::Scatter)
@@ -202,20 +212,12 @@ function draw_atomic(screen::GLScreen, scene::Scene, plot::Scatter)
     gl_attributes[:uv_offset_width] = Vec4f0(0)
     gl_attributes[:use_pixel_marker] = lift(x-> x <: Pixel, plot.markerspace)
     positions = apply_transform(plot.basics.transformation.transform_func, plot.position)
-
     robj = visualize((plot[:marker], positions), Style(:default), gl_attributes)
-
-    for key in (:pixel_space, :view, :projection, :eyeposition, :projectionview)
-        if !haskey(robj.uniforms, key)
-            robj[key] = getfield(scene.camera, key)
-        end
-    end
-    robj[:resolution] = map(x-> Vec2f0(widths(x)), scene.camera.pixel_area)
-    push!(screen, scene, robj)
+    add_to_screen!(screen, scene, robj)
     return robj
 end
 
-function draw_atomic(screen::GLScreen, scene::Scene, @nospecialize(x::Union{Scatter, MeshScatter}))
+function draw_atomic(screen::GLScreen, scene::Scene, x::MeshScatter)
     robj = cached_robj!(screen, scene, x) do gl_attributes
         # signals not supported for shading yet
         gl_attributes[:shading] = to_value(get(gl_attributes, :shading, true))
@@ -245,131 +247,124 @@ function draw_atomic(screen::GLScreen, scene::Scene, @nospecialize(x::Union{Scat
     end
 end
 
-function draw_atomic(screen::GLScreen, scene::Scene, @nospecialize(x::Lines))
-    robj = cached_robj!(screen, scene, x) do gl_attributes
-        linestyle = pop!(gl_attributes, :linestyle)
-        data = Dict{Symbol, Any}(gl_attributes)
-        ls = to_value(linestyle)
-        if isnothing(ls)
-            data[:pattern] = ls
-        else
-            linewidth = gl_attributes[:thickness]
-            data[:pattern] = ls .* (to_value(linewidth) * 0.25)
-        end
-        positions = handle_view(x[1], data)
-        positions = apply_transform(transform_func_obs(x), positions)
-        handle_intensities!(data)
-        visualize(positions, Style(:lines), data)
+function draw_atomic(screen::GLScreen, scene::Scene, plot::Lines)
+    gl_attributes = Dict{Symbol, Any}()
+    linestyle = plot.linestyle
+    linewidth = plot.linewidth
+    @show linestyle[]
+    if isnothing(linestyle[])
+        gl_attributes[:pattern] = linestyle[]
+    else
+        gl_attributes[:pattern] = map((ls, lw)-> ls .* (lw * 0.25), linestyle, linewidth)
     end
+    gl_attributes[:thickness] = linewidth
[…]
