### 2021-07-13 e2fa4b5
conditional scene update
