### 2021-07-14 a4fb441
add rectplot to ax for custom interaction
