+++ Artifacts.toml
@@ -1,5 +1,5 @@
 [assets]
-git-tree-sha1 = "8f3a7c9e37a7dfeb46b1162f0767592f4a01b8f7"
+git-tree-sha1 = "ddf525c5a18d6f4fc17d11a2c362fb5fae0627b8"
 
     [[assets.download]]
     sha256 = "58dc1795f85407196ebfe87f98865d0905037b14d56c94748823ca626ffd8b2b"
