### 2021-07-14 1ae1377
add scatter example with airports
