### 2021-07-14 5418e3a
jk/airportlocations (#1122)
* move artifact generation script into folder with own env

* new artifact with airportlocations.csv

* add comments and small fixes

* change tree hash

* add scatter example with airports
