### 2021-07-15 97e70c9
Orthographic projection seems pretty unusable for that example
