### 2021-07-15 ed312d5
bump minor makie version, patch backend versions, tag news
