### 2021-07-15 40c356b
Update NEWS.md
