### 2021-07-16 7edd95b
try fixing text updates
