### 2021-07-17 16a4c32
add copy of docs and barebones franklin
