+++ docs_franklin/site/plotting_functions/scatter.md
@@ -1,0 +1,43 @@
+# scatter
+
+```julia:code
+scatter
+```
+
+### Examples
+
+\examplefigure{name = basic_scatter, svg = true
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+Axis(f[1, 1])
+
+xs = LinRange(0, 10, 20)
+ys = 0.5 .* sin.(xs)
+
+scatter!(xs, ys, color = :red)
+scatter!(xs, ys .- 1, color = xs)
+scatter!(xs, ys .- 2, markersize = LinRange(5, 30, 20))
+scatter!(xs, ys .- 3, marker = 'a':'t', strokewidth = 0, color = :black)
+
+f
+```
+}
+
+
+\examplefigure{
+```julia
+using CairoMakie
+using DelimitedFiles
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+a = readdlm(assetpath("airportlocations.csv"))
+
+scatter(a[1:50:end, :], marker = '✈',
+    markersize = 20, color = :black)
+```
+}
