+++ docs_franklin/site/plotting_functions/index.md
@@ -1,0 +1,1 @@
+{{list_plotting_functions}}
