+++ docs_franklin/site/Project.toml
@@ -1,0 +1,6 @@
+[deps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+Chain = "8be319e6-bccf-4806-a6f7-6fae938471bc"
+Franklin = "713c75ef-9fc9-4b05-94a9-213340da978e"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+NodeJS = "2bd173c7-0d6d-553b-b6af-13a54713934c"

[…]
