+++ docs_franklin/site/plotting_functions.md
@@ -1,1 +1,3 @@
+# Plotting functions!
+
 {{list_plotting_functions}}
