### 2021-07-18 7d2c9ed
specify how to add a title to a LScene
