### 2021-07-18 bc18631
fix projection matrix computation when using an orthographic projection
