### 2021-07-19 0d3d642
add basic docstrings with style
