+++ docs_franklin/site/plotting_functions/scatter.md
@@ -1,8 +1,6 @@
 # scatter
 
-```julia:code
-scatter
-```
+{{doc scatter}}
 
 ### Examples
 
