### 2021-07-19 a962c9c
add some more example plots
