+++ docs_franklin/site/plotting_functions/contour.md
@@ -1,0 +1,24 @@
+# contour
+
+{{doc contour}}
+
+### Examples
+
+\examplefigure{
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+Axis(f[1, 1])
+
+xs = LinRange(0, 10, 100)
+ys = LinRange(0, 15, 100)
+zs = [cos(x) * sin(y) for x in xs, y in ys]
+
+contour!(xs, ys, zs)
+
+f
+```
+}
