### 2021-07-19 708b032
clean up text functions
