+++ src/layouting/boundingbox.jl
@@ -142,45 +142,20 @@ end
 
 function boundingbox(x::Text{<:Tuple{<:GlyphCollection}})
     boundingbox(
-        to_value(x[1]),
-        to_ndim(Point3f0, to_value(x[:position]), 0),
-        convert_attribute(to_value(x[:rotation]), key"rotation"())
+        x[1][],
+        to_ndim(Point3f0, x.position[], 0),
+        to_rotation(x.rotation)
     )
 end
 
 function boundingbox(x::Text{<:Tuple{<:AbstractArray{<:GlyphCollection}}})
     boundingbox(
-        to_value(x[1]),
-        to_ndim.(Point3f0, to_value(x[:position]), 0),
-        convert_attribute(to_value(x[:rotation]), key"rotation"())
+        x[1][],
+        to_ndim.(Point3f0, x.position[], 0),
+        to_rotation(x.rotation[])
     )
 end
 
-
-#     if x.space[] == :data
-#         if x[1][] isa AbstractArray
-#             bb = FRect3D()
-#             for
-#         else
-#             boundingbox(x, to_value(x[1]))
-#         end
-#     elseif x.space[] == :screen
-#         data_limits(x)
-#     end
-# end
-
-
-# boundingbox(x::Text) = boundingbox(x, to_value(x[1]))
-
-# function boundingbox(
-#         text::String, position, textsize;
-#         font = "default", align = (:left, :bottom), rotation = 0.0
-#     )
-#     return boundingbox(
-#         text, position, textsize,
-#         to_font(font), to_align(align), to_rotation(rotation)
-#     )
-# end
 
 """
 Calculate an approximation of a tight rectangle around a 2D rectangle rotated by `angle` radians.
@@ -205,69 +180,4 @@ function rotatedrect(rect::Rect{2}, angle)
     rmaxs = maximum(rotated, dims = 2)
 
     return Rect2D(rmins..., (rmaxs .- rmins)...)
-end
-
-function quaternion_to_2d_angle(quat)
-    # this assumes that the quaternion was calculated from a simple 2d rotation as well
-    return 2acos(quat[4]) * (signbit(quat[1]) ? -1 : 1)
-end
-
-function boundingbox(
-        text::String, position, textsize, font,
-        align, rotation, model, justification, lineheight;
-        # use the font's ascenders and descenders for the bounding box
-        # this means that a string's boundingbox doesn't change in the vertical
-        # dimension when characters change (for example numbers during an animation)
-        # this is not wanted in most cases because of the jitter it creates when
-        # the boundingbox slightly changes size in each frame (in MakieLayout mostly)
-        use_vertical_dimensions_from_font = true
-    )
-    atlas = get_texture_atlas()
-    N = length(text)
-    ctext_state1 = iterate(text)
-    ctext_state1 === nothing && return FRect3D()
-
-    # call the layouting algorithm to find out where all the glyphs end up
-    # this is kind of a doubling, maybe it could be avoided if at creation all
-    # positions would be populated in the text object, but that seems convoluted
-    if position isa VecTypes
-        glyph_offsets = layout_text(text, textsize, font, align,
-            rotation, model, justification, lineheight)
-        position = [go .+ to_ndim(Point3f0, position, 0) for go in glyph_offsets]
-    end
-
-    bbox = Ref(FRect3D())            # use ref to avoid julia#15276
-    ctext_state = Ref{Union{Nothing,typeof(ctext_state1)}}(ctext_state1)
-
-    let position = position
-        broadcast_foreach(1:N, rotation, font, textsize) do i, rotation, font, scale
-            c, text_state = ctext_state[]::typeof(ctext_state1)
-            ctext_state[] = iterate(text, text_state)
-
-            if !(c in ('\r', '\n'))
[…]
