+++ src/basic_recipes/text.jl
@@ -1,16 +1,16 @@
 function plot!(plot::Text)
     # attach a function to any text that calculates the glyph layout and stores it
     glyphcollection = lift(plot[1], plot.textsize, plot.font, plot.align,
-            plot.rotation, plot.model, plot.justification, plot.lineheight,
+            plot.rotation, plot.justification, plot.lineheight,
             plot.color, plot.strokecolor, plot.strokewidth) do str,
-                ts, f, al, rot, mo, jus, lh, col, scol, swi
+                ts, f, al, rot, jus, lh, col, scol, swi
         ts = to_textsize(ts)
         f = to_font(f)
         rot = to_rotation(rot)
         col = to_color(col)
         scol = to_color(scol)
 
-        layout_text(str, ts, f, al, rot, mo, jus, lh, col, scol, swi)
+        layout_text(str, ts, f, al, rot, jus, lh, col, scol, swi)
     end
 
     text!(plot, glyphcollection; plot.attributes...)
@@ -44,9 +44,9 @@ function plot!(plot::Text{<:Tuple{<:AbstractArray{<:AbstractString}}})
         scol = to_color(scol)
 
         gcs = GlyphCollection[]
-        broadcast_foreach(str, ts, f, al, rot, Ref(mo), jus, lh, col, scol, swi) do str,
-                ts, f, al, rot, mo, jus, lh, col, scol, swi
-            subgl = layout_text(str, ts, f, al, rot, mo, jus, lh, col, scol, swi)
+        broadcast_foreach(str, ts, f, al, rot, jus, lh, col, scol, swi) do str,
+                ts, f, al, rot, jus, lh, col, scol, swi
+            subgl = layout_text(str, ts, f, al, rot, jus, lh, col, scol, swi)
             push!(gcs, subgl)
         end
         position.val = pos
