+++ GLMakie/src/drawing_primitives.jl
@@ -259,13 +259,29 @@ value_or_first(x::AbstractArray) = first(x)
 value_or_first(x::StaticArray) = x
 value_or_first(x) = x
 
+function collect_glyph_data(res, gcollection, projview, transfunc, pos, space, offset)
+    res = Vec2f0(widths(pixelarea(scene)[]))
+    preprojected_glyph_arrays(pos, gcollection, space, projview, res, offset, transfunc)
+end
+
+function collect_glyph_data(res, gcollection, projview, transfunc, pos, space, offset)
+    preprojected_glyph_arrays(pos, to_value(gcollection), space, projview, res, offset, transfunc)
+end
+
+
 function draw_atomic(screen::GLScreen, scene::Scene,
         x::Text{<:Tuple{<:Union{<:Makie.GlyphCollection, <:AbstractVector{<:Makie.GlyphCollection}}}})
 
     robj = cached_robj!(screen, scene, x) do gl_attributes
         glyphcollection = x[1]
-        liftkeys = (:position, :space, :offset)
-        args = getindex.(Ref(gl_attributes), liftkeys)
+
+
+        res = map(x->Vec2f0(widths(x)), pixelarea(scene))
+        projview = scene.camera.projectionview
+        transfunc =  Makie.transform_func_obs(scene)
+        pos = gl_attributes[:position]
+        space = gl_attributes[:space]
+        offset = gl_attributes[:offset]
 
         # TODO: This is a hack before we get better updating of plot objects and attributes going.
         # Here we only update the glyphs when the glyphcollection changes, if it's a singular glyphcollection.
@@ -274,20 +290,16 @@ function draw_atomic(screen::GLScreen, scene::Scene,
         # vector is triggered, but basically all internal objects use the vector of tuples version, and that triggers
         # both glyphcollection and position, so it still works
         if glyphcollection[] isa Makie.GlyphCollection
-            # here we lift the glyph collection
-            function collect_glyph_data(gcollection, projview, transfunc, pos, space, offset)
-                res = Vec2f0(widths(pixelarea(scene)[]))
-                preprojected_glyph_arrays(pos, gcollection, space, projview, res, offset, transfunc)
-            end
-            glyph_data = lift(collect_glyph_data, glyphcollection, scene.camera.projectionview, Makie.transform_func_obs(scene), args...)
+            # here we use the glyph collection observable directly
+            gcollection = glyphcollection
         else
-            # and here we don't because it triggers dimension mismatches
-            function collect_glyph_data(projview, transfunc, pos, space, offset)
-                gcollection = glyphcollection[]
-                res = Vec2f0(widths(pixelarea(scene)[]))
-                preprojected_glyph_arrays(pos, gcollection, space, projview, res, offset, transfunc)
-            end
-            glyph_data = lift(collect_glyph_data, scene.camera.projectionview, Makie.transform_func_obs(scene), args...)
+            # and here we wrap it into another observable
+            # so it doesn't triggers dimension mismatches
+            # the actual, new value gets then taken in the below lift with to_value
+            gcollection = Observable(glyphcollection)
+        end
+        glyph_data = lift(pos, gcollection, space, projview, res, offset, transfunc) do pos, gc, args...
+            preprojected_glyph_arrays(pos, to_value(gc), args...)
         end
 
         # unpack values from the one signal:
