+++ docs_franklin/site/plotting_functions/boxplot.md
@@ -4,7 +4,7 @@
 
 ### Examples
 
-\examplefigure{
+\begin{examplefigure}{}
 ```julia
 using CairoMakie
 CairoMakie.activate!() # hide
@@ -15,9 +15,9 @@ ys = randn(1000)
 
 boxplot(xs, ys)
 ```
-}
+\end{examplefigure}
 
-\examplefigure{
+\begin{examplefigure}{}
 ```julia
 using CairoMakie
 CairoMakie.activate!() # hide
@@ -29,7 +29,7 @@ dodge = rand(1:2, 1000)
 
 boxplot(xs, ys, dodge = dodge, show_notch = true, color = dodge)
 ```
-}
+\end{examplefigure}
 
 Colors are customizable. The `color` attribute refers to the color of the boxes, whereas
 `outliercolor` refers to the color of the outliers. If not scalars (e.g. `:red`), these attributes
@@ -40,7 +40,7 @@ same color as their box, as shown above.
     For all indices corresponding to points within the same box, `color` (but not `outliercolor`)
     must have the same value.
 
-\examplefigure{
+\begin{examplefigure}{}
 ```julia
 using CairoMakie
 CairoMakie.activate!() # hide
@@ -52,4 +52,4 @@ dodge = rand(1:2, 1000)
 
 boxplot(xs, ys, dodge = dodge, show_notch = true, color = map(d->d==1 ? :blue : :red, dodge) , outliercolor = rand([:red, :green, :blue, :black, :yellow], 1000))
 ```
-}
+\end{examplefigure}
