### 2021-07-19 ea3bb58
other plotting functions
