+++ docs_franklin/site/plotting_functions/hist.md
@@ -1,0 +1,35 @@
+# hist
+
+{{doc hist}}
+
+### Examples
+
+\begin{examplefigure}{}
+```julia
+using GLMakie
+GLMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+data = randn(1000)
+
+f = Figure()
+hist(f[1, 1], data, bins = 10)
+hist(f[1, 2], data, bins = 20, color = :red, strokewidth = 1, strokecolor = :black)
+hist(f[2, 1], data, bins = [-5, -2, -1, 0, 1, 2, 5], color = :gray)
+hist(f[2, 2], data, normalization = :pdf)
+f
+```
+\end{examplefigure}
+
+#### Histogram with labels
+
+You can use all the same arguments as [`barplot`](@ref):
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!()
+hist(data, normalization = :pdf, bar_labels = :values,
+     label_formatter=x-> round(x, digits=2), label_size = 15,
+     strokewidth = 0.5, strokecolor = (:black, 0.5), color = :values)
+```
+\end{examplefigure}
