+++ docs_franklin/site/plotting_functions/meshscatter.md
@@ -1,0 +1,19 @@
+# meshscatter
+
+{{doc meshscatter}}
+
+### Examples
+
+\begin{examplefigure}{}
+```julia
+using GLMakie
+GLMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+xs = cos.(1:0.5:20)
+ys = sin.(1:0.5:20)
+zs = LinRange(0, 3, length(xs))
+
+meshscatter(xs, ys, zs, markersize = 0.1, color = zs)
+```
+\end{examplefigure}
