+++ docs_franklin/site/plotting_functions/streamplot.md
@@ -1,0 +1,31 @@
+# streamplot
+
+{{doc streamplot}}
+
+### Examples
+
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+struct FitzhughNagumo{T}
+    ϵ::T
+    s::T
+    γ::T
+    β::T
+end
+
+P = FitzhughNagumo(0.1, 0.0, 1.5, 0.8)
+
+f(x, P::FitzhughNagumo) = Point2f0(
+    (x[1]-x[2]-x[1]^3+P.s)/P.ϵ,
+    P.γ*x[1]-x[2] + P.β
+)
+
+f(x) = f(x, P)
+
+streamplot(f, -1.5..1.5, -1.5..1.5, colormap = :magma)
+```
+\end{examplefigure}

[…]
