+++ docs_franklin/site/tutorials/basic-tutorial.md
@@ -1,0 +1,396 @@
+# Basic Tutorial
+
+## Preface
+
+Here is a quick tutorial to get you started with Makie!
+
+We assume you have [Julia](https://julialang.org/) and `CairoMakie.jl` (or one of the other backends, i.e. `GLMakie.jl` or `WGLMakie.jl`) installed already.
+
+This tutorial uses CairoMakie, but the code can be executed with any backend.
+CairoMakie can output beautiful static vector graphics, but it doesn't have the native ability to open interactive windows.
+
+To see the output of plotting commands when using CairoMakie, we recommend you either use an IDE which supports png or svg output, such as VSCode, Atom/Juno, Jupyter, Pluto, etc., or try using a viewer package such as [ElectronDisplay.jl](https://github.com/queryverse/ElectronDisplay.jl), or alternatively save your plots to files directly.
+The Julia REPL by itself does not have the ability to show plots.
+
+GLMakie can open interactive windows, or alternatively display bitmaps inline if `Makie.inline!(true)` is called.
+
+WGLMakie shows interactive plots in environments that support interactive html displays, such as VSCode, Atom/Juno, Jupyter, Pluto, etc.
+
+For more information, have a look at [Backends & Output](@ref).
+
+Ok, now that this is out of the way, let's get started!
+
+## First plot
+
+First, we import CairoMakie.
+
+\begin{examplefigure}{}
+```julia
+
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+nothing # hide
+```
+\end{examplefigure}
+
+Makie has many different plotting functions, one of the most common ones is [lines](@ref).
+You can just call such a function and your plot will appear if your coding environment can show png or svg files.
+
+!!! note
+    Objects such as [Figure](@ref), `FigureAxisPlot` or `Scene` are usually displayed whenever they are returned in global scope (e.g. in the REPL).
+    To display such objects from within a local scope, like from within a function, you can directly call `display(figure)`, for example.
+
+\begin{examplefigure}{}
+```julia
+
+x = range(0, 10, length=100)
+y = sin.(x)
+lines(x, y)
+```
+\end{examplefigure}
+
+Another common function is [scatter](@ref).
+
+\begin{examplefigure}{}
+```julia
+
+using CairoMakie
+
+x = range(0, 10, length=100)
+y = sin.(x)
+scatter(x, y)
+```
+\end{examplefigure}
+
+## Multiple plots
+
+Every plotting function has a version with and one without `!`.
+For example, there's `scatter` and `scatter!`, `lines` and `lines!`, etc.
+The functions without a `!` always create a new axis with a plot inside, while the functions with `!` plot into an already existing axis.
+
+Here's how you could plot two lines on top of each other.
+
+\begin{examplefigure}{}
+```julia
+
+using CairoMakie
+
+x = range(0, 10, length=100)
+y1 = sin.(x)
+y2 = cos.(x)
+
+lines(x, y1)
+lines!(x, y2)
+current_figure()
+```
+\end{examplefigure}
+
+The second `lines!` call plots into the axis created by the first `lines` call.
+If you don't specify an axis to plot into, it's as if you had called `lines!(current_axis(), ...)`.
+
+The call to `current_figure` is necessary here, because functions with `!` return only the newly created plot object, but this alone does not cause the figure to display when returned.
+
+## Attributes
+
+Every plotting function has attributes which you can set through keyword arguments.
+The lines in the previous example both have the same default color, which we can change easily.
+
+\begin{examplefigure}{}
+```julia
+
[…]
