+++ src/camera/old_camera3d.jl
@@ -4,9 +4,9 @@
 struct OldCamera3D <: AbstractCamera
     rotationspeed::Node{Float32}
     translationspeed::Node{Float32}
-    eyeposition::Node{Vec3f0}
-    lookat::Node{Vec3f0}
-    upvector::Node{Vec3f0}
+    eyeposition::Node{Vec3f}
+    lookat::Node{Vec3f}
+    upvector::Node{Vec3f}
     fov::Node{Float32}
     near::Node{Float32}
     far::Node{Float32}
@@ -28,9 +28,9 @@ function old_cam3d_cad!(scene; kw_args...)
         Attributes(
             rotationspeed = 0.01,
             translationspeed = 1.0,
-            eyeposition = Vec3f0(3),
-            lookat = Vec3f0(0),
-            upvector = Vec3f0(0, 0, 1),
+            eyeposition = Vec3f(3),
+            lookat = Vec3f(0),
+            upvector = Vec3f(0, 0, 1),
             fov = 45f0,
             near = 0.01f0,
             far = 100f0,
@@ -64,9 +64,9 @@ function old_cam3d_turntable!(scene; kw_args...)
         Attributes(
             rotationspeed = 0.01,
             translationspeed = 1.0,
-            eyeposition = Vec3f0(3),
-            lookat = Vec3f0(0),
-            upvector = Vec3f0(0, 0, 1),
+            eyeposition = Vec3f(3),
+            lookat = Vec3f(0),
+            upvector = Vec3f(0, 0, 1),
             fov = 45f0,
             near = 0.01f0,
             far = 100f0,
@@ -99,7 +99,7 @@ the plot's axis.
 const old_cam3d! = old_cam3d_turntable!
 
 function projection_switch(
-        wh::Rect2D,
+        wh::Rect2,
         fov::T, near::T, far::T,
         projectiontype::ProjectionEnum, zoom::T
     ) where T <: Real
@@ -138,7 +138,7 @@ end
 
 # TODO switch button and key because this is the wrong order
 function add_translation!(scene, cam, key, button, zoom_shift_lookat::Bool)
-    last_mousepos = RefValue(Vec2f0(0, 0))
+    last_mousepos = RefValue(Vec2f(0, 0))
     dragging = RefValue(false)
     on(camera(scene), scene.events.mousebutton) do event
         if event.button == key[] && ispressed(scene, button[])
@@ -151,7 +151,7 @@ function add_translation!(scene, cam, key, button, zoom_shift_lookat::Bool)
                 dragging[] = false
                 diff = (last_mousepos[] - mousepos) * cam.translationspeed[]
                 last_mousepos[] = mousepos
-                translate_cam!(scene, cam, Vec3f0(0f0, diff[1], diff[2]))
+                translate_cam!(scene, cam, Vec3f(0f0, diff[1], diff[2]))
                 return Consume(true)
             end
         end
@@ -163,7 +163,7 @@ function add_translation!(scene, cam, key, button, zoom_shift_lookat::Bool)
             mousepos = screen_relative(scene, mp)
             diff = (last_mousepos[] .- mousepos) * cam.translationspeed[]
             last_mousepos[] = mousepos
-            translate_cam!(scene, cam, Vec3f0(0f0, diff[1], diff[2]))
+            translate_cam!(scene, cam, Vec3f(0f0, diff[1], diff[2]))
             return Consume(true)
         end
         return Consume(false)
[…]
