+++ src/basic_recipes/text.jl
@@ -69,7 +69,7 @@ function plot!(plot::Text{<:Tuple{<:AbstractArray{<:Tuple{<:AbstractString, <:Po
     strings_and_positions = plot[1]
 
     strings = Node(first.(strings_and_positions[]))
-    positions = Node(to_ndim.(Ref(Point3f0), last.(strings_and_positions[]), 0))
+    positions = Node(to_ndim.(Ref(Point3f), last.(strings_and_positions[]), 0))
 
     attrs = plot.attributes
     pop!(attrs, :position)
@@ -79,7 +79,7 @@ function plot!(plot::Text{<:Tuple{<:AbstractArray{<:Tuple{<:AbstractString, <:Po
     # update both text and positions together
     on(strings_and_positions) do str_pos
         strs = first.(str_pos)
-        poss = to_ndim.(Ref(Point3f0), last.(str_pos), 0)
+        poss = to_ndim.(Ref(Point3f), last.(str_pos), 0)
         # first mutate strings without triggering redraw
         t[1].val = strs
         # then update positions with trigger
@@ -102,7 +102,7 @@ function plot!(plot::Text{<:Tuple{<:Union{LaTeXString, AbstractVector{<:LaTeXStr
         if latexstring isa AbstractVector
             tex_elements = []
             glyphcollections = GlyphCollection[]
-            offsets = Point2f0[]
+            offsets = Point2f[]
             broadcast_foreach(latexstring, ts, al, rot, color, scolor, swidth) do latexstring,
                 ts, al, rot, color, scolor, swidth
 
@@ -122,7 +122,7 @@ function plot!(plot::Text{<:Tuple{<:Union{LaTeXString, AbstractVector{<:LaTeXStr
     glyphcollection = @lift($lineels_glyphcollection_offset[2])
 
 
-    linepairs = Node(Tuple{Point2f0, Point2f0}[])
+    linepairs = Node(Tuple{Point2f, Point2f}[])
     linewidths = Node(Float32[])
 
     scene = Makie.parent_scene(plot)
@@ -146,7 +146,7 @@ function plot!(plot::Text{<:Tuple{<:Union{LaTeXString, AbstractVector{<:LaTeXStr
             allels = [allels]
         end
         broadcast_foreach(allels, offs, pos, ts, rot) do allels, offs, pos, ts, rot
-            offset = Point2f0(pos)
+            offset = Point2f(pos)
 
             els = map(allels) do el
                 el[1] isa VLine || el[1] isa HLine || return nothing
@@ -156,16 +156,16 @@ function plot!(plot::Text{<:Tuple{<:Union{LaTeXString, AbstractVector{<:LaTeXStr
 
                 ps = if el[1] isa VLine
                     h = el[1].height
-                    (Point2f0(p[1], p[2]) .* ts, Point2f0(p[1], p[2] + h) .* ts) .- Ref(offs)
+                    (Point2f(p[1], p[2]) .* ts, Point2f(p[1], p[2] + h) .* ts) .- Ref(offs)
                 else
                     w = el[1].width
-                    (Point2f0(p[1], p[2]) .* ts, Point2f0(p[1] + w, p[2]) .* ts) .- Ref(offs)
+                    (Point2f(p[1], p[2]) .* ts, Point2f(p[1] + w, p[2]) .* ts) .- Ref(offs)
                 end
-                ps = Ref(rot) .* to_ndim.(Point3f0, ps, 0)
+                ps = Ref(rot) .* to_ndim.(Point3f, ps, 0)
                 # TODO the points need to be projected to work inside Axis
                 # ps = project ps with projview somehow
 
[…]
