+++ src/makielayout/layoutables/axis3d.jl
@@ -28,7 +28,7 @@ function layoutable(::Type{<:Axis3}, fig_or_scene::Union{Figure, Scene}; bbox = 
 
     finallimits = Node(Rect3f(Vec3f(0f0, 0f0, 0f0), Vec3f(100f0, 100f0, 100f0)))
 
-    scenearea = lift(round_to_IRect2D, layoutobservables.computedbbox)
+    scenearea = lift(round_to_Rect2i, layoutobservables.computedbbox)
 
     scene = Scene(topscene, scenearea, raw = true, clear = false, backgroundcolor = attrs.backgroundcolor)
 
