+++ docs_franklin/site/tutorials.md
@@ -1,0 +1,4 @@
+@def title = "Tutorials"
+
+- [Basic Tutorial](basic-tutorial)
+- [Layout Tutorial](layout-tutorial)
