+++ docs_franklin/site/makielayout/gridlayout.md
@@ -1,0 +1,295 @@
+```@eval
+using CairoMakie
+CairoMakie.activate!()
+```
+
+# GridLayout
+
+## Setting column and row sizes correctly
+
+There are four different types of sizes you can give rows and columns.
+
+### Fixed
+
+`Fixed(scene_units)` is used to set a column or row to an absolute size, independent of its content.
+This only really makes sense if there is variable width content in the column or row, that can shrink or expand to meet this size. You will probably not need `Fixed` sizes very often.
+
+```@example
+using CairoMakie
+
+scene, layout = layoutscene(resolution = (1200, 900))
+
+layout[1, 1] = Axis(scene, title = "My column has size Fixed(400)")
+layout[1, 2] = Axis(scene, title = "My column has size Auto()")
+
+colsize!(layout, 1, Fixed(400))
+# colsize!(layout, 1, 400) would also work
+
+scene
+```
+
+### Relative
+
+`Relative(fraction)` is used to set a column or row to a size that is a certain fraction of the available width or height.
+This is useful, e.g., if you want a column to span 50% of the available width, no matter what other content is there.
+In this case, you would use `Relative(1/2)`. The available width is the width of the GridLayout minus the space taken by row or column gaps including protrusions.
+
+```@example
+using CairoMakie
+
+scene, layout = layoutscene(resolution = (1200, 900))
+
+layout[1, 1] = Axis(scene, title = "My column has size Relative(2/3)")
+layout[1, 2] = Axis(scene, title = "My column has size Auto()")
+layout[1, 3] = Colorbar(scene)
+
+colsize!(layout, 1, Relative(2/3))
+
+scene
+```
+
+
+### Auto
+
+The `Auto` size is a bit more complex to understand. It has two parameters, the Boolean `trydetermine` and the number `ratio`. The default is `Auto() == Auto(true, 1)`. This is also the default row height and column width.
+
+A column or row that is sized `Auto(true)` tries to fit its own size to its content.
+
+When a `GridLayout` is solved, it looks at the content in each row or column and checks if it reports a fixed size.
+Many objects can report their own width or height because their content has a specific size, such as `Label`.
+Other objects are often used with a user-set width or height, for example `Colorbar(scene, width = 30)`.
+In this case, the `GridLayout` can also see that the colorbar has a width of 30 units and fit the column width to that value.
+Objects like `Axis` on the other hand are usually not set to a specific size.
+
+Only objects that span a *single* row or column report their width or height, respectively. If *multiple* rows or columns are spanned, it's not well defined how the space that the object needs should be distributed.
+
+If there is more than one object with a fixed width or height in an `Auto(true)` sized column, the maximum size is used.
+
+If a column or row is sized `Auto(false)`, fixed-size objects are ignored. It can also happen of course, that there is no fixed-size object in a row or column with `Auto(true)` size. In both these cases, columns or rows determine their size by what remains after all `Fixed`, `Relative`, `Aspect` and size-inferred `Auto(true)` columns or rows have been calculated. Each undetermined `Auto` column gets a share of the remaining space that is proportional to its `ratio` parameter.
+
+For example, let's say there are two columns left with undetermined `Auto` size when 300 units space remain.
[…]
