+++ docs_franklin/site/makielayout/axis.md
@@ -1,0 +1,754 @@
+```@eval
+using CairoMakie
+CairoMakie.activate!()
+```
+
+# Axis
+
+## Creating an Axis
+
+The `Axis` is a 2D axis that works well with automatic layouts.
+Here's how you create one
+
+```@example laxis
+using CairoMakie
+
+f = Figure()
+
+ax = Axis(f[1, 1], xlabel = "x label", ylabel = "y label",
+    title = "Title")
+
+f
+```
+
+## Plotting into an Axis
+
+You can use all the normal mutating 2D plotting functions with an `Axis`.
+These functions return the created plot object.
+Omitting the `ax` argument plots into the `current_axis()`, which is usually the axis that was last created.
+
+```@example laxis
+lineobject = lines!(ax, 0..10, sin, color = :red)
+scatobject = scatter!(0:0.5:10, cos, color = :orange)
+
+f
+```
+
+## Deleting plots
+
+You can delete a plot object directly via `delete!(ax, plotobj)`.
+You can also remove all plots with `empty!(ax)`.
+
+```@example
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+
+axs = [Axis(f[1, i]) for i in 1:3]
+
+scatters = map(axs) do ax
+    [scatter!(ax, 0:0.1:10, x -> sin(x) + i) for i in 1:3]
+end
+
+delete!(axs[2], scatters[2][2])
+empty!(axs[3])
+
+f
+```
+
+## Setting Axis limits and reversing axes
+
+You can set axis limits with the functions `xlims!`, `ylims!` or `limits!`. The
+numbers are meant in the order left right for `xlims!`, and bottom top for `ylims!`.
+Therefore, if the second number is smaller than the first, the respective axis
+will reverse. You can manually reverse an axis by setting `ax.xreversed = true` or
+`ax.yreversed = true`.
+
+Note that if you enforce an aspect ratio between x-axis and y-axis using `autolimitaspect`,
+the values you set with these functions will probably not be exactly what you get,
+but they will be changed to fit the chosen ratio.
+
+```@example
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+
+axes = [Axis(f[i, j]) for j in 1:3, i in 1:2]
+
+for (i, ax) in enumerate(axes)
+    ax.title = "Axis $i"
+    poly!(ax, Point2f0[(9, 9), (3, 1), (1, 3)],
+        color = cgrad(:inferno, 6, categorical = true)[i])
+end
+
+xlims!(axes[1], [0, 10]) # as vector
+xlims!(axes[2], 10, 0) # separate, reversed
+ylims!(axes[3], 0, 10) # separate
+ylims!(axes[4], (10, 0)) # as tuple, reversed
+limits!(axes[5], 0, 10, 0, 10) # x1, x2, y1, y2
+limits!(axes[6], BBox(0, 10, 0, 10)) # as rectangle
+
+f
+```
+
+### Setting half-automatic limits
+
+You can set half limits by either giving one argument as `nothing` or by using the keyword syntax where only `low` or `high` is given.
+
+```@example
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+
+data = rand(100, 2) .* 0.7 .+ 0.15
+
+Axis(f[1, 1], title = "xlims!(nothing, 1)")
+scatter!(data)
+xlims!(nothing, 1)
+
[…]

[…]
