### 2021-07-19 565180d
remove unused functions
