+++ docs_franklin/site/makielayout.md
@@ -1,0 +1,3 @@
+# MakieLayout
+
+{{list_folder_with_images makielayout}}
