+++ docs_franklin/site/plotting_functions.md
@@ -1,3 +1,3 @@
 # Plotting functions
 
-{{list_plotting_functions}}
+{{list_folder_with_images plotting_functions}}

[…]
