### 2021-07-20 590e114
fix text_bb
