+++ .github/workflows/Deploy.yml
@@ -49,4 +49,4 @@ jobs:
       with:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         BRANCH: gh-pages
-        FOLDER: __site
+        FOLDER: docs_franklin/site/__site
