### 2021-07-20 343a107
disable software rendering loop
