+++ docs_franklin/site/config.md
@@ -15,7 +15,8 @@ ignore = ["node_modules/"]
 generate_rss = true
 website_title = "Franklin Template"
 website_descr = "Example website using Franklin"
-website_url   = "https://tlienart.github.io/FranklinTemplates.jl/"
+website_url   = "https://jkrumbiegel.github.io/Makie.jl/"
+prepath = "Makie.jl"
 +++
 
 <!--
