### 2021-07-20 82ad29a
give WGLMakie same clean up
