+++ WGLMakie/src/particles.jl
@@ -160,11 +160,13 @@ value_or_first(x::StaticArray) = x
 value_or_first(x) = x
 
 function create_shader(scene::Scene, plot::Makie.Text{<:Tuple{<:Union{<:Makie.GlyphCollection, <:AbstractVector{<:Makie.GlyphCollection}}}})
-
     glyphcollection = plot[1]
-    liftkeys = (:position, :rotation, :model, :space, :offset)
-
-    args = getindex.(Ref(plot), liftkeys)
+    res = map(x->Vec2f0(widths(x)), pixelarea(scene))
+    projview = scene.camera.projectionview
+    transfunc =  Makie.transform_func_obs(scene)
+    pos = plot.position
+    space = plot.space
+    offset = plot.offset
 
     # TODO: This is a hack before we get better updating of plot objects and attributes going.
     # Here we only update the glyphs when the glyphcollection changes, if it's a singular glyphcollection.
@@ -173,32 +175,21 @@ function create_shader(scene::Scene, plot::Makie.Text{<:Tuple{<:Union{<:Makie.Gl
     # vector is triggered, but basically all internal objects use the vector of tuples version, and that triggers
     # both glyphcollection and position, so it still works
     if glyphcollection[] isa Makie.GlyphCollection
-        # here we lift the glyph collection
-        collect_glyph_data = (gcollection, projview, transfunc, pos, rotation, model, space, offset) -> begin
-            res = Vec2f0(widths(pixelarea(scene)[]))
-            Makie.preprojected_glyph_arrays(pos, gcollection, space, projview, res, offset, transfunc)
-        end
-        glyph_data = lift(collect_glyph_data, glyphcollection, scene.camera.projectionview, Makie.transform_func_obs(scene), args...)
+        # here we use the glyph collection observable directly
+        gcollection = glyphcollection
     else
-        # and here we don't because it triggers dimension mismatches
-        collect_glyph_data = (projview, transfunc, pos, rotation, model, space, offset) -> begin
-            gcollection = glyphcollection[]
-            res = Vec2f0(widths(pixelarea(scene)[]))
-            Makie.preprojected_glyph_arrays(pos, gcollection, space, projview, res, offset, transfunc)
-        end
-        glyph_data = lift(collect_glyph_data, scene.camera.projectionview, Makie.transform_func_obs(scene), args...)
+        # and here we wrap it into another observable
+        # so it doesn't triggers dimension mismatches
+        # the actual, new value gets then taken in the below lift with to_value
+        gcollection = Observable(glyphcollection)
     end
 
+    glyph_data = lift(pos, gcollection, space, projview, res, offset, transfunc) do pos, gc, args...
+        Makie.preprojected_glyph_arrays(pos, to_value(gc), args...)
+    end
     # unpack values from the one signal:
     positions, offset, uv_offset_width, scale = map((1, 2, 3, 4)) do i
         lift(getindex, glyph_data, i)
-    end
-
-    atlas = get_texture_atlas()
-    keys = (:color, :strokecolor, :rotation)
-
-    signals = map(keys) do key
-        Makie.get_attribute(plot, key)
     end
 
     uniform_color = lift(glyphcollection) do gc
@@ -209,14 +200,7 @@ function create_shader(scene::Scene, plot::Makie.Text{<:Tuple{<:Union{<:Makie.Gl
             Makie.collect_vector(gc.colors, length(gc.glyphs))
         end
     end
-    uniform_stroke_color = lift(glyphcollection) do gc
-        if gc isa AbstractArray
-            reduce(vcat, (Makie.collect_vector(g.strokecolors, length(g.glyphs)) for g in gc),
-                init = RGBAf0[])
[…]
