+++ GLMakie/src/drawing_primitives.jl
@@ -284,7 +284,7 @@ function draw_atomic(screen::GLScreen, scene::Scene,
             gcollection = glyphcollection
         else
             # and here we wrap it into another observable
-            # so it doesn't triggers dimension mismatches
+            # so it doesn't trigger dimension mismatches
             # the actual, new value gets then taken in the below lift with to_value
             gcollection = Observable(glyphcollection)
         end
