+++ docs_franklin/site/Project.toml
@@ -6,5 +6,6 @@ ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 Franklin = "713c75ef-9fc9-4b05-94a9-213340da978e"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
+ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 NodeJS = "2bd173c7-0d6d-553b-b6af-13a54713934c"
