### 2021-07-20 73b232f
make hist more flexible
