+++ docs/src/plotting_functions/hist.md
@@ -31,3 +31,16 @@ hist(data, normalization = :pdf, bar_labels = :values,
      label_formatter=x-> round(x, digits=2), label_size = 15,
      strokewidth = 0.5, strokecolor = (:black, 0.5), color = :values)
 ```
+
+#### Moving histograms
+
+With `scale_to`, `offset`, `fillto` and `flip`, one can put multiple histograms into the same plot:
+
+```@example hist
+fig = Figure()
+ax = Axis(fig[1, 1])
+for i in 1:5
+     hist!(ax, randn(1000), scale_to=0.6, offset=i, fillto=i, direction=:x, flip=true)
+end
+fig
+```
