### 2021-07-20 6466dfc
CompatHelper: bump compat for "GeometryBasics" to "0.4"
