+++ Project.toml
@@ -65,7 +65,7 @@ FixedPointNumbers = "0.6, 0.7, 0.8"
 Formatting = "0.4"
 FreeType = "3.0, 4.0"
 FreeTypeAbstraction = "0.8, 0.9"
-GeometryBasics = "0.3.5"
+GeometryBasics = "0.3.5, 0.4"
 GridLayoutBase = "0.5.5"
 ImageIO = "0.2, 0.3, 0.4, 0.5"
 IntervalSets = "0.3, 0.4, 0.5"
