### 2021-07-20 15f6528
CompatHelper: bump compat for "GeometryBasics" to "0.4" for package CairoMakie
