### 2021-07-20 cadbcf8
CompatHelper: bump compat for "GeometryBasics" to "0.4" for package WGLMakie
