+++ docs_franklin/site/documentation/figure.md
@@ -1,0 +1,172 @@
+# `Figure`
+
+The `Figure` object contains a top-level `Scene` and a `GridLayout`, as well as a list of layoutables that have been placed into it, like `Axis`, `Colorbar`, `Slider`, `Legend`, etc.
+
+!!! note
+    Wherever you see the old `scene, layout = layoutscene()` workflow from MakieLayout, you can imagine that
+    the `Figure` takes over the role of both `scene` and `layout`, plus additional conveniences like keeping
+    track of layoutables.
+
+## Creating a `Figure`
+
+You can create a figure explicitly with the `Figure()` function, and set attributes of the underlying scene.
+The most important one of which is the `resolution`.
+
+```julia
+f = Figure()
+f = Figure(resolution = (600, 400))
+```
+
+A figure is also created implicitly when you use simple, non-mutating plotting commands like `plot()`, `scatter()`, `lines()`, etc.
+Because these commands also create an axis for the plot to live in and the plot itself, they return a compound object `FigureAxisPlot`, which just stores these three parts.
+To access the figure you can either destructure that object into its three parts or access the figure field directly.
+
+```julia
+figureaxisplot = scatter(rand(100, 2))
+figure = figureaxisplot.figure
+
+# destructuring syntax
+figure, axis, plot = scatter(rand(100, 2))
+
+# you can also ignore components
+figure, = scatter(rand(100, 2))
+```
+
+You can pass arguments to the created figure in a dict-like object to the special `figure` keyword:
+
+```julia
+scatter(rand(100, 2), figure = (resolution = (600, 400),))
+```
+
+## Placing layoutables into a `Figure`
+
+All layoutables take their parent figure as the first argument, then you can place them in the figure layout via indexing syntax.
+
+```julia
+f = Figure()
+ax = f[1, 1] = Axis(f)
+sl = f[2, 1] = Slider(f)
+```
+
+## `GridPosition`s and `GridSubposition`s
+
+The indexing syntax of `Figure` is implemented to work seamlessly with layouting.
+If you index into the figure, a `GridPosition` object that stores this indexing operation is created.
+This object can be used to plot a new axis into a certain layout position in the figure, for example like this:
+
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+
+f = Figure()
+pos = f[1, 1]
+scatter(pos, rand(100, 2))
+
+pos2 = f[1, 2]
+lines(pos2, cumsum(randn(100)))
+
+# you don't have to store the position in a variable first, of course
+heatmap(f[1, 3], randn(10, 10))
+
+f
+```
+\end{examplefigure}
+
+
+You can also index further into a `GridPosition`, which creates a `GridSubposition`.
+With `GridSubposition`s you can describe positions in arbitrarily nested grid layouts.
+Often, a desired plot layout can only be achieved with nesting, and repeatedly indexing makes this easy.
+
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+
+f = Figure()
+
+f[1, 1] = Axis(f, title = "I'm not nested")
+f[1, 2][1, 1] = Axis(f, title = "I'm nested")
+
+# plotting into nested positions also works
+heatmap(f[1, 2][2, 1], randn(20, 20))
+
+f
+```
+\end{examplefigure}
+
+
+All nested GridLayouts that don't exist yet, but are needed for a nested plotting call, are created in the background automatically.
+
+!!! note
+    The `GridLayout`s that are implicitly created when using `GridSubpositions` are not directly available in the return
+    value for further manipulation. You can instead retrieve them after the fact with the `content` function, for example,
[…]
