+++ docs_franklin/site/Project.toml
@@ -3,9 +3,11 @@ Animations = "27a7e980-b3e6-11e9-2bcd-0b925532e340"
 CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Chain = "8be319e6-bccf-4806-a6f7-6fae938471bc"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 Franklin = "713c75ef-9fc9-4b05-94a9-213340da978e"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
+JSServe = "824d6782-a2ef-11e9-3a09-e5662e0c26f9"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 NodeJS = "2bd173c7-0d6d-553b-b6af-13a54713934c"

[…]
