### 2021-07-20 17b0af7
begin block logic
