### 2021-07-21 d210073
Add deprecatation warnings for RGB(A)f0, Quaternionf0
