+++ src/deprecated.jl
@@ -1,0 +1,5 @@
+using Base: @deprecate_binding
+ 
+@deprecate_binding Quaternionf0 Quaternionf
+@deprecate_binding RGBf0 RGBf
+@deprecate_binding RGBAf0 RGBAf
