### 2021-07-21 39c854e
fix zoom_shift_lookat for Orthographic projections
