### 2021-07-21 b998975
modify drag speed when using Orthographic projection
