### 2021-07-21 b03b79a
fix default zoom level
