### 2021-07-21 46151d8
remove duplicate code
