### 2021-07-22 d0820b1
remove wglmakie live examples
