+++ docs_franklin/site/utils.jl
@@ -7,7 +7,7 @@ using ImageTransformations
 
 # Pause renderloop for slow software rendering.
 # This way, we only render if we actualy save e.g. an image
-@show GLMakie.set_window_config!(;
+GLMakie.set_window_config!(;
     framerate = 15.0,
     pause_rendering = true
 )
@@ -77,6 +77,54 @@ function env_examplefigure(com, _)
 
   s
 end
+
+
+function lx_video(lxc, _)
+    if length(lxc.braces) == 1
+      rpath = Franklin.stent(lxc.braces[1])
+      alt   = ""
+    elseif length(lxc.braces) == 2
+      rpath = Franklin.stent(lxc.braces[2])
+      alt   = Franklin.stent(lxc.braces[1])
+    end
+    
+    path  = Franklin.parse_rpath(rpath; canonical=false, code=true)
+    fdir, fext = splitext(path)
+
+    # there are several cases
+    # A. a path with no extension --> guess extension
+    # B. a path with extension --> use that
+    # then in both cases there can be a relative path set but the user may mean
+    # that it's in the subfolder /output/ (if generated by code) so should look
+    # both in the relpath and if not found and if /output/ not already last dir
+    candext = ifelse(isempty(fext),
+                     (".mp4", ".mkv", ".mov"), (fext,))
+    for ext ∈ candext
+        candpath = fdir * ext
+        syspath  = joinpath(Franklin.PATHS[:site], split(candpath, '/')...)
+        isfile(syspath) && return html_video(candpath, alt)
+    end
+    # now try in the output dir just in case (provided we weren't already
+    # looking there)
+    p1, p2 = splitdir(fdir)
+    if splitdir(p1)[2] != "output"
+        for ext ∈ candext
+            candpath = joinpath(p1, "output", p2 * ext)
+            syspath  = joinpath(Franklin.PATHS[:site], split(candpath, '/')...)
+            isfile(syspath) && return html_video(candpath, alt)
+        end
+    end
+    return Franklin.html_err("Video matching '$path' not found.")
+  end
+
+function html_video(path, alt)
+  """
+  ~~~
+  <video src="$path" controls="true" loop="true"></video>
+  ~~~
+  """
+end
+
 
 @delay function hfun_list_folder_with_images(param)
 
