### 2021-07-22 e96cf33
wglmakie, admonitions, font-awesome, index page
