+++ docs_franklin/site/index.md
@@ -1,0 +1,84 @@
+![Makie](assets/logo.png)
+
+## Welcome to Makie!
+
+Makie is a high-performance, extendable, and multi-platform plotting ecosystem for the [Julia](https://julialang.org/) programming language.
+
+## Installation and Import
+
+Add one or more of the Makie backend packages [`GLMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/GLMakie) (OpenGL), [`CairoMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/CairoMakie) (Cairo), or [`WGLMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/WGLMakie) (WebGL).
+
+```julia
+]add GLMakie
+using GLMakie
+```
+
+If you want to switch to a different backend in a running session, call that backend's `activate!` function.
+
+```julia
+]add CairoMakie
+using CairoMakie
+# if, for example, GLMakie is activated already
+CairoMakie.activate!()
+```
+
+## First Steps
+
+- Learn the basics of plotting with Makie in the [Basic Tutorial](@ref)
+- Check out how to make more complex plots and layouts in the [Layout Tutorial](@ref)
+- A [gallery of examples](https://lazarusa.github.io/BeautifulMakie/) is maintained in a separate repository.
+
+```julia:autogen_notice
+# hideall
+using Markdown
+using Pkg
+
+function print_dep_version(depname)
+    deps = Pkg.dependencies()
+    version = first(d for d in deps if d.second.name == depname).second.version
+
+    "$depname: v$version"
+end
+
+println(
+"""
+!!! note
+    These docs were autogenerated using:
+      - $(print_dep_version("Makie"))
+      - $(print_dep_version("GLMakie"))
+      - $(print_dep_version("CairoMakie"))
+      - $(print_dep_version("WGLMakie"))
+
+    If you encounter errors with the code examples, check first that your installed versions are appropriate.
+"""
+)
+```
+
+\textoutput{autogen_notice}
+
+## Makie Ecosystem
+
+| Package                                                        | Description                                                                                         |
+| :------------------------------------------------------------- | :-------------------------------------------------------------------------------------------------- |
+| [`Makie.jl`](https://github.com/JuliaPlots/Makie.jl)           | Defines all infrastructure objects which can be visualized using one of the three backend packages. |
+| [`GLMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/GLMakie)       | GPU-powered, interactive 2D and 3D plotting in standalone `GLFW.jl` windows.                        |
+| [`CairoMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/CairoMakie) | `Cairo.jl` based, non-interactive 2D backend for publication-quality vector graphics.               |
+| [`WGLMakie.jl`](https://github.com/JuliaPlots/Makie.jl/tree/master/WGLMakie)     | WebGL-based interactive 2D and 3D plotting that runs within browsers.                               |
+
+The differences between backends are explained in more details under [Backends & Output](@ref).
+
+### Extension Packages
+
+These packages are maintained by third parties. If you install them, keep an eye on version conflicts or downgrades as the Makie ecosystem is developing quickly so things break occasionally.
+
+| Package                                                                       | Description                                              |
+| :---------------------------------------------------------------------------- | :------------------------------------------------------- |
[…]
