+++ docs_franklin/site/Project.toml
@@ -11,3 +11,4 @@ ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 JSServe = "824d6782-a2ef-11e9-3a09-e5662e0c26f9"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 NodeJS = "2bd173c7-0d6d-553b-b6af-13a54713934c"
+WGLMakie = "276b4fcb-3e11-5398-bf8b-a0c2d153d008"
