### 2021-07-22 7331000
Fix edge case in z vector conversion for SurfaceLike
