+++ docs_franklin/site/config.md
@@ -2,7 +2,7 @@
 Add here global page variables to use throughout your website.
 -->
 +++
-author = "Septimia Zenobia"
+author = "Makie.jl"
 mintoclevel = 2
 
 # Add here files or directories that should be ignored by Franklin, otherwise
