### 2021-07-25 fb0ce8d
change thumb width
