### 2021-07-25 9f2bb4f
move examples, fix styling and titles
