+++ docs_franklin/site/examples/makielayout.md
@@ -1,3 +1,0 @@
-# MakieLayout
-
-{{list_folder_with_images makielayout}}
