+++ docs_franklin/site/examples/layoutables.md
@@ -1,0 +1,3 @@
+# Layoutables & Widgets
+
+{{list_folder_with_images layoutables}}

[…]
