+++ GLMakie/Project.toml
@@ -28,7 +28,7 @@ FileIO = "1.1"
 FixedPointNumbers = "0.7, 0.8"
 FreeTypeAbstraction = "0.8, 0.9"
 GLFW = "3"
-GeometryBasics = "0.3"
+GeometryBasics = "0.4.1"
 Makie = "=0.15.0"
 MeshIO = "0.4"
 ModernGL = "1"
