+++ CairoMakie/Project.toml
@@ -22,15 +22,15 @@ Colors = "0.10, 0.11, 0.12"
 FFTW = "1"
 FileIO = "1.1"
 FreeType = "3, 4.0"
-GeometryBasics = "0.2, 0.3"
+GeometryBasics = "0.4.1"
 Makie = "=0.15.0"
 StaticArrays = "0.12, 1.0"
 julia = "1.3"
 
 [extras]
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 ReferenceTests = "d37af2e0-5618-4e00-9939-d430db56ee94"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test", "Pkg", "ReferenceTests"]
