+++ docs_franklin/site/tutorials/basic-tutorial.md
@@ -31,11 +31,11 @@ Makie.inline!(true) # hide
 nothing # hide
 ```
 
-Makie has many different plotting functions, one of the most common ones is [lines](@ref).
+Makie has many different plotting functions, one of the most common ones is \myreflink{lines}.
 You can just call such a function and your plot will appear if your coding environment can show png or svg files.
 
 !!! note
-    Objects such as [Figure](@ref), `FigureAxisPlot` or `Scene` are usually displayed whenever they are returned in global scope (e.g. in the REPL).
+    Objects such as [Figure](\reflink{Figures}), `FigureAxisPlot` or `Scene` are usually displayed whenever they are returned in global scope (e.g. in the REPL).
     To display such objects from within a local scope, like from within a function, you can directly call `display(figure)`, for example.
 
 \begin{examplefigure}{}
@@ -47,7 +47,7 @@ lines(x, y)
 ```
 \end{examplefigure}
 
-Another common function is [scatter](@ref).
+Another common function is \myreflink{scatter}.
 
 \begin{examplefigure}{}
 ```julia
@@ -169,7 +169,7 @@ current_figure()
 \end{examplefigure}
 
 Note that the color array does not actually contain colors, rather the numerical values are mapped to the plot's `colormap`.
-There are many different colormaps to choose from, take a look on the [Colors](@ref) page.
+There are many different colormaps to choose from, take a look on the \myreflink{Colors} page.
 
 The values are mapped to colors via the `colorrange` attribute, which by default goes from the minimum to the maximum color value, but we can also limit or expand the range manually.
 For example, we can constrain the previous scatter plot's color range to (0.25, 0.75), which will clip the colors at the bottom and the top quarters.
@@ -222,10 +222,10 @@ current_figure()
 ## Subplots
 
 Makie uses a powerful layout system under the hood, which allows you to create very complex figures with many subplots.
-For the easiest way to do this, we need a [Figure](@ref) object.
+For the easiest way to do this, we need a [Figure](\reflink{Figures}) object.
 So far, we haven't seen this explicitly, it was created in the background in the first plotting function call.
 
-We can also create a [Figure](@ref) directly and then continue working with it.
+We can also create a [Figure](\reflink{Figures}) directly and then continue working with it.
 We can make subplots by giving the location of the subplot in our layout grid as the first argument to our plotting function.
 The basic syntax for specifying the location in a figure is `fig[row, col]`.
 
@@ -250,10 +250,10 @@ Each `lines` call creates a new axis in the position given as the first argument
 
 ## Constructing axes manually
 
-Like [Figure](@ref)s, we can also create axes manually.
+Like [Figure](\reflink{Figures})s, we can also create axes manually.
 This is useful if we want to prepare an empty axis to then plot into it later.
 
-The default 2D axis that we have created implicitly so far is called [Axis](@ref) and can also be created in a specific position in the figure by passing that position as the first argument.
+The default 2D axis that we have created implicitly so far is called \myreflink{Axis} and can also be created in a specific position in the figure by passing that position as the first argument.
 
 For example, we can create a figure with three axes.
 
@@ -302,11 +302,11 @@ fig
 
 ## Legend and Colorbar
 
-We have seen two `Layoutables` so far, the [Axis](@ref) and the [Legend](@ref) which was created by the function `axislegend`.
+We have seen two `Layoutables` so far, the \myreflink{Axis} and the \myreflink{Legend} which was created by the function `axislegend`.
 All `Layoutable`s can be placed into the layout of a figure at arbitrary positions, which makes it easy to assemble complex figures.
 
-In the same way as with the [Axis](@ref) before, you can also create a [Legend](@ref) manually and then place it freely, wherever you want, in the figure.
[…]
