+++ .github/workflows/cairomakie.yaml
@@ -27,6 +27,8 @@ jobs:
           - ubuntu-latest
         arch:
           - x64
+    env:
+      JULIA_PKG_SERVER: ""
     steps:
       - name: Checkout
         uses: actions/checkout@v2
