### 2021-07-27 bf16f08
Replace f0 type suffix with f and rename Rect types
