+++ GLMakie/src/events.jl
@@ -55,7 +55,7 @@ function window_area(scene::Scene, screen::Screen)
         # TODO put back window position, but right now it makes more trouble than it helps#
         # x, y = GLFW.GetWindowPos(window)
         # if minimum(rect) != Vec(x, y)
-        #     event[] = IRect(x, y, framebuffer_size(window))
+        #     event[] = Recti(x, y, framebuffer_size(window))
         # end
         w, h = GLFW.GetFramebufferSize(window)
         if Vec(w, h) != widths(rect)
@@ -64,7 +64,7 @@ function window_area(scene::Scene, screen::Screen)
             # dpi of a monitor should be the same in x y direction.
             # if not, minimum seems to be a fair default
             dpievent[] = minimum(props.dpi)
-            event[] = IRect(minimum(rect), w, h)
+            event[] = Recti(minimum(rect), w, h)
         end
     end
     return

[…]
