### 2021-07-27 2be37f9
Use "dev ./<submodule>" syntax in CI
