+++ .github/workflows/ci.yml
@@ -51,7 +51,7 @@ jobs:
         run: |
           using Pkg;
           # dev mono repo versions
-          pkg"dev . MakieCore"
+          pkg"dev . ./MakieCore"
           Pkg.test("Makie"; coverage=true)
 
       - uses: julia-actions/julia-processcoverage@v1
