+++ .github/workflows/wglmakie.yaml
@@ -42,7 +42,7 @@ jobs:
         run: |
           using Pkg;
           # dev mono repo versions
-          pkg"dev . MakieCore WGLMakie ReferenceTests"
+          pkg"dev . ./MakieCore ./WGLMakie ./ReferenceTests"
       - name: Run the tests
         continue-on-error: true
         run: >
