+++ .github/workflows/glmakie.yaml
@@ -52,7 +52,7 @@ jobs:
         run: |
           using Pkg;
           # dev mono repo versions
-          pkg"dev . MakieCore GLMakie ReferenceTests"
+          pkg"dev . ./MakieCore ./GLMakie ./ReferenceTests"
       - name: Run the tests
         continue-on-error: true
         run: >
