+++ .github/workflows/cairomakie.yaml
@@ -51,7 +51,7 @@ jobs:
         run: |
           using Pkg;
           # dev mono repo versions
-          pkg"dev . MakieCore CairoMakie ReferenceTests"
+          pkg"dev . ./MakieCore ./CairoMakie ./ReferenceTests"
       - name: Run the tests
         continue-on-error: true
         run: >
