+++ .github/workflows/docs.yml
@@ -50,7 +50,7 @@ jobs:
            pkg"registry add https://github.com/JuliaRegistries/General";
            pkg"registry up General";
            # dev mono repo versions
-           pkg"dev . MakieCore CairoMakie WGLMakie GLMakie; add Documenter#master; instantiate"'
+           pkg"dev . ./MakieCore ./CairoMakie ./WGLMakie ./GLMakie; add Documenter#master; instantiate"'
       - name: Build and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
