+++ docs_franklin/site/examples/plotting_functions/image.md
@@ -11,12 +11,22 @@ CairoMakie.activate!() # hide
 Makie.inline!(true) # hide
 using FileIO
 
-img = rotr90(load(assetpath("cow.png")))
+img = load(assetpath("cow.png"))
 
 f = Figure()
-Axis(f[1, 1], aspect = DataAspect())
 
-image!(img)
+image(f[1, 1], img,
+    axis = (title = "Default",))
+
+image(f[1, 2], img,
+    axis = (aspect = DataAspect(), title = "DataAspect()",))
+
+image(f[2, 1], rotr90(img),
+    axis = (aspect = DataAspect(), title = "rotr90",))
+
+image(f[2, 2], img',
+    axis = (aspect = DataAspect(), yreversed = true,
+        title = "img' and reverse y-axis",))
 
 f
 ```
