### 2021-07-28 c9b0238
replace nonworking lifts
