+++ docs_franklin/site/examples/layoutables/button.md
@@ -1,5 +1,3 @@
-
-
 # Button
 
 \begin{examplefigure}{}
