+++ docs_franklin/site/examples/layoutables/button.md
@@ -11,7 +11,7 @@ fig[2, 1] = buttongrid = GridLayout(tellwidth = false)
 
 counts = Node([1, 4, 3, 7, 2])
 
-buttonlabels = [@lift("Count: $($counts[i])") for i in 1:5]
+buttonlabels = [lift(x -> "Count: $(x[i])", counts) for i in 1:5]
 
 buttons = buttongrid[1, 1:5] = [Button(fig, label = l) for l in buttonlabels]
 
