+++ Project.toml
@@ -65,7 +65,7 @@ Formatting = "0.4"
 FreeType = "3.0, 4.0"
 FreeTypeAbstraction = "0.8, 0.9"
 GeometryBasics = "0.3.5"
-GridLayoutBase = "0.5.5"
+GridLayoutBase = "0.5.5, 0.6"
 ImageIO = "0.2, 0.3, 0.4, 0.5"
 IntervalSets = "0.3, 0.4, 0.5"
 Isoband = "0.1"
