### 2021-07-29 0cf37d1
CompatHelper: bump compat for "GeometryBasics" to "0.4"
