### 2021-07-30 7aef1ee
update icons + move logos to assets
