### 2021-07-30 c15a683
fix layout tutorial image
