### 2021-07-30 90cb191
new layout tutorial bare version
