+++ docs_franklin/site/tutorials/layout-tutorial.md
@@ -3,588 +3,115 @@
 In this tutorial, we will see some of the capabilities of layouts in Makie while
 building a complex figure step by step. This is the final result we will create:
 
-\outputimage{layout_tutorial_final.png}
 
-~~~
-<details><summary>You can click here to see the full code</summary>
-~~~
-
+\begin{examplefigure}{}
 ```julia
 using CairoMakie
+using FileIO
 CairoMakie.activate!() # hide
 
 noto_sans = assetpath("fonts", "NotoSans-Regular.ttf")
 noto_sans_bold = assetpath("fonts", "NotoSans-Bold.ttf")
 
-fig = Figure(backgroundcolor = RGBf0(0.98, 0.98, 0.98),
+f = Figure(backgroundcolor = RGBf0(0.98, 0.98, 0.98),
     resolution = (1000, 700), font = noto_sans)
 
-ax1 = fig[1, 1] = Axis(fig, title = "Pre Treatment")
+s1 = f[1, 1] = GridLayout(default_rowgap = 10, default_colgap = 10)
 
-data1 = randn(50, 2) * [1 2.5; 2.5 1] .+ [10 10]
+ax = Axis(s1[2, 1])
+axtop = Axis(s1[1, 1])
+axright = Axis(s1[2, 2])
 
-line1 = lines!(ax1, 5..15, x -> x, color = :red, linewidth = 2)
-scat1 = scatter!(ax1, data1,
-    color = (:red, 0.3), markersize = 15px, marker = '■')
+linkxaxes!(ax, axtop)
+linkyaxes!(ax, axright)
 
-ax2, line2 = lines(fig[1, 2], 7..17, x -> -x + 26,
-    color = :blue, linewidth = 2,
-    axis = (title = "Post Treatment",))
+data = randn(3, 100, 2) .+ [1, 3, 5]
+labels = ["alpha", "beta", "gamma"]
 
-data2 = randn(50, 2) * [1 -2.5; -2.5 1] .+ [13 13]
-
-scat2 = scatter!(data2,
-    color = (:blue, 0.3), markersize = 15px, marker = '▲')
-
-linkaxes!(ax1, ax2)
-
-hideydecorations!(ax2, grid = false)
-
-ax1.xlabel = "Weight [kg]"
-ax2.xlabel = "Weight [kg]"
-ax1.ylabel = "Maximum Velocity [m/sec]"
-
-leg = fig[1, end+1] = Legend(fig,
-    [line1, scat1, line2, scat2],
-    ["f(x) = x", "Data", "f(x) = -x + 26", "Data"])
-
-fig[2, 1:2] = leg
-
-trim!(fig.layout)
-
-leg.tellheight = true
-
-leg.orientation = :horizontal
-
-hm_axes = fig[1:2, 3] = [Axis(fig, title = t) for t in ["Cell Assembly Pre", "Cell Assembly Post"]]
-
-heatmaps = [heatmap!(ax, i .+ rand(20, 20)) for (i, ax) in enumerate(hm_axes)]
-
-hm_sublayout = GridLayout()
-fig[1:2, 3] = hm_sublayout
-
-# there is another shortcut for filling a GridLayout vertically with
-# a vector of content
-hm_sublayout[:v] = hm_axes
-
-hidedecorations!.(hm_axes)
-
-for hm in heatmaps
-    hm.colorrange = (1, 3)
+for (label, col) in zip(labels, eachslice(data, dims = 1))
+    scatter!(ax, col, label = label)
+    density!(axtop, col[:, 1])
+    density!(axright, col[:, 2], direction = :y)
 end
 
[…]
