### 2021-07-31 50a41b9
correctly check for identity transform
