+++ paper.bib
@@ -26,3 +26,14 @@
   journal={arXiv preprint arXiv:1209.5145},
   year={2012}
 }
+
+@software{Breloff:2021,
+  author       = {Tom Breloff},
+  title        = {Plots.jl},
+  month        = jun,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v1.16.3},
+  doi          = {10.5281/zenodo.4907285},
+  url          = {https://doi.org/10.5281/zenodo.4907285}
+}
