### 2021-07-31 #3349 @jkrumbiegel
@fverdugo I have added a citation to Plots.jl and a short paragraph describing the differences.

I have also added contribution guidelines here https://github.com/JuliaPlots/Makie.jl/blob/master/CONTRIBUTING.md

The file "airportlocations.csv" is in the released assets now, you can see an example in the docs here http://makie.juliaplots.org/stable/plotting_functions/scatter.html. The example should therefore work.

The contribution guidelines also mention now that you need to dev `ReferenceTests` from the monorepo to run the tests, although Simon is also trying to improve that process a bit for local testing.

Let me know if there are any other issues!
