+++ docs_franklin/site/tutorials/layout-tutorial.md
@@ -1,8 +1,27 @@
+@def reeval = true
+
 # Layout Tutorial
 
-In this tutorial, we will see some of the capabilities of layouts in Makie while
-building a complex figure step by step. This is the final result we will create:
+In this tutorial, you will learn how to create a complex figure using Makie's layout tools.
 
+Let's say that we want to create the following figure:
+
+\outputimage{final_result.png}
+
+How do we approach this task?
+
+In the following sections, we'll go over the process step by step.
+We're not always going to use the shortest possible syntax, as the main goal is to get a better understanding of the logic and the available options.
+
+## Basic layout plan
+
+When building figures, you always think in terms of rectangular boxes. We want to find the biggest boxes that enclose meaningful groups of content, and then we realize those boxes either using `GridLayout` or by placing content objects there.
+
+If we look at our target figure, there is one box around each of the labelled areas A, B, C and D. But A and C are not in one row, neither are B and D. This means that we don't use a 2x2 GridLayout, but have to be a little more creative.
+
+We can say that A and B are in one column, and C and D are in one column. We can have different row heights for both groups by making one big nested `GridLayout` within the second column, in which we place C and D. This way the rows of column 2 are decoupled from column 1.
+
+Ok, let's create the figure first with a gray backgroundcolor, and a predefined font:
 
 \begin{examplefigure}{}
 ```julia
@@ -15,102 +34,307 @@ noto_sans_bold = assetpath("fonts", "NotoSans-Bold.ttf")
 
 f = Figure(backgroundcolor = RGBf0(0.98, 0.98, 0.98),
     resolution = (1000, 700), font = noto_sans)
+```
+\end{examplefigure}
 
-s1 = f[1, 1] = GridLayout(default_rowgap = 10, default_colgap = 10)
+Now, let's make the four nested GridLayouts that are going to hold the objects of A, B, C and D. There's also the layout that holds C and D together, so the rows are separate from A and B. We are not going to see anything yet as we have no visible content, but that will come soon.
 
-ax = Axis(s1[2, 1])
-axtop = Axis(s1[1, 1])
-axright = Axis(s1[2, 2])
+```julia:grids
+ga = f[1, 1] = GridLayout()
+gb = f[2, 1] = GridLayout()
+gcd = f[1:2, 2] = GridLayout()
+gc = gcd[1, 1] = GridLayout()
+gd = gcd[2, 1] = GridLayout()
+```
 
-linkxaxes!(ax, axtop)
-linkyaxes!(ax, axright)
+Now we can start placing objects into the figure. We start with A.
 
+There are three axes and a legend. We can place the axes first and plot the first data into them.
+
+\begin{examplefigure}{}
+```julia
+axtop = Axis(ga[1, 1])
+axmain = Axis(ga[2, 1], xlabel = "before", ylabel = "after")
+axright = Axis(ga[2, 2])
+
+labels = ["treatment", "placebo", "control"]
 data = randn(3, 100, 2) .+ [1, 3, 5]
-labels = ["alpha", "beta", "gamma"]
 
 for (label, col) in zip(labels, eachslice(data, dims = 1))
-    scatter!(ax, col, label = label)
+    scatter!(axmain, col, label = label)
     density!(axtop, col[:, 1])
     density!(axright, col[:, 2], direction = :y)
 end
 
-hidedecorations!(axright, grid = false)
+f
[…]
