### 2021-08-01 fdcb4c6
improve start page with images
