+++ docs_franklin/site/utils.jl
@@ -4,6 +4,7 @@ using GLMakie
 using FileIO
 using ImageTransformations
 using Colors
+using Pkg
 
 include("colormap_generation.jl")
 
@@ -269,4 +270,17 @@ function lx_outputimage(lxc, _)
     rpath = Franklin.stent(lxc.braces[1])
     path = Franklin.parse_rpath("output/" * rpath; canonical=false, code=true)
     return "![$rpath]($path)"
+end
+
+function hfun_generating_versions()
+  
+  function dep_version(depname)
+      deps = Pkg.dependencies()
+      version = first(d for d in deps if d.second.name == depname).second.version
+  
+      "$depname: v$version"
+  end
+  
+  "<p>These docs were autogenerated using " * join([dep_version(x)
+    for x in ["Makie", "GLMakie", "CairoMakie", "WGLMakie"]], ", ") * "</p>"
 end
