+++ docs_franklin/site/index.md
@@ -11,57 +11,87 @@ Add one or more of the Makie backend packages [`GLMakie.jl`](https://github.com/
 using GLMakie
 ```
 
-If you want to switch to a different backend in a running session, call that backend's `activate!` function.
-
-```julia
-]add CairoMakie
-using CairoMakie
-# if, for example, GLMakie is activated already
-CairoMakie.activate!()
-```
+To switch to a different backend, for example `CairoMakie`, call `CairoMakie.activate!()`.
 
 ## First Steps
 
-- Learn the basics of plotting with Makie in the \myreflink{Basic Tutorial}
-- Check out how to make more complex plots and layouts in the \myreflink{Layout Tutorial}
-- A [gallery of examples](https://lazarusa.github.io/BeautifulMakie/) is maintained in a separate repository.
+@@box-container
+  @@box
+    @@title \myreflink{Basic Tutorial} @@
+    @@box-content
+      @@description
+      Learn the basics of plotting with Makie.
+      @@
+      ~~~
+      <img src="/assets/basic_tutorial_example.png">
+      ~~~
+    @@
+  @@
 
-```julia:autogen_notice
-# hideall
-using Markdown
-using Pkg
+  @@box
+    @@title \myreflink{Layout Tutorial} @@
+    @@box-content
+      @@description
+      Check out how to make complex plots and layouts.
+      @@
+      ~~~
+      <img src="/assets/tutorials/layout-tutorial/code/output/final_result.png">
+      ~~~
+    @@
+  @@
 
-function print_dep_version(depname)
-    deps = Pkg.dependencies()
-    version = first(d for d in deps if d.second.name == depname).second.version
+  @@box
+    @@title [Beautiful Makie](https://lazarusa.github.io/BeautifulMakie/) @@
+    @@box-content
+      @@description
+      A gallery that is maintained in a separate repository.
+      @@
+      ~~~
+      <img src="/assets/beautifulmakie_example.png">
+      ~~~
+    @@
+  @@
+@@
 
-    "$depname: v$version"
-end
+## Makie Backends
 
-println(
-"""
-!!! note
-    These docs were autogenerated using:
-      - $(print_dep_version("Makie"))
-      - $(print_dep_version("GLMakie"))
-      - $(print_dep_version("CairoMakie"))
-      - $(print_dep_version("WGLMakie"))
+@@box-container
+  @@box
+    @@title [GLMakie.jl](https://github.com/JuliaPlots/Makie.jl/tree/master/GLMakie) @@
+    @@box-content
+      @@description
+      GPU-powered, interactive 2D and 3D plotting in standalone `GLFW.jl` windows.
+      @@
+      ~~~
+      <img src="/assets/examples/plotting_functions/surface/code/output/example_9574460083465564948.png">
+      ~~~
+    @@
+  @@
 
-    If you encounter errors with the code examples, check first that your installed versions are appropriate.
-"""
-)
-```
+  @@box
+    @@title [CairoMakie.jl](https://github.com/JuliaPlots/Makie.jl/tree/master/CairoMakie) @@
+    @@box-content
+      @@description
+      `Cairo.jl` based, non-interactive 2D backend for publication-quality vector graphics.
+      @@
+      ~~~
+      <img src="/assets/examples/plotting_functions/density/code/output/example_7784305977099122996.png">
+      ~~~
+    @@
+  @@
 
-\textoutput{autogen_notice}
-
-## Makie Ecosystem
-
-| Package                                                        | Description                                                                                         |
-| :------------------------------------------------------------- | :-------------------------------------------------------------------------------------------------- |
[…]
