+++ docs_franklin/site/tutorials/layout-tutorial.md
@@ -1,4 +1,4 @@
-@def reeval = true
+@def reeval = false
 
 # Layout Tutorial
 
