### 2021-08-03 db18eba
make tests runnable locally
