+++ Project.toml
@@ -73,7 +73,7 @@ KernelDensity = "0.5, 0.6"
 LaTeXStrings = "1.2"
 MakieCore = "0.1.3"
 Match = "1.1"
-MathTeXEngine = "0.1.2, 0.2"
+MathTeXEngine = "0.2"
 Observables = "0.4"
 Packing = "0.4"
 PlotUtils = "1"
