### 2021-08-03 5d051f8
clean up scene
