### 2021-08-03 b6c8749
clean up constructor
