### 2021-08-03 1665017
move offset to barplot+clean up
