### 2021-08-03 #3349 @SimonDanisch
Awesome :) I also fixed the local tests: https://github.com/JuliaPlots/Makie.jl/pull/1190
But we still need to tag a new version for this to work! Will update the contribution docs before we tag...
