+++ src/conversions.jl
@@ -949,7 +949,6 @@ function convert_attribute(s::VecTypes{N}, ::key"rotation") where N
     elseif N == 3
         rotation_between(Vec3f0(0, 0, 1), to_ndim(Vec3f0, s, 0.0))
     elseif N == 2
-
         rotation_between(Vec3f0(0, 1, 0), to_ndim(Vec3f0, s, 0.0))
     else
         error("The $N dimensional vector $s can't be converted to a rotation.")
