### 2021-08-03 ee2dc55
remove unused attributes
