+++ src/basic_recipes/convenience_functions.jl
@@ -44,32 +44,28 @@ function showgradients(
     scene = Scene(resolution = resolution)
 
     map(collect(cgrads)) do cmap
+        c = to_colormap(cmap)
 
-         c = to_colormap(cmap)
+        cbar = image!(
+            scene,
+            range(0, stop = 10, length = length(c)),
+            range(0, stop = 1, length = length(c)),
+            reshape(c, (length(c),1))
+        )[end]
 
-         cbar = image!(
-             scene,
-             range(0, stop = 10, length = length(c)),
-             range(0, stop = 1, length = length(c)),
-             reshape(c, (length(c),1)),
-             show_axis = false
-         )[end]
+        cmapstr = monospace ? UnicodeFun.to_latex("\\mono{$cmap}") : string(cmap, ":")
 
-         cmapstr = monospace ? UnicodeFun.to_latex("\\mono{$cmap}") : string(cmap, ":")
+        text!(
+            scene,
+            cmapstr,
+            position = Point2f0(-0.1, 0.5 + h),
+            align = (:right, :center),
+            textsize = textsize
+        )
 
-         text!(
-             scene,
-             cmapstr,
-             position = Point2f0(-0.1, 0.5 + h),
-             align = (:right, :center),
-             show_axis = false,
-             textsize = textsize
-         )
+        translate!(cbar, 0, h, 0)
 
-         translate!(cbar, 0, h, 0)
-
-         h -= (1 + offset)
-
+        h -= (1 + offset)
     end
 
     scene
