+++ src/makielayout/layoutables/button.jl
@@ -41,7 +41,7 @@ function layoutable(::Type{Button}, fig_or_scene::FigureLike; bbox = nothing, kw
     bcolors = (; out = buttoncolor, active = buttoncolor_active, hover = buttoncolor_hover)
     bcolor = lift((s,_...)->bcolors[s][], Any, mousestate, values(bcolors)...)
     button = poly!(subscene, roundedrectpoints, strokewidth = strokewidth, strokecolor = strokecolor,
-        color = bcolor, raw = true, inspectable = false)
+        color = bcolor, inspectable = false)
     decorations[:button] = button
 
 
@@ -49,7 +49,7 @@ function layoutable(::Type{Button}, fig_or_scene::FigureLike; bbox = nothing, kw
     lcolors = (; out = labelcolor, active = labelcolor_active, hover = labelcolor_hover)
     lcolor = lift((s,_...)->lcolors[s][], Any, mousestate, values(lcolors)...)
     labeltext = text!(subscene, label, position = textpos, textsize = textsize, font = font,
-        color = lcolor, align = (:center, :center), raw = true, space = :data, inspectable = false)
+        color = lcolor, align = (:center, :center), space = :data, inspectable = false)
 
     decorations[:label] = labeltext
 
