### 2021-08-03 #3349 @gaelforget
> Functionality: Have the functional claims of the software been confirmed?
> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

I am about to check these two items in the list but would suggest reviewing the text or example in the paper in regard to functionality / performance since :

-  trying to run the 100 million points example crashed my laptop several times (2018 MacBook Pro). New potential user of Makie who might copy and paste the code snippet from the paper might be turned off if this happens to them too (?), especially if they just read that about this _size that many other plotting frameworks would struggle with_ in the paper. Worked for me with `1_000_000` though, so maybe that's better for the code snippet. And maybe mention system requirements for running the example with `100_000_000`. Not sure.

- about _build up a figure piece by piece in GLMakie, explore the data interactively, and then save a publication-quality version with CairoMakie._ it seems unfortunate that when I do this with the code snippet in the paper the two panels on the left hand side turn into blanks. I believe the scatter disappears due to the markersize spec (maybe these are handled differently in Cairo and GL?). The 3D contour issued a warning that _Volume{Tuple{IntervalSets.ClosedInterval{Float32} ... not supported by cairo right now_ (seems fair enough although not ideal since that's the example in the paper).
