### 2021-08-03 #3349 @gaelforget
> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

Very nice overall so I checked this one in the list too. A few small issues caught my attention though:

- Sentence _To effectively understand ... powerful plotting software._ seems to be missing a verb / seems broken
- `airports_rep` should be `a_rep` , I believe, in the code snippet
- _Makie’s core package is AbstractPlotting.jl_ will need to be revised to reflect the reorganization that took place over the course of the review. I think this is now in the `MakieCore` subfolder. Not sure if the subfolder structure (MakieCore, CairoMakie, etc) is useful to explain in the paper, but may be that would be useful.
