### 2021-08-03 #3349 @gaelforget
> State of the field: Do the authors describe how this software compares to other commonly-used packages?
> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

LGTM but maybe for completeness it could be useful to mention GR.jl and Plots.jl
