### 2021-08-03 #3349 @gaelforget
I see now in https://github.com/JuliaPlots/Makie.jl/tree/joss-paper that the `Plots.jl` and `AbstractPlotting.jl` bits have already been addressed
