### 2021-08-03 #3349 @gaelforget
> @gaelforget can you update us on where we stand with this review? Thanks again for your help.

Hi @Kevin-Mattheus-Moerman et al

I just went through the checklist and also recommend that the paper be accepted for publication. Great package and great work by the authors!!

ps. a few small issues that should easily be addressed are listed above
