### 2021-08-04 #3349 @jkrumbiegel
>     * trying to run the 100 million points example crashed my laptop several times

This is a good point, I made the example with a more powerful GPU, but this could indeed be confusing. It's just that 1 million points are not as impressive of a feat, and we wanted to show what can be possible. Maybe the text should qualify that the GPU can be the bottleneck?

> it seems unfortunate that when I do this with the code snippet in the paper the two panels on the left hand side turn into blanks

Also a very good point. 2D vector graphics can never handle 3D data, so maybe the text should make clearer that CairoMakie will handle the 2D subset (with some 3D things working with caveats). So the scatter didn't work for you but the brain did? Maybe the point size is too small for Cairo to build up to an image, while GLMakie handles this case a little differently. Not ideal for an example.
