### 2021-08-04 6eb7bea
clean up constructors
