+++ src/scenes.jl
@@ -25,7 +25,7 @@ mutable struct Scene <: AbstractScene
     camera::Camera
 
     "The controls for the camera of the Scene."
-    camera_controls::RefValue
+    camera_controls::AbstractCamera
 
     "The [`Transformation`](@ref) of the Scene."
     transformation::Transformation
@@ -76,121 +76,89 @@ function Base.show(io::IO, scene::Scene)
 end
 
 function Scene(
-        events::Events,
-        px_area::Observable{IRect2D},
-        clear::Bool,
-        camera::Camera,
-        camera_controls::RefValue,
-        transformation::Transformation,
-        plots::Vector{AbstractPlot},
-        theme::Attributes, # the default values a scene owns
-        children::Vector{Scene},
-        current_screens::Vector{AbstractScreen},
-        parent=nothing,
-        backgroundcolor=Observable(RGBAf0(1, 1, 1, 1)),
-        visible=Observable(true)
+        px_area::Union{Observable{IRect2D}, Nothing} = nothing;
+        events::Events = Events(),
+        clear::Bool = true,
+        transform_func=identity,
+        camera::Union{Function, Camera, Nothing} = nothing,
+        camera_controls::AbstractCamera = EmptyCamera(),
+        transformation::Transformation = Transformation(transform_func),
+        plots::Vector{AbstractPlot} = AbstractPlot[],
+        theme::Attributes = Attributes(),
+        children::Vector{Scene} = Scene[],
+        current_screens::Vector{AbstractScreen} = AbstractScreen[],
+        parent = nothing,
+        backgroundcolor = Observable(RGBAf0(1, 1, 1, 1)),
+        visible = Observable(true),
+        theme_kw...
     )
     bg = map(to_color, backgroundcolor)
-    return Scene(
-        parent, events, px_area, clear, camera, camera_controls,
-        transformation, plots, theme,
-        children, current_screens, bg, visible
-    )
-end
-
-
-function Scene(;clear=true, transform_func=identity, backgroundcolor=Observable(RGBAf0(1, 1, 1, 1)), visible=Observable(true), theme...)
-    events = Events()
-    m_theme = current_default_theme(; theme...)
-    px_area = lift(m_theme.resolution) do res
-        IRect(0, 0, res)
+    m_theme = current_default_theme(; theme..., theme_kw...)
+    if isnothing(px_area)
+        px_area = lift(m_theme.resolution) do res
+            IRect(0, 0, res)
+        end
     end
+
+    cam = camera isa Camera ? camera : Camera(px_area)
+
     on(events.window_area, priority = typemax(Int8)) do w_area
         if !any(x -> x ≈ 0.0, widths(w_area)) && px_area[] != w_area
             px_area[] = w_area
         end
         return Consume(false)
     end
+
     scene = Scene(
-        events,
-        px_area,
-        clear,
-        Camera(px_area),
-        RefValue{Any}(EmptyCamera()),
-        Transformation(transform_func),
-        AbstractPlot[],
-        m_theme,
-        Scene[],
-        AbstractScreen[],
-        nothing,
-        backgroundcolor,
-        visible
+        parent, events, px_area, clear, cam, camera_controls,
+        transformation, plots, m_theme,
+        children, current_screens, bg, visible
     )
-    # Set the transformation parent
     scene.transformation.parent[] = scene
+    cam = if camera isa Function
+        camera(scene)
+    end
     return scene
 end
 
 function Scene(
-        scene::Scene;
-        events=scene.events,
-        px_area=scene.px_area,
[…]
