+++ src/interfaces.jl
@@ -34,7 +34,6 @@ function color_and_colormap!(plot, intensity = plot[:color])
     end
 end
 
-
 function calculated_attributes!(::Type{<: Mesh}, plot)
     need_cmap = color_and_colormap!(plot)
     need_cmap || delete!(plot, :colormap)
@@ -266,24 +265,6 @@ const PlotFunc = Union{Type{Any}, Type{<: AbstractPlot}}
 ######################################################################
 # In this section, the plotting functions have P as the first argument
 # These are called from type recipes
-
-# # non-mutating, without positional attributes
-
-# function plot(P::PlotFunc, args...; kw_attributes...)
-#     attributes = Attributes(kw_attributes)
-#     plot(P, attributes, args...)
-# end
-
-# # with positional attributes
-
-# function plot(P::PlotFunc, attrs::Attributes, args...; kw_attributes...)
-#     attributes = merge!(Attributes(kw_attributes), attrs)
-#     scene_attributes = extract_scene_attributes!(attributes)
-#     scene = Scene(; scene_attributes...)
-#     plot!(scene, P, attributes, args...)
-# end
-
-# mutating, without positional attributes
 
 function plot!(P::PlotFunc, scene::SceneLike, args...; kw_attributes...)
     attributes = Attributes(kw_attributes)
