### 2021-08-04 504be72
Update src/basic_recipes/barplot.jl
Co-authored-by: Pietro Vertechi <pietro.vertechi@neuro.fchampalimaud.org>
