### 2021-08-04 9f4d2aa
clean up + fix constructor
