+++ src/makielayout/layoutables/axis.jl
@@ -72,7 +72,7 @@ function layoutable(::Type{<:Axis}, fig_or_scene::Union{Figure, Scene}; bbox = n
 
     scenearea = sceneareanode!(layoutobservables.computedbbox, finallimits, aspect)
 
-    scene = Scene(topscene, px_area=scenearea)
+    scene = Scene(topscene, px_area=scenearea, camera=campixel!)
 
     background = poly!(topscene, scenearea, color = backgroundcolor, strokewidth = 0, inspectable = false)
     translate!(background, 0, 0, -100)
@@ -82,8 +82,6 @@ function layoutable(::Type{<:Axis}, fig_or_scene::Union{Figure, Scene}; bbox = n
 
     xaxislinks = Axis[]
     yaxislinks = Axis[]
-
-    campixel!(scene)
 
     xgridnode = Node(Point2f0[])
     xgridlines = linesegments!(
