### 2021-08-04 579c758
smarter defaults for tellheight and tellwidth
