+++ src/makielayout/layoutables/colorbar.jl
@@ -58,7 +58,6 @@ end
 
 function layoutable(::Type{<:Colorbar}, fig_or_scene; bbox = nothing, kwargs...)
     topscene = get_topscene(fig_or_scene)
-    attrs = merge!(Attributes(kwargs), default_attributes(Colorbar, topscene).attributes)
 
     default_attrs = default_attributes(Colorbar, topscene).attributes
     theme_attrs = subtheme(topscene, :Colorbar)
