### 2021-08-04 #3349 @gaelforget
> > ```
> > * trying to run the 100 million points example crashed my laptop several times
> > ```
>
>
> This is a good point, I made the example with a more powerful GPU, but this could indeed be confusing. It's just that 1 million points are not as impressive of a feat, and we wanted to show what can be possible. Maybe the text should qualify that the GPU can be the bottleneck?

Would make sense to me. Could also maybe add a commented version of this line of code in the snippet with `1_000_000`, right after the one with `100_000_000`, as a hint to the user. Not sure what's best.

> > it seems unfortunate that when I do this with the code snippet in the paper the two panels on the left hand side turn into blanks
>
>
> Also a very good point. 2D vector graphics can never handle 3D data, so maybe the text should make clearer that CairoMakie will handle the 2D subset (with some 3D things working with caveats). So the scatter didn't work for you but the brain did? Maybe the point size is too small for Cairo to build up to an image, while GLMakie handles this case a little differently. Not ideal for an example.

I believe the scatter worked but unless I remove `markersize = 0.5` I see what's shown below

![paper_example_e](https://user-images.githubusercontent.com/20276764/128201601-e5f61236-4e22-4640-9deb-ab3bc4263829.png)
