### 2021-08-05 bd394a5
Typo fix in legend.md docs
