+++ docs/src/makielayout/legend.md
@@ -63,7 +63,7 @@ f
 
 With the keywords `merge` and `unique` you can control how plot objects with the same labels are treated.
 If `merge` is `true`, all plot objects with the same label will be layered on top of each other into one legend entry.
-If `unique` is `true`, all plot objects with the same plot type and label will be reduced to one occurance.
+If `unique` is `true`, all plot objects with the same plot type and label will be reduced to one occurrence.
 
 ```@example
 using CairoMakie
