+++ src/makielayout/layoutables/legend.jl
@@ -532,7 +532,7 @@ Create a single-group legend with all plots from `axis` that have the
 attribute `label` set.
 
 If `merge` is `true`, all plot objects with the same label will be layered on top of each other into one legend entry.
-If `unique` is `true`, all plot objects with the same plot type and label will be reduced to one occurance.
+If `unique` is `true`, all plot objects with the same plot type and label will be reduced to one occurrence.
 """
 function layoutable(::Type{Legend}, fig_or_scene, axis::Union{Axis, Scene, LScene}, title = nothing; merge = false, unique = false, kwargs...)
     plots, labels = get_labeled_plots(axis, merge = merge, unique = unique)
