+++ .github/workflows/Deploy.yml
@@ -23,7 +23,7 @@ jobs:
         echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://makie.juliaplots.org/previews/PR${{ github.event.number }}/" >> $GITHUB_ENV
         echo "PREVIEW_FRANKLIN_PREPATH=previews/PR${{ github.event.number }}" >> $GITHUB_ENV
     - name: Fix URLs for dev deployment
-      if: github.event_name == 'push' && github.ref == 'refs/heads/dev'
+      if: github.event_name == 'push' && github.ref == 'refs/heads/master'
       run: |
         echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://makie.juliaplots.org/dev/" >> $GITHUB_ENV
         echo "PREVIEW_FRANKLIN_PREPATH=dev" >> $GITHUB_ENV
@@ -80,7 +80,7 @@ jobs:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           TARGET_FOLDER: "previews/PR${{ github.event.number }}" # The website preview is going to be stored in a subfolder
     - name: Deploy (dev)
-      if: github.event_name == 'push' && github.ref == 'refs/heads/dev'
+      if: github.event_name == 'push' && github.ref == 'refs/heads/master'
       uses: JamesIves/github-pages-deploy-action@releases/v3
       with:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
