+++ .github/workflows/Deploy.yml
@@ -17,22 +17,24 @@ jobs:
       uses: actions/checkout@v2
       with:
         persist-credentials: false
+    - name: Get tag version
+      if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
+      id: tag
+      run: echo ::set-output name=VERSION::${GITHUB_REF/refs\/tags\//}
     - name: Fix URLs for PR preview deployment (pull request previews)
       if: github.event_name == 'pull_request'
       run: |
-        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://makie.juliaplots.org/previews/PR${{ github.event.number }}/" >> $GITHUB_ENV
+        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://jkrumbiegel.github.io/Makie.jl/previews/PR${{ github.event.number }}/" >> $GITHUB_ENV
         echo "PREVIEW_FRANKLIN_PREPATH=previews/PR${{ github.event.number }}" >> $GITHUB_ENV
     - name: Fix URLs for dev deployment
       if: github.event_name == 'push' && github.ref == 'refs/heads/master'
       run: |
-        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://makie.juliaplots.org/dev/" >> $GITHUB_ENV
+        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://jkrumbiegel.github.io/Makie.jl/dev/" >> $GITHUB_ENV
         echo "PREVIEW_FRANKLIN_PREPATH=dev" >> $GITHUB_ENV
     - name: Fix URLs for tagged version deployment
-      id: tag
       if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
       run: |
-        echo ::set-output name=VERSION::${GITHUB_REF/refs\/tags\//}
-        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://makie.juliaplots.org/${{ steps.tag.outputs.VERSION }}/" >> $GITHUB_ENV
+        echo "PREVIEW_FRANKLIN_WEBSITE_URL=http://jkrumbiegel.github.io/Makie.jl/${{ steps.tag.outputs.VERSION }}/" >> $GITHUB_ENV
         echo "PREVIEW_FRANKLIN_PREPATH=${{ steps.tag.outputs.VERSION }}" >> $GITHUB_ENV
     # NOTE: Python is necessary for the pre-rendering (minification) step
     - name: Install python
