### 2021-08-05 d21e724
correct typo and clarify glmakie / cairomakie abilities
