+++ paper.md
@@ -34,7 +34,7 @@ Finally, users can extend every step of this pipeline for their custom types thr
 
 Visualizations are crucial for all scientists around the world who want to effectively understand and communicate their data.
 Flexible and powerful plotting software is required to create publication-quality vector graphics, animated movies and interactive data exploration tools.
-Most options that are available today lack one or more of the following attributes: High performance for the responsive rendering of large datasets, interactive visualizations using flexible dynamic languages, the ability to extend the plotting pipeline to handle user-defined data structures and types, implementations of both 2D and 3D rendering, and the power to construct complex figures without having to tweak subplot positions after the fact in image editing software.
+Most options that are available today lack one or more of the following attributes: High performance for the responsive rendering of large datasets, interactive visualizations that can be updated live using dynamic languages, the ability to extend the plotting pipeline to handle user-defined data structures and types, implementations of both 2D and 3D rendering, and the power to construct complex figures without having to tweak subplot positions after the fact in image editing software.
 Therefore, researchers have to switch between tools which means they have to spend more time to learn unfamiliar syntax and redo work if one software turns out to lack critical abilities for the task at hand.
 
 `Makie.jl` is a new plotting package which is built from the ground up to leverage the power of `Julia`, a relatively young programming language which excels at technical computing and has seen steady growth of its user base since reaching the 1.0 milestone in 2018.
