### 2021-08-05 #3349 @jkrumbiegel
Oh and the version here still says ` Version: v.0.17.4` which was AbstractPlotting, we should probably change that to `v.0.15.0` which is the latest Makie.jl
