### 2021-08-07 fce32c8
remove unnecessary project.toml
