+++ docs_franklin/Project.toml
@@ -1,3 +1,0 @@
-[deps]
-Franklin = "713c75ef-9fc9-4b05-94a9-213340da978e"
-NodeJS = "2bd173c7-0d6d-553b-b6af-13a54713934c"
