+++ docs_franklin/site/examples/layoutables/axis3.md
@@ -11,20 +11,20 @@ using FileIO
 GLMakie.activate!() # hide
 Makie.inline!(true) # hide
 
-f = Figure(resolution = (1000, 1000), fontsize = 10)
+f = Figure()
 
 brain = load(assetpath("brain.stl"))
 colors = [tri[1][2] for tri in brain for i in 1:3]
 
-azimuths = LinRange(0, 2pi, 8)[1:7]
-elevations = LinRange(-pi/2 + 0.01, pi/2 - 0.01, 8)[1:7]
+azimuths = [0, 0.2pi, 0.4pi]
+elevations = [-0.2pi, 0, 0.2pi]
 
 for (i, elevation) in enumerate(elevations)
     for (j, azimuth) in enumerate(azimuths)
         ax = Axis3(f[i, j], aspect = :data,
         title = "elevation = $(round(elevation/pi, digits = 2))π\nazimuth = $(round(azimuth/pi, digits = 2))π",
         elevation = elevation, azimuth = azimuth,
-        protrusions = (0, 0, 0, 20))
+        protrusions = (0, 0, 0, 40))
 
         hidedecorations!(ax)
         mesh!(brain, color = colors, colormap = :thermal)
@@ -54,7 +54,7 @@ using FileIO
 GLMakie.activate!() # hide
 Makie.inline!(true) # hide
 
-f = Figure(resolution = (1000, 800), fontsize = 14)
+f = Figure()
 
 brain = load(assetpath("brain.stl"))
 
