+++ docs_franklin/site/examples/layoutables/axis3.md
@@ -90,7 +90,7 @@ using GLMakie
 GLMakie.activate!() # hide
 Makie.inline!(true) # hide
 
-f = Figure(resolution = (1200, 1000), fontsize = 14)
+f = Figure()
 
 r = LinRange(-1, 1, 100)
 cube = [(x.^2 + y.^2 + z.^2) for x = r, y = r, z = r]
@@ -100,10 +100,11 @@ viewmodes = [:fitzoom, :fit, :stretch]
 
 for (j, viewmode) in enumerate(viewmodes)
     for (i, azimuth) in enumerate([1.1, 1.275, 1.45] .* pi)
-        Box(f[i, j], color = :transparent, strokecolor = :gray80)
         ax = Axis3(f[i, j], aspect = :data,
             azimuth = azimuth,
             viewmode = viewmode, title = "$viewmode")
+        hidedecorations!(ax)
+        ax.protrusions = (0, 0, 0, 20)
         volume!(cube_with_holes, algorithm = :iso, isorange = 0.05, isovalue = 1.7)
     end
 end
