### 2021-08-08 e05cbdc
fix example sizes
