### 2021-08-08 95d9d50
Update transformation.jl
Fix bug in the definition of the inv_symlog10 function. It wasn't able to be evaluated before.
