### 2021-08-08 4655439
Update paper.md (#1194)
