### 2021-08-08 ccf3e86
move prepath env settings to makedocs
