+++ docs/make.jl
@@ -1,288 +1,0 @@
-using ImageMagick
-using FileIO
-using Documenter
-using Highlights
-using Markdown
-using Random
-using WGLMakie
-using GLMakie
-using CairoMakie
-using Makie
-using JSServe
-import Makie: to_string
-
-Makie.inline!(true)
-
-# Pause renderloop for slow software rendering.
-# This way, we only render if we actualy save e.g. an image
-GLMakie.set_window_config!(;
-    framerate = 15.0,
-    pause_rendering = true
-)
-
-# use svgs for CairoMakie which look crisper by default
-CairoMakie.activate!(type = "svg")
-
-"""
-    print_table(io::IO, dict::Dict)
-
-Print a Markdown-formatted table with the entries from `dict` to specified `io`.
-"""
-function print_table(io::IO, dict::Dict)
-    # get max length of the keys
-    k = string.("`", collect(keys(dict)), "`")
-    maxlen_k = max(length.(k)...)
-
-    # get max length of the values
-    v = string.(collect(values(dict)))
-    maxlen_v = max(length.(v)...)
-
-    j = sort(collect(dict), by = x -> x[1])
-
-    # column labels
-    labels = ["Symbol", "Description"]
-
-    # print top header
-    print(io, "|")
-    print(io, "$(labels[1])")
-    print(io, " "^(maxlen_k - length(labels[1])))
-    print(io, "|")
-    print(io, "$(labels[2])")
-    print(io, " "^(maxlen_v - length(labels[2])))
-    print(io, "|")
-    print(io, "\n")
-
-    # print second line (toprule)
-    print(io, "|")
-    print(io, "-"^maxlen_k)
-    print(io, "|")
-    print(io, "-"^maxlen_v)
-    print(io, "|")
-    print(io, "\n")
-
-    for (idx, entry) in enumerate(j)
-        print(io, "|")
-        print(io, "`$(entry[1])`")
-        print(io, " "^(maxlen_k - length(string(entry[1])) - 2))
-        print(io, "|")
-        print(io, "$(entry[2])")
-        print(io, " "^(maxlen_v - length(entry[2])))
-        print(io, "|")
-        print(io, "\n")
-    end
-end
-
-
-
-################################################################################
-#                              Utility functions                               #
-################################################################################
-
-
-################################################################################
-#                                    Setup                                     #
-################################################################################
-
-pathroot  = normpath(@__DIR__, "..")
-docspath  = joinpath(pathroot, "docs")
-srcpath   = joinpath(docspath, "src")
-srcgenpath   = joinpath(docspath, "src_generation")
-buildpath = joinpath(docspath, "build")
-genpath   = joinpath(srcpath, "generated")
-
-mkpath(genpath)
-
-# copy news file into src directory
-cp(joinpath(pathroot, "NEWS.md"), joinpath(srcpath, "news.md"))
-
-################################################################################
-#                          Syntax highlighting theme                           #
-################################################################################
-
-@info("Writing highlighting stylesheet")
-
-open(joinpath(srcpath, "assets", "syntaxtheme.css"), "w") do io
-    Highlights.stylesheet(io, MIME("text/css"), Highlights.Themes.DefaultTheme)
-end
-
-################################################################################
-#                      Automatic Markdown page generation                      #
-################################################################################
-
-
-########################################
-#       Plot attributes overview       #
-########################################
-
-# automatically generate an overview of the plot attributes (keyword arguments), using a source md file
-@info("Generating attributes page")
-# axis_attr_list = []
-# for a in (Axis3D,)
-#     attr = keys(default_theme(nothing, a))
-#     push!(axis_attr_list, attr...)
-# end
[…]
