### 2021-08-08 a856637
fix hist example
