+++ docs/examples/layoutables/gridlayout.md
@@ -20,8 +20,8 @@ f = Figure()
 Axis(f[1, 1], title = "My column has size Fixed(400)")
 Axis(f[1, 2], title = "My column has size Auto()")
 
-colsize!(layout, 1, Fixed(400))
-# colsize!(layout, 1, 400) would also work
+colsize!(f.layout, 1, Fixed(400))
+# colsize!(f.layout, 1, 400) would also work
 
 f
 ```
@@ -36,15 +36,15 @@ In this case, you would use `Relative(1/2)`. The available width is the width of
 ```julia
 using CairoMakie
 
-scene, layout = layoutscene(resolution = (1200, 900))
+f = Figure()
 
-layout[1, 1] = Axis(scene, title = "My column has size Relative(2/3)")
-layout[1, 2] = Axis(scene, title = "My column has size Auto()")
-layout[1, 3] = Colorbar(scene)
+Axis(f[1, 1], title = "My column has size Relative(2/3)")
+Axis(f[1, 2], title = "My column has size Auto()")
+Colorbar(f[1, 3])
 
-colsize!(layout, 1, Relative(2/3))
+colsize!(f.layout, 1, Relative(2/3))
 
-scene
+f
 ```
 \end{examplefigure}
 
@@ -75,16 +75,16 @@ The first column will get `1 / (1 + 2) * 300 == 100` units, while the second col
 ```julia
 using CairoMakie
 
-scene, layout = layoutscene(resolution = (1200, 900))
+f = Figure()
 
-layout[1, 1] = Axis(scene, title = "My column infers my width\nof 200 units")
-layout[1, 2] = Axis(scene, title = "My column gets 1/3rd\nof the remaining space")
-layout[1, 3] = Axis(scene, title = "My column gets 2/3rds\nof the remaining space")
+Axis(f[1, 1], title = "My column infers my width\nof 200 units")
+Axis(f[1, 2], title = "My column gets 1/3rd\nof the remaining space")
+Axis(f[1, 3], title = "My column gets 2/3rds\nof the remaining space")
 
-colsize!(layout, 2, Auto(1)) # equivalent to Auto(true, 1)
-colsize!(layout, 3, Auto(2)) # equivalent to Auto(true, 2)
+colsize!(f.layout, 2, Auto(1)) # equivalent to Auto(true, 1)
+colsize!(f.layout, 3, Auto(2)) # equivalent to Auto(true, 2)
 
-scene
+f
 ```
 \end{examplefigure}
 
@@ -98,19 +98,19 @@ Aspect sized columns or rows are very useful when you want to constrain the aspe
 ```julia
 using CairoMakie
 
-scene, layout = layoutscene(resolution = (1200, 900))
+f = Figure()
 
-layout[1, 1] = Axis(scene, title = "I'm square and aligned")
-layout[1, 2] = Box(scene, color = (:blue, 0.1), strokecolor = :transparent)
-layout[1, 2] = Axis(scene, aspect = AxisAspect(1),
+Axis(f[1, 1], title = "I'm square and aligned")
+Box(f[1, 2], color = (:blue, 0.1), strokecolor = :transparent)
+Axis(f[1, 2], aspect = AxisAspect(1),
     title = "I'm square but break the layout.\nMy actual cell is the blue rect.")
-layout[2, 1] = Axis(scene)
-layout[2, 2] = Axis(scene)
+Axis(f[2, 1])
+Axis(f[2, 2])
 
-rowsize!(layout, 2, Relative(2/3))
-colsize!(layout, 1, Aspect(1, 1))
[…]
