### 2021-08-09 2d6e419
density stroke
