### 2021-08-09 49ec316
jk/franklin docs (#1181)
* add copy of docs and barebones franklin

* delay plotting_functions

* add basic docstrings with style

* add some more example plots

* change to env function

* other plotting functions

* remove cruft

* remove cruft

* add basic-tutorial

* admonitions

* layout tutorial

* favicon

* makielayout

* add makielayout docs and add workflow (disable others)

* syntax fix

* fix folder

* disable software rendering loop

* url and prepath

* specify output

* anchor

* thumbnails

* other docs pages basic version

* begin block logic

* remove wglmakie live examples


[…]
