+++ docs/utils.jl
@@ -1,0 +1,286 @@
+using Chain
+using Markdown
+using GLMakie
+using FileIO
+using ImageTransformations
+using Colors
+using Pkg
+
+include("colormap_generation.jl")
+
+# Pause renderloop for slow software rendering.
+# This way, we only render if we actualy save e.g. an image
+GLMakie.set_window_config!(;
+    framerate = 15.0,
+    pause_rendering = true
+)
+
+
+function hfun_bar(vname)
+  val = Meta.parse(vname[1])
+  return round(sqrt(val), digits=2)
+end
+
+function hfun_m1fill(vname)
+  var = vname[1]
+  return pagevar("index", var)
+end
+
+
+function hfun_doc(params)
+  fname = params[1]
+  head = length(params) > 1 ? params[2] : fname
+  type = length(params) == 3 ? params[3] : ""
+  doc = eval(Meta.parse("using Makie; @doc Makie.$fname"))
+  txt = Markdown.plain(doc)
+  # possibly further processing here
+  #body = Markdown.html(Markdown.parse(txt))
+  body = fd2html(txt, internal = true)
+  return """
+    <div class="docstring">
+        <h2 class="doc-header" id="$fname">
+          <a href="#$fname">$head</a>
+          <div class="doc-type">$type</div>
+        </h2>
+        <div class="doc-content">$body</div>
+    </div>
+  """
+end
+
+
+function env_examplefigure(com, _)
+  content = Franklin.content(com)
+
+  _, middle = split(content, r"```(julia)?", limit = 2)
+  kwargs = eval(Meta.parse("Dict(pairs((;" * Franklin.content(com.braces[1]) * ")))"))
+
+  name = get(kwargs, :name, "example_" * string(hash(content)))
+  svg = get(kwargs, :svg, false)::Bool
+
+  pngfile = "$name.png"
+  svgfile = "$name.svg"
+
+  # add the generated png name to the list of examples for this page, which
+  # can later be used to assemble an overview page
+  # for some reason franklin needs a pair as the content?
+  pngsvec, _ = get!(Franklin.LOCAL_VARS, "examplefigures_png", String[] => Vector{String})
+  push!(pngsvec, pngfile)
+
+  middle, _ = split(middle, r"```\s*$")
+  s = """
+    ```julia:$name
+    __result = begin # hide
+    $middle
+    end # hide
+    save(joinpath(@OUTPUT, "$pngfile"), __result) # hide
+    $(svg ? "save(joinpath(@OUTPUT, \"$svgfile\"), __result) # hide" : "")
+    nothing # hide
+    ```
+    ~~~
+    <a id="$name">
+    ~~~
+    \\fig{$name.$(svg ? "svg" : "png")}
+    ~~~
+    </a>
+    ~~~
+  """
+
+  s
+end
+
+
+function lx_video(lxc, _)
+    if length(lxc.braces) == 1
+        rpath = Franklin.stent(lxc.braces[1])
+        alt   = ""
+    elseif length(lxc.braces) == 2
+        rpath = Franklin.stent(lxc.braces[2])
+        alt   = Franklin.stent(lxc.braces[1])
+    end
+    
+    path  = Franklin.parse_rpath(rpath; canonical=false, code=true)
+    fdir, fext = splitext(path)
+
+    # there are several cases
+    # A. a path with no extension --> guess extension
+    # B. a path with extension --> use that
+    # then in both cases there can be a relative path set but the user may mean
+    # that it's in the subfolder /output/ (if generated by code) so should look
+    # both in the relpath and if not found and if /output/ not already last dir
+    candext = ifelse(isempty(fext),
+                     (".mp4", ".mkv", ".mov"), (fext,))
+    for ext ∈ candext
+        candpath = fdir * ext
[…]
