+++ docs/tutorials.md
@@ -1,0 +1,3 @@
+# Tutorials
+
+{{list_folder tutorials}}
