+++ docs/examples/plotting_functions.md
@@ -1,0 +1,3 @@
+# Plotting functions
+
+{{list_folder_with_images plotting_functions}}
