+++ docs/examples/plotting_functions/lines.md
@@ -1,0 +1,26 @@
+# lines
+
+{{doc lines}}
+
+### Examples
+
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+Axis(f[1, 1])
+
+xs = 0:0.01:10
+ys = 0.5 .* sin.(xs)
+
+lines!(xs, ys)
+lines!(xs, ys .- 1, linewidth = 5)
+lines!(xs, ys .- 2, linewidth = 5, color = ys)
+lines!(xs, ys .- 3, linestyle = :dash)
+
+f
+```
+\end{examplefigure}
