+++ docs/examples/plotting_functions/band.md
@@ -1,0 +1,59 @@
+# band
+
+{{doc band}}
+
+### Examples
+
+\begin{examplefigure}{}
+```julia
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+Axis(f[1, 1])
+
+xs = 1:0.2:10
+ys_low = -0.2 .* sin.(xs) .- 0.25
+ys_high = 0.2 .* sin.(xs) .+ 0.25
+
+band!(xs, ys_low, ys_high)
+band!(xs, ys_low .- 1, ys_high .-1, color = :red)
+
+f
+```
+\end{examplefigure}
+
+\begin{examplefigure}{}
+```julia
+using Statistics
+using CairoMakie
+CairoMakie.activate!() # hide
+Makie.inline!(true) # hide
+
+f = Figure()
+Axis(f[1, 1])
+
+n, m = 100, 101
+t = range(0, 1, length=m)
+X = cumsum(randn(n, m), dims = 2)
+X = X .- X[:, 1]
+μ = vec(mean(X, dims=1)) # mean
+lines!(t, μ)              # plot mean line
+σ = vec(std(X, dims=1))  # stddev
+band!(t, μ + σ, μ - σ)   # plot stddev band
+f
+```
+\end{examplefigure}
+
+\begin{examplefigure}{}
+```julia
+using GLMakie
+GLMakie.activate!() # hide
+Makie.inline!(true) # hide
+lower = fill(Point3f0(0,0,0), 100)
+upper = [Point3f0(sin(x), cos(x), 1.0) for x in range(0,2pi, length=100)]
+col = repeat([1:50;50:-1:1],outer=2)
+band(lower, upper, color=col, axis=(type=Axis3,))
+```
+\end{examplefigure}
