+++ docs/src/transformations.md
@@ -1,10 +1,0 @@
-# Transformations
-
-Plots, Scenes and Subscenes are all `Transformable`, meaning that `Transformation`s can be applied to transform them.
-
-There are three main convenience functions provided for transformation:
-
-[`translate!`](@ref)
-[`rotate!`](@ref)
-[`scale!`](@ref)
-

[…]
