### 2021-08-09 b9a0887
fix typo (#1211)
* fix typo

* remove double makedocs
