+++ docs/utils.jl
@@ -284,3 +284,115 @@ function hfun_generating_versions()
   "<p>These docs were autogenerated using " * join([dep_version(x)
     for x in ["Makie", "GLMakie", "CairoMakie", "WGLMakie"]], ", ") * "</p>"
 end
+
+
+function hfun_sidebar()
+  items = join("""<li><a href="#$key">$(val[1])</a></li>""" for (key, val) in Franklin.PAGE_HEADERS)
+  "<ul class=\"sidebar\">$items</ul>"
+end
+
+struct NavEntry
+  name::String
+  children::Vector{NavEntry}
+  metadata::Dict
+end
+
+@delay function hfun_navigation()
+  all_pages = sort!(collect(keys(Franklin.ALL_PAGE_VARS)))
+
+  naventries = NavEntry[]
+
+  for page in all_pages
+      parts = split(page, '/')
+
+      this_page_vars = Franklin.ALL_PAGE_VARS[page]
+
+      d = naventries
+      for (j, part) in enumerate(parts)
+          i = findfirst(x -> x.name == part, d)
+          if i === nothing
+            n = NavEntry(part, [], Dict())
+            push!(d, n)
+          else
+            n = d[i]
+          end
+          d = n.children
+
+          if j == length(parts)
+            n.metadata["title"] = first(get(this_page_vars, "title", "" => nothing))
+            n.metadata["hidden"] = first(get(this_page_vars, "hidden", false => nothing))
+            n.metadata["order"] = first(get(this_page_vars, "order", 999 => nothing))
+            n.metadata["icon"] = first(get(this_page_vars, "icon", "" => nothing))
+            n.metadata["page"] = page
+
+            # if !isempty(title) && !hidden
+            #   route = ""
+            #   if length(parts) == 1
+            #       route = string('/', page === "index" ? "" : page)
+            #   elseif length(parts) == 2 && parts[end] == "index"
+            #       route = string('/', parts[1])
+            #   else
+            #       println(parts)
+            #   end
+            #   push!(nav, (
+            #   page = page,
+            #   title = title,
+            #   order = order,
+            #   route = route,
+            #   icon = icon
+            #   ))
+            # end
+
+            # pretty_url = match(r"(.*)/index.html", first(Franklin.LOCAL_VARS["fd_url"]))
+            # n.metadata["active"] = pretty_url !== nothing && pretty_url[1] == item.route || (pretty_url[1] == "" && item.route == "/")
+          end
+      end
+
+
+
+      
+  end
+
+  # sort!(nav, by = x -> x.order)
+  output = IOBuffer()
+
+  function printlist(io, naventries)
+    print(io, "<ul>")
+    for naventry in naventries
+      print(io, "<li>")
+
+      if haskey(naventry.metadata, "page")
+        print(io, """<a href="/$(naventry.metadata["page"])">$(naventry.metadata["title"])</a>""")
+      else
+        print(io, "<span>$(naventry.metadata["title"])</a>")
+      end
+
+      if get(naventry.metadata, "active", false)
+        print(io, "active")
+      end
+      printlist(io, naventry.children)
+      print(io, "</li>")
+    end
+    print(io, "</ul>")
+  end
+
+  printlist(output, naventries)
+  # for item in nav
+  
+  #     println(output,
+  #     """
+  #     <li><a href="$(item.route)" class="$(is_active ? "active" : "")">$(item.page)</a></li>
+  #     """)
+  # end
+
+  return String(take!(output))
+end
+
+
+function hfun_contenttable()
+  """
+  <ul>
+  $(join(["<li>$(val[1])</li>" for (key, val) in Franklin.PAGE_HEADERS]))
+  </ul>
+  """
+end
