+++ docs/franklin.md
@@ -1,187 +1,0 @@
-@def title = "Franklin Example"
-@def tags = ["syntax", "code"]
-
-# How to use Franklin
-
-\tableofcontents <!-- you can use \toc as well -->
-
-This section is meant as a refresher if you're new to Franklin.
-Have a look at both how the website renders and the corresponding markdown (`index.md`).
-Modify at will to get a feeling for how things work!
-
-Ps: if you want to modify the header or footer or the general look of the website, adjust the files in
-* `src/_css/` and
-* `src/_html_parts/`.
-
-## The base with Markdown
-
-The [standard markdown syntax](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet) can be used such as titles using `#`, lists:
-
-* element with **bold**
-* element with _emph_
-
-or code-blocks `inline` or with highlighting (note the `@def hascode = true` in the source to allow [highlight.js](https://highlightjs.org/) to do its job):
-
-```julia
-abstract type Point end
-struct PointR2{T<:Real} <: Point
-    x::T
-    y::T
-end
-struct PointR3{T<:Real} <: Point
-    x::T
-    y::T
-    z::T
-end
-function len(p::T) where T<:Point
-  sqrt(sum(getfield(p, η)^2 for η ∈ fieldnames(T)))
-end
-```
-
-You can also quote stuff
-
-> You must have chaos within you to ...
-
-or have tables:
-
-| English         | Mandarin   |
-| --------------- | ---------- |
-| winnie the pooh | 维尼熊      |
-
-Note that you may have to do a bit of CSS-styling to get these elements to look the way you want them (the same holds for the whole page in fact).
-
-### Symbols and html entities
-
-If you want a dollar sign you have to escape it like so: \$, you can also use html entities like so: &rarr; or &pi; or, if you're using Juno for instance, you can use `\pi[TAB]` to insert the symbol as is: π (it will be converted to a html entity).[^1]
-
-If you want to show a backslash, just use it like so: \ ; if you want to force a line break, use a ` \\ ` like \\ so (this is on a new line).[^blah]
-
-If you want to show a backtick, escape it like so: \` and if you want to show a tick in inline code use double backticks like ``so ` ...``.
-
-Footnotes are nice too:
-
-[^1]: this is the text for the first footnote, you can style all this looking at `.fndef` elements; note that the whole footnote definition is _expected to be on the same line_.
-[^blah]: and this is a longer footnote with some blah from veggie ipsum: turnip greens yarrow ricebean rutabaga endive cauliflower sea lettuce kohlrabi amaranth water spinach avocado daikon napa cabbage asparagus winter purslane kale. Celery potato scallion desert raisin horseradish spinach carrot soko.
-
-## Basic Franklin extensions
-
-### Divs
-
-It is sometimes useful to have a short way to make a part of the page belong to a div so that it can be styled separately.
-You can do this easily with Franklin by using `@@divname ... @@`.
-For instance, you could want a blue background behind some text.
-
-@@colbox-blue
-Here we go! (this is styled in the css sheet with name "colbox-blue").
-@@
-
-Since it's just a `<div>` block, you can put this construction wherever you like and locally style your text.
-
-### LaTeX and Maths
-
-Essentially three things are imitated from LaTeX
-
-1. you can introduce definitions using `\newcommand`
-1. you can use hyper-references with `\eqref`, `\cite`, ...
-1. you can show nice maths (via KaTeX)
-
-The definitions can be introduced in the page or in the `config.md` (in which case they're available everywhere as opposed to just in that page).
[…]
