### 2021-08-15 e847914
add logical expressions for `ispressed`
