### 2021-08-16 70133dc
computing default tellheight more carefully
