+++ src/makielayout/defaultattributes.jl
@@ -742,9 +742,9 @@ function default_attributes(::Type{Legend}, scene)
         "The height setting of the legend."
         height = Auto()
         "Controls if the parent layout can adjust to this element's width"
-        tellwidth = true
+        tellwidth = automatic
         "Controls if the parent layout can adjust to this element's height"
-        tellheight = false
+        tellheight = automatic
         "The font family of the legend group titles."
         titlefont = lift_parent_attribute(scene, :font, "DejaVu Sans")
         "The font size of the legend group titles."
