+++ docs/index.md
@@ -1,4 +1,5 @@
 @def title = "Home"
+@def order = 0
 
 ~~~
 <img class="banner" src="/assets/makie_banner.png"/>
