+++ docs/documentation.md
@@ -1,3 +1,5 @@
+@def order = 3
+
 # Documentation
 
 {{list_folder documentation}}
