### 2021-08-16 231bcf4
jk/sidebar (#1216)
* sidebar

* improve sidebar

* collapse script

* sidebar nesting

* add content bar javascript

* add js for navbar

* fix home

* mobile first css

* improve spacing

* chevrons

* layout fixes

* margin

* max height

* footer
