+++ docs/utils.jl
@@ -284,3 +284,159 @@ function hfun_generating_versions()
   "<p>These docs were autogenerated using " * join([dep_version(x)
     for x in ["Makie", "GLMakie", "CairoMakie", "WGLMakie"]], ", ") * "</p>"
 end
+
+
+function hfun_sidebar()
+  items = join("""<li><a href="#$key">$(val[1])</a></li>""" for (key, val) in Franklin.PAGE_HEADERS)
+  "<ul class=\"sidebar\">$items</ul>"
+end
+
+struct NavEntry
+  name::String
+  children::Vector{NavEntry}
+  metadata::Dict
+end
+
+@delay function hfun_navigation()
+  all_pages = sort!(collect(keys(Franklin.ALL_PAGE_VARS)))
+
+  naventries = NavEntry[]
+
+  for page in all_pages
+      parts = splitpath(page)
+
+      this_page_vars = Franklin.ALL_PAGE_VARS[page]
+
+      hidden = first(get(this_page_vars, "hidden", false => nothing))
+      hidden && continue
+
+      d = naventries
+      for (j, part) in enumerate(parts)
+          i = findfirst(x -> x.name == part, d)
+          if i === nothing
+            n = NavEntry(part, [], Dict())
+            push!(d, n)
+          else
+            n = d[i]
+          end
+          d = n.children
+
+          if j == length(parts)
+            n.metadata["title"] = first(get(this_page_vars, "title", "" => nothing))
+            n.metadata["order"] = first(get(this_page_vars, "order", 999 => nothing))
+            n.metadata["icon"] = first(get(this_page_vars, "icon", "" => nothing))
+            n.metadata["page"] = page === "index" ? "" : page
+
+            pretty_url = match(r"(.*)/index.html", first(Franklin.LOCAL_VARS["fd_url"]))
+            pretty_url = pretty_url === nothing ? nothing : pretty_url[1]
+
+            n.metadata["isactive"] = pretty_url == "/" * join(parts, "/")
+
+
+            # n.metadata["active"] = pretty_url !== nothing && pretty_url[1] == item.route || (pretty_url[1] == "" && item.route == "/")
+          end
+      end      
+  end
+
+  function should_collapse(naventry)
+    !naventry.metadata["isactive"] && all(should_collapse, naventry.children)
+  end
+
+  function navsort!(entries)
+    sort!(entries, by = e -> e.metadata["order"])
+    foreach(entries) do entry
+      navsort!(entry.children)
+    end
+  end
+
+  navsort!(naventries)
+
+  output = IOBuffer()
+
+  function printlist(io, naventries, level = "")
+    isempty(naventries) && return
+
+    print(io, "<ul>")
+
+    for (i, naventry) in enumerate(naventries)
+
+      this_level = join([level, string(i)], "-")
+
+      has_children = !isempty(naventry.children)
+
+      print(io, "<li>")
+
+      active = naventry.metadata["isactive"]
+
+      inputid = "menuitem$this_level"
+      if has_children
+        print(io, """<input class="collapse-toggle" id="$inputid" type="checkbox" $(should_collapse(naventry) ? "" : "checked")>""")
+      end
+
+      print(io, """<div class="tocitem-container">""")
+
+      if haskey(naventry.metadata, "page")
+        print(io, """<a $(active ? "class = active" : "") href="/$(naventry.metadata["page"])">$(naventry.metadata["title"])</a>""")
+      else
+        print(io, "<span $(active ? "class = active" : "")>$(naventry.metadata["title"])</span>")
+      end
+
+      if has_children
[…]
