### 2021-08-16 091c0f9
smarter defaults for tellheight and tellwidth (#1195)
* smarter defaults for tellheight and tellwidth

* computing default tellheight more carefully
