### 2021-08-16 23472d0
simplify plotting barplot by group
