### 2021-08-16 02b6efa
add x_distance example for barplot
