### 2021-08-17 09243b2
remove contentbar (#1226)
