### 2021-08-18 4df004c
include lunr in makedocs and remove old workflow
