### 2021-08-18 d6f7149
jk/lunr search (#1231)
