### 2021-08-19 ece904e
latex docs
