### 2021-08-19 #3349 @gaelforget
> @gaelforget Ok I think I have implemented your suggestions now. Fixed the typo, improved the broken sentence, and included more information about the feasibility of the large plot (a comment that a user might need more gpu power for 100 million) and what's possible with CairoMakie vs GLMakie (2D subset with some 3D capabilities).
>
> If you have any other suggestions, let me know!

LGTM
