### 2021-08-19 b60a738
(terrible?) workaround for 1 elem hists
