### 2021-08-20 4f876ef
Update src/stats/hist.jl
Co-authored-by: Pietro Vertechi <pietro.vertechi@neuro.fchampalimaud.org>
