+++ .github/workflows/ci.yml
@@ -21,6 +21,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
+          - '1.3'
           - '1.5'
           - '1.6'
         os:
