### 2021-08-21 c08a4d1
cleanup docstrings
