### 2021-08-21 470172e
call update_cam in user camera tranformations
