### 2021-08-29 #3349 @jkrumbiegel
The problem is that we currently can't edit the zenodo Metadata, as this ability is linked to the account of a former collaborator. I'm still waiting on zenodo's support to transfer ownership.
