### 2021-08-31 #3349 @jkrumbiegel
We've checked the paper again and added Simon's orcid, so from my side everything is ready. As far as I understand we can still amend the zenodo metadata later, right? And that wrong metadata won't be reflected in the JOSS paper. I'm not sure what the function of the zenodo doi is in relation to the JOSS paper
