### 2021-08-31 #3349 @arfon
> I'm not sure what the function of the zenodo doi is in relation to the JOSS paper

The primary function is to have an archive of the software that is outside of GitHub (or any other non-archival venue/platform).
