### 2021-08-31 #3349 @Kevin-Mattheus-Moerman
@jkrumbiegel just to add to what @arfon said: The Zenodo archive helps create a long term permanent/persistant storage of the software (e.g. since GitHub repositories may be renamed/removed or code can be moved, so these do not share this permanence/stability). Since the archive is there now we are happy to proceed but it is more clearly linked to the JOSS paper if the meta-data (like the title and authors) also matches. Let me know if you have any other questions.
