### 2021-08-31 #3349 @Kevin-Mattheus-Moerman
@jkrumbiegel I see archived version for `v0.15.2` have also been created https://zenodo.org/record/5269012. Let me know if we should use that archive and upgrade the version tag here again.
