### 2021-08-31 #3349 @jkrumbiegel
> @jkrumbiegel I see archived version for `v0.15.2` have also been created https://zenodo.org/record/5269012. Let me know if we should use that archive and upgrade the version tag here again.

I think we can just keep 0.15, there are patch versions coming in all the time. At least that number is round.
