### 2021-08-31 #3349 @jkrumbiegel
The DOI is correct, https://www.doi.org/10.1109/mcse.2007.55 also works. But we need to make one more correction with the example plot, we there's a typo in one subplot title. We'll let you know once that's fixed.
