### 2021-08-31 #3349 @SimonDanisch
I pushed a fix for the example plot!
