### 2021-09-01 #3349 @Kevin-Mattheus-Moerman
@jkrumbiegel  @arfon I'll leave this issue open to discuss the finalization of the archive editing. Please update us if you hear from Zenodo or once you've edited the information.
