### 2021-09-23 #3349 @arfon
> @kthyng there is an issue with the Zenodo archive, it needs to be edited (e.g. title does not match): https://zenodo.org/record/5106448 but I believe the authors are having permission issues that prevent them from editing it. We are leaving this open until it is resolved.

Having the titles match is a strong preference but not essential for JOSS. As such I propose we close this thread unless the authors especially care about this.
